R6 Class Representing PosteriorBSVART

The class PosteriorBSVART contains posterior output and the specification including the last MCMC draw for the bsvar model with t-distributed structural shocks. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BSVART with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bsvar_t$new()

specify_posterior_bsvar_t$get_posterior()

specify_posterior_bsvar_t$get_last_draw()

specify_posterior_bsvar_t$is_normalised()

specify_posterior_bsvar_t$set_normalised()

specify_posterior_bsvar_t$clone()

Method new()

Create a new posterior output PosteriorBSVART.

Usage

specify_posterior_bsvar_t$new(specification_bsvar, posterior_bsvar)

Arguments

specification_bsvar

an object of class BSVART with the last draw of the current MCMC run as the starting value.

posterior_bsvar

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBSVART.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bsvar_t$get_posterior()

Examples

data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
set.seed(123)
estimate       = estimate(specification, 10)
estimate$get_posterior()

Method get_last_draw()

Returns an object of class BSVART with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bsvar_t$get_last_draw()

Examples

data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)

# run the burn-in
set.seed(123)
burn_in        = estimate(specification, 10)

# estimate the model
posterior      = estimate(burn_in, 10)

Method is_normalised()

Returns TRUE if the posterior has been normalised using normalise_posterior() and FALSE otherwise.

Usage

specify_posterior_bsvar_t$is_normalised()

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)

# estimate the model
set.seed(123)
posterior      = estimate(specification, 10)

# check normalisation status beforehand
posterior$is_normalised()

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) %*% BB         # set negative diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()

Method set_normalised()

Sets the private indicator normalised to TRUE.

Usage

specify_posterior_bsvar_t$set_normalised(value)

Arguments

value

(optional) a logical value to be passed to indicator normalised.

Examples

# This is an internal function that is run while executing normalise_posterior()
# Observe its working by analysing the workflow:

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10)

# check normalisation status beforehand
posterior$is_normalised()

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) %*% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bsvar_t$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# This is a function that is used within estimate()
data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)
estimate       = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(estimate)
#> [1] "PosteriorBSVART" "R6"             

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$get_posterior`
## ------------------------------------------------

data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)
estimate       = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
estimate$get_posterior()
#> $B
#> , , 1
#> 
#>           [,1]      [,2]    [,3]
#> [1,] 26.870575  0.000000  0.0000
#> [2,] 24.394002 14.784509  0.0000
#> [3,] -2.359019 -3.313245 10.6949
#> 
#> , , 2
#> 
#>           [,1]      [,2]   [,3]
#> [1,] 24.276615   0.00000  0.000
#> [2,] 37.287813   7.26658  0.000
#> [3,] -5.391225 -18.58121 56.516
#> 
#> , , 3
#> 
#>           [,1]       [,2]     [,3]
#> [1,]  42.21391   0.000000   0.0000
#> [2,] -32.05149   4.254093   0.0000
#> [3,] -10.04929 -35.200747 107.1725
#> 
#> , , 4
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  54.855193   0.00000   0.0000
#> [2,]  -8.688224  10.59015   0.0000
#> [3,] -14.136722 -40.57327 126.8057
#> 
#> , , 5
#> 
#>           [,1]      [,2]     [,3]
#> [1,]  58.14233   0.00000   0.0000
#> [2,] -10.21387  10.66300   0.0000
#> [3,] -21.63257 -44.19195 139.2766
#> 
#> , , 6
#> 
#>           [,1]      [,2]    [,3]
#> [1,]  56.47574   0.00000   0.000
#> [2,] -18.65909  11.54526   0.000
#> [3,] -20.70394 -41.92282 134.644
#> 
#> , , 7
#> 
#>           [,1]      [,2]     [,3]
#> [1,]  60.15480   0.00000   0.0000
#> [2,] -23.71172  12.38342   0.0000
#> [3,] -18.54389 -42.35785 139.5014
#> 
#> , , 8
#> 
#>           [,1]      [,2]     [,3]
#> [1,]  57.96529   0.00000   0.0000
#> [2,] -29.14534  13.86362   0.0000
#> [3,] -18.22901 -38.12763 135.2273
#> 
#> , , 9
#> 
#>           [,1]      [,2]     [,3]
#> [1,]  58.25936   0.00000   0.0000
#> [2,] -28.52806  13.72920   0.0000
#> [3,] -26.94407 -42.87016 157.6055
#> 
#> , , 10
#> 
#>           [,1]      [,2]     [,3]
#> [1,]  58.16362   0.00000   0.0000
#> [2,] -22.65097  15.92612   0.0000
#> [3,] -32.27370 -41.42029 159.6195
#> 
#> 
#> $A
#> , , 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.4804702 -0.04503067 -0.4425444  0.7969551
#> [2,] -2.3308842  1.71228006  2.1439891  1.2196554
#> [3,] -0.2391456  0.36828874  0.7818442 -0.5198334
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.5716152 -0.1324149 -0.5442413 -0.0193477
#> [2,] -1.0246751  1.2317472  0.9051312 -0.5924543
#> [3,] -0.1658356  0.1537583  1.0770827  0.5242285
#> 
#> , , 3
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,]  1.1571278 -0.02486692 -0.1317656 0.2539785
#> [2,] -0.6704007  1.37093451  0.4582159 0.7402311
#> [3,] -0.1817455  0.12842039  1.1058511 0.3390234
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.9587770 0.02198567 0.04253967 0.1540261
#> [2,] -0.5727379 1.35653321 0.36495939 0.8177823
#> [3,] -0.1896258 0.12421942 1.12041828 0.3330203
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.9322192 0.02655012 0.07397823 0.1863531
#> [2,] -0.5741270 1.31723004 0.37742498 0.5099146
#> [3,] -0.1666877 0.12438587 1.09199619 0.3392703
#> 
#> , , 6
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9284099 0.01561624 0.07322476 0.03534699
#> [2,] -0.4710026 1.32135232 0.25998643 0.64442177
#> [3,] -0.1577667 0.11765424 1.08705648 0.31965796
#> 
#> , , 7
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.9278112 0.02861172 0.06608255 0.1033858
#> [2,] -0.4713327 1.28264375 0.27942948 0.4056981
#> [3,] -0.1464684 0.09785360 1.08028868 0.1794600
#> 
#> , , 8
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9339919 0.02831593 0.04912579 0.02969466
#> [2,] -0.3904908 1.26538839 0.18372668 0.27891893
#> [3,] -0.1131104 0.08941772 1.04459186 0.14277188
#> 
#> , , 9
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.92360356 0.03071039 0.06157173 0.04877192
#> [2,] -0.33070859 1.25378068 0.12103866 0.25632653
#> [3,] -0.09904554 0.08894107 1.03224892 0.17989758
#> 
#> , , 10
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.95645730 0.03094050 0.02337748 0.07268629
#> [2,] -0.30545481 1.27098433 0.10033340 0.51574293
#> [3,] -0.08557641 0.08455731 1.02412119 0.20423410
#> 
#> 
#> $hyper
#> , , 1
#> 
#>           [,1]       [,2]
#> [1,]  3.326552  0.4174181
#> [2,]  7.369414  2.4304551
#> [3,]  2.516382  0.9161348
#> [4,] 23.353356 10.0762454
#> [5,] 33.049659 18.8228103
#> [6,] 35.569478 15.0666624
#> [7,]  2.615414  1.0190860
#> 
#> , , 2
#> 
#>           [,1]       [,2]
#> [1,] 56.847022  0.3883003
#> [2,] 70.502288  7.7517370
#> [3,] 19.930982  1.3885801
#> [4,] 26.556838  7.9341364
#> [5,] 56.009050 32.5470045
#> [6,] 24.341357 17.2957458
#> [7,]  4.409854  2.3903528
#> 
#> , , 3
#> 
#>            [,1]       [,2]
#> [1,]  48.081533  0.9034924
#> [2,] 343.118707  1.7330607
#> [3,] 189.629403  1.0446011
#> [4,] 115.063247  9.6572192
#> [5,] 156.017799 16.8548264
#> [6,]  74.147795 13.9543233
#> [7,]   7.081297  1.7913465
#> 
#> , , 4
#> 
#>            [,1]       [,2]
#> [1,]   82.84392  0.6135754
#> [2,]   91.74652  1.3738528
#> [3,] 1338.19923  2.3016583
#> [4,]  223.15115  9.4000360
#> [5,]  156.00805 19.5661520
#> [6,]  252.07165 18.0923637
#> [7,]   13.09037  1.3121502
#> 
#> , , 5
#> 
#>            [,1]      [,2]
#> [1,]  156.97849  1.056376
#> [2,]   37.96808  1.128493
#> [3,] 1774.99296  1.566079
#> [4,]  278.19774 13.140372
#> [5,]  337.20785 15.773098
#> [6,]  252.55418 28.763629
#> [7,]   22.42122  1.955250
#> 
#> , , 6
#> 
#>            [,1]       [,2]
#> [1,]  724.78506  1.0558917
#> [2,]   85.06300  0.6696044
#> [3,] 1331.27913  1.1006033
#> [4,]  429.50292 12.7292690
#> [5,]  398.71093 11.3567229
#> [6,]  418.10161 16.0821181
#> [7,]   32.02853  1.7226517
#> 
#> , , 7
#> 
#>            [,1]       [,2]
#> [1,]  334.73168  0.5297545
#> [2,]   66.19846  0.8451408
#> [3,] 1320.32127  3.3152765
#> [4,]  672.96938 11.7860823
#> [5,]  585.15346  9.8581069
#> [6,]  879.47082 17.3700652
#> [7,]   44.04476  1.4833461
#> 
#> , , 8
#> 
#>           [,1]       [,2]
#> [1,]  634.3884  1.3565989
#> [2,]  114.3976  0.9138923
#> [3,] 1646.8058  0.8213030
#> [4,] 2003.7475 12.2898287
#> [5,] 1086.9674 13.5592045
#> [6,] 1601.6555 21.0513067
#> [7,]  105.8914  1.7462631
#> 
#> , , 9
#> 
#>           [,1]       [,2]
#> [1,]  509.8000  0.4076954
#> [2,]  136.9066  0.7769130
#> [3,] 2093.2840  0.8403957
#> [4,] 2518.7749  6.3035145
#> [5,]  675.9234 13.4348042
#> [6,] 1628.2611 16.7638744
#> [7,]  180.6883  1.3197935
#> 
#> , , 10
#> 
#>           [,1]       [,2]
#> [1,] 1344.8224  0.8920807
#> [2,]  220.9185  0.5071264
#> [3,] 2463.9122  0.8275948
#> [4,] 2327.8140  9.0235227
#> [5,] 2305.4515 10.2630265
#> [6,] 2886.2256  9.4632208
#> [7,]  176.8318  1.2356995
#> 
#> 
#> $lambda
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>   [1,]  7.949416  0.8132173  2.5042021  1.7215826  2.5748178  7.9835659
#>   [2,] 40.039958  2.1707698  4.0694795  0.6662830  1.1918390  3.1984474
#>   [3,] 12.440581  2.4925702  3.7687474  2.7130622  1.2772048  0.8453295
#>   [4,]  6.443517  7.2052466  3.2915567  1.5627542  2.0673129  4.1173580
#>   [5,] 10.515118  2.4709679  4.7207755  1.3307017  1.1554496  1.1090064
#>   [6,] 28.106257  2.3010934  2.4320033  1.7783639  2.6153049  6.0560250
#>   [7,] 40.939334 15.4520507  1.4744984  0.8014008  4.6682736  2.6430036
#>   [8,]  7.761822 10.5831133 11.5386433 15.0658520 11.5020165  7.0131570
#>   [9,] 11.116515  6.3382427 13.3591628  4.7844407  2.1332395  1.8330117
#>  [10,] 10.973081  3.1482423 10.3918351  6.2925453 10.9945432 15.1511833
#>  [11,] 12.345675  8.7844103 14.4266894  5.3270206  3.9322395  2.1338708
#>  [12,] 17.037334  8.7523680 22.7291119 10.6692469 11.1923166  3.7444883
#>  [13,]  7.040674  2.5196825  2.8454071 10.2933564  6.7203196  1.5246935
#>  [14,]  8.266596  3.6182723  5.0078917  4.1623636  5.8450650  9.1469225
#>  [15,] 12.244820 10.2180757  6.0537257  3.2018192  2.7193363  5.5741219
#>  [16,] 15.253326  3.3594303  2.7270353  3.9062118  2.4969121  1.6759473
#>  [17,] 13.150858  3.9129427  8.1706642  2.3226169  2.5698179  2.9684471
#>  [18,] 21.260259  3.6201503  2.0205827  2.0229602  0.8938979  1.2728359
#>  [19,] 12.196792  4.6631869  3.7552483  3.4776974  1.4729928  2.0691816
#>  [20,] 34.893639  3.9434904  1.1703209  0.8292661  2.5398041  0.9997848
#>  [21,] 23.129176  8.5419369  1.9446747  1.0995800  1.2477479  1.1259965
#>  [22,] 22.880294  6.0485825  3.8509152  3.5598133  0.6817711  0.8905734
#>  [23,] 13.873519 11.6956806  3.9489488  6.1445322  1.6714025  2.9539165
#>  [24,] 13.825634 15.6777509  2.2228835  1.3033932  3.8137598  4.0155540
#>  [25,] 12.748303  7.5943496  2.9477518  1.4046635  2.7517089  3.5123733
#>  [26,]  8.263517  4.8678565  6.6542237  0.8367692  2.9538046  1.2034156
#>  [27,]  8.734584 11.3789944 18.8746667  1.3734879  1.7342155  1.5567372
#>  [28,]  9.267945 32.2517159 10.5730407  3.6470145  3.7759347  2.9372550
#>  [29,] 12.406562  7.6712874  4.1661734  2.2350953  1.6325424  0.9206611
#>  [30,] 10.457580  7.2418461 12.5200186 18.9732912  0.9496880  1.6697783
#>  [31,] 25.045277  5.8212793  3.4285450  2.4717376  1.6643493  1.1474360
#>  [32,] 13.459182  6.0469539  4.6760076  1.2156896  1.9155789  0.8429818
#>  [33,] 13.069005  4.2096737  3.5355567  3.1063302  1.1035274  7.1698394
#>  [34,]  7.134092  9.6700532  2.9275743  2.7464071  1.3449327  1.2023675
#>  [35,] 23.196496  7.0135303  2.5641292  1.2019901  1.3567291  1.5849977
#>  [36,]  9.868568  4.6157350  9.3918415  2.6232754  2.0030436  1.2843540
#>  [37,] 19.825516  6.2629531  2.7788717  2.1475280  0.9108793  1.6180085
#>  [38,] 12.509933  7.6482016  2.0590777  1.5658924  1.5991316  1.1872822
#>  [39,] 15.049912  4.8927878  3.0601253  5.0378765  2.0125664  1.4313322
#>  [40,] 12.334131  7.5165917  1.7636487  2.0234371  4.7111310  1.8851323
#>  [41,] 10.470176  3.7339808  3.3425843  1.5079828  0.8405856  2.1644906
#>  [42,] 21.972158  5.1818418  3.6443585  2.3605901  1.6479938  1.9086297
#>  [43,] 31.728752 17.4161298  2.1840041  3.8363276  1.6204524  1.0985043
#>  [44,] 22.286257  7.3584847 10.9312312  5.6052211  5.8033684  2.7003391
#>  [45,] 10.738572  7.6613688  2.1174077  3.4626278  2.3443864  0.6891587
#>  [46,]  9.957839 14.3753733  3.2669771  1.7050152  1.6037448  2.0758264
#>  [47,] 15.658230 16.2940803  2.7897115  1.7600534  2.2576288  1.0663468
#>  [48,]  5.095093  3.7675287  5.7186450  3.3652965  1.5737772  7.9484242
#>  [49,] 27.182604  2.7711578  3.8870534  1.3471230  0.7893527  2.3260983
#>  [50,] 10.355597  1.2660560  1.1143812  1.8086875  1.2538060  0.9173120
#>  [51,]  9.700111  3.9187892  0.8962574  4.8203439  3.7573196  0.7239788
#>  [52,] 11.685185  4.8509323  0.9334527  1.5263372  1.1084380  1.0182876
#>  [53,]  7.906141  1.6426787  2.4786207  0.4014983  0.9169953  0.8003631
#>  [54,]  5.214345  2.2002627  1.0541692  1.0337959  0.7827290  0.4905043
#>  [55,] 15.443622  3.6063111  2.4278145  1.9474995  0.7572841  2.0777373
#>  [56,] 19.655321  2.2993796  4.4170642  1.9167547  1.1600414  2.5001357
#>  [57,] 17.447540 15.4200071  3.2130111  1.6323530  0.6977410  0.4754619
#>  [58,]  8.808066  6.7752513  1.5765687  2.2281558  2.1134121  1.3459932
#>  [59,] 11.799574  4.5468928  2.1608536  1.9514245  1.8724796  1.4984145
#>  [60,] 24.454866  5.7338197  2.5670043  2.2461307  1.2897930  0.8470422
#>  [61,] 18.284350  5.9180405  1.5604907  3.4704618  0.6453886  0.5212646
#>  [62,] 11.695053  2.0112236  1.3294767  1.0427003  0.7985830  0.7359724
#>  [63,]  9.800052  5.2659801  6.8220530  1.8548456  1.6713177  1.9960841
#>  [64,] 14.312513  5.0975618  1.0738518  5.3626820  2.4608032  3.4928622
#>  [65,] 18.099926  4.0024731  1.0142082  1.4606475  2.8307281  2.8522183
#>  [66,] 10.045664  4.4981138  4.1093433  2.0924329  1.7578238  1.2967709
#>  [67,]  7.283288  6.9283211  3.5962463  1.5894348  1.7115563  1.0288431
#>  [68,]  9.619917  5.6669497  5.5544487  8.1467173  2.2304947  1.8058245
#>  [69,] 14.040944  2.6634629  3.2892052  2.9698599  0.8708022  0.8422093
#>  [70,] 10.637036  2.2777386  1.7085687  1.0876350  1.4925852  1.3209430
#>  [71,] 15.160894  3.9803159  9.1901993  3.4326351  0.9330211  0.6164836
#>  [72,] 10.346901  4.9621320 10.4264580  4.0990978  5.5322924  2.1948200
#>  [73,] 16.334212  6.8743135  1.2923897  1.6626316  1.6929514  1.8643749
#>  [74,]  6.582826  3.9662140  1.6425672  1.0625234  2.6341445  1.5179483
#>  [75,] 14.752268  7.0923353  2.3315486  0.6125462  1.8148055  0.9630327
#>  [76,]  5.969686  9.8594572  4.8554034  1.2321681  0.9015310  1.0320680
#>  [77,] 12.941983  3.2224493  2.3269890  0.9484798  0.6857370  1.3324185
#>  [78,] 10.443256  4.6574858  5.1456216  0.6828809  0.8948403  1.1236951
#>  [79,]  5.473777  6.6073498  2.8616976  1.5906144  0.9181521  3.0530882
#>  [80,] 11.174853 14.5874420  4.3415013  1.9507540  1.1275719  2.0402225
#>  [81,]  6.810104  2.1931366  2.0947930  0.8657873  0.7010762  0.5413730
#>  [82,] 15.385185  3.6155601 10.1361647 47.5932815  1.8243355  3.9216242
#>  [83,] 11.692416  1.4292486  1.4703378  2.0713782  0.8739475  0.8874329
#>  [84,]  9.232004  1.8528411  3.1194184  2.1716745  1.3947882  1.5197121
#>  [85,]  8.966507  3.0695793  1.0876417  1.1649873  0.5755682  1.1125811
#>  [86,] 12.918368  1.9094851  1.8425633  1.9057597  1.6400786  2.4861457
#>  [87,] 15.740989  3.6677945  1.3258766  1.0822523  0.9052043  1.0225147
#>  [88,]  9.188174  2.7055072  1.1942951  0.8533250  1.2513559  2.2461247
#>  [89,] 13.330687  3.4451093  1.1880785  1.0908039  1.3170435  0.7678015
#>  [90,]  9.120186  1.9422222  0.8589157  1.5936919  1.6072041  1.4207081
#>  [91,] 10.089198  1.7378603  1.2708775  2.3120391  1.3402866  1.9850630
#>  [92,] 11.096455  2.7137349  1.2682108  3.8548869  1.4890663  3.9568741
#>  [93,] 16.614794  5.6377691  0.8436056  0.8379131  2.1682097  1.6758292
#>  [94,]  7.620603  1.8321288  1.2462946  1.3705955  1.1528004  0.7125329
#>  [95,]  7.992990  3.3693730  1.1023971  0.8379728  1.7002173  1.5201769
#>  [96,]  9.567312  3.0610566  7.0623480  9.1943518  3.9138003  2.2965603
#>  [97,]  9.924322  1.2499651  0.7760736  0.9643153  1.2823619  0.9117849
#>  [98,]  8.944613  2.3734005  1.0144232  1.3397893  1.2865544  0.9504046
#>  [99,]  5.464103  1.8570488  0.8813388  0.8364655  1.0455192  1.0885444
#> [100,]  8.849860  1.9278728  4.0689017  0.9772679  1.2365993  2.5542605
#> [101,] 18.037499  1.6920824  0.7307249  0.7325127  1.0951595  1.5946168
#> [102,] 11.430109  3.3788328  1.5257087  0.9090282  1.4042461  2.4992190
#> [103,] 20.562411  2.5633515  0.8270780  0.4758077  1.2750268  0.5061473
#> [104,]  7.236893  2.5115459  1.4003611  1.3148210  2.5299097  2.8783523
#> [105,]  7.070161  1.2061336  1.2331164  0.7550891  1.5944723  0.5440389
#> [106,]  5.964154  3.8246180  1.2619788  0.9358690  2.5732624  2.7491350
#> [107,] 10.805554  1.0545628  0.7925821  1.6631058  2.4738521  1.4578610
#> [108,] 14.666338  3.5889268  1.4388908  1.1516035  1.5198057  2.6684541
#> [109,]  7.307170 15.9570735 25.6840544 17.0474549 12.3906551 16.7661019
#> [110,] 29.403352  6.1629701  8.9048379  6.3961657  5.6465566  5.9825588
#> [111,]  8.878189  1.3203651  0.5487706  0.8836734  0.8604264  1.0448111
#> [112,] 12.421472  2.8235562  2.1574963  2.0852579  1.4272905  1.1782405
#> [113,]  8.631129  5.2369589  1.4168629  1.5335704  1.1261949  0.6776020
#> [114,] 13.667098  2.8795637  0.9965058  0.4205483  1.2705447  0.4130017
#> [115,] 13.664798  1.3793843  0.8579030  0.9871709  0.7740125  1.0691884
#> [116,] 33.129200  1.0192692  1.7674202  1.2736902  1.5415374  1.8170159
#> [117,] 25.677681  2.5938891  1.8021892  1.5993464  1.1354984  0.6291561
#> [118,]  8.667603  1.4964028  2.4521679  1.0399980  3.2789213  1.2227253
#> [119,]  5.123934  1.1205561  1.1176034  0.3391906  0.6218728  1.2008729
#> [120,] 25.926890  2.9074623  0.9675734  1.4957571  1.5411990  0.9760175
#> [121,] 69.584459  2.9457813  2.4997033  1.3150359  1.4591901  3.3962796
#> [122,] 10.121403  0.5412379  0.8318384  0.8360418  1.2933824  1.9942158
#> [123,] 21.221697  1.3248121  2.4966430  1.6085594  0.7906672  0.4308855
#> [124,] 12.005077  1.4834355  4.0720358  0.9439259  1.0097431  1.3053485
#> [125,]  9.143438  0.9478600  1.1882737  0.6783369  1.6164478  1.6198930
#> [126,] 14.462086  1.7833763  0.8424526  0.7864245  2.0646483  2.3984177
#> [127,] 17.625125  1.6138507  1.8671784  6.1545446  2.2211067  2.3171768
#> [128,]  7.432488  3.7989343  2.3756099  7.4543507  1.5713628  1.9005306
#> [129,] 21.694514  2.6461221  1.9392013  3.5037622  3.7850568  9.7892418
#> [130,]  6.456702  1.5236253  1.6976793  1.2159543  0.7877647  0.8081995
#> [131,]  8.805943  1.5532519  1.3448365  1.1914609  0.8271980  4.3256220
#> [132,] 15.259023  5.2757997  2.0688484  1.0448387  2.1873212  1.2235886
#> [133,]  9.893369  7.4048952  5.6923052  2.6716212  7.5944238  6.2462313
#> [134,]  5.213691  2.8724489  1.1977753  1.6497695  1.0294182  0.7692860
#> [135,] 23.634248  3.2167334  2.3555837  5.2786297  1.8551053  2.2876550
#> [136,] 11.973362  3.6381191  1.5114198  2.2295857  2.9327835  2.9786868
#> [137,] 13.602898  2.9100319  2.3976291  1.4342393  1.6273737  1.9004858
#> [138,] 22.282123  3.2112299  3.0932848  2.6410265  1.8868090  1.5209466
#> [139,] 20.524366  2.5538562  2.5909504  2.3496299  2.7175461  1.9224588
#> [140,]  9.486888  1.2530811  2.1406175  1.2843572  0.9219370  0.8084645
#> [141,] 14.823150  1.3367546  0.9436410  0.9823430  0.6740914  1.6344491
#> [142,] 18.907491  3.8093163  2.1393293  1.3396317  3.0962893  1.5063054
#> [143,] 13.997862  0.8257470  1.0759613  1.1907658  0.4989761  1.4467390
#> [144,] 51.658522  0.7465590  1.0406375  3.0138943  0.9549873  2.1997573
#> [145,] 12.797662  2.7279619  0.8331734  1.1777457  1.1519955  3.1020887
#> [146,] 12.248595  3.0904683  1.0301431  1.2201128  0.9448074  1.0738701
#> [147,]  8.138772  2.9618106  0.9553955  1.8546343  0.8012380  2.1020797
#> [148,] 15.011660  6.1473604  2.2356998  1.2754743  0.8676589  1.6087522
#> [149,]  8.302054  1.2223873  2.2195398  2.5625495  3.3660851  5.5669707
#> [150,]  8.131932  2.3140653  1.4735571  1.5706065  0.9082141  1.5157846
#> [151,]  6.351028  3.6020986  1.8760476  1.9711273  0.9989882  0.4169986
#> [152,] 15.650118  3.6103549  1.7365786  0.6233479  0.8008592  2.6004120
#> [153,]  5.075960  2.9710771  0.9088481  0.9427502  2.1675387  3.0745174
#> [154,]  8.576575  2.9971482  1.1951900  0.8672742  1.4377507  0.6781425
#> [155,] 24.127221  1.6409697  0.7969763  0.5832973  0.9988818  1.9430518
#> [156,] 24.592972  3.4734102  1.9097825  1.8496545  3.0328199  0.4510527
#> [157,]  4.727349  1.4185003  1.9386043  2.3553707  1.5154395  2.3995633
#> [158,]  6.462471  1.4783656  1.1782157  7.6641106  1.1751711  1.5431319
#> [159,] 13.027050  1.6709414  0.8653849  1.3686422  1.4870038  0.4189261
#> [160,] 26.797041  1.3759073  1.6359989  1.0981570  1.6127888  2.0060555
#> [161,] 14.737812  1.8252797  2.0653704  0.9187466  1.9346342  0.8888094
#> [162,] 11.495334  1.9682381  1.1672328  1.3259503  0.6911286  0.6244654
#> [163,]  5.712074  1.8383378  4.1769135  0.9395255  1.2768429  0.5956880
#> [164,] 11.763675  1.1955480  1.9993784  2.1534410  1.0073570  1.0453078
#> [165,]  9.569355  4.3169270  1.6252510  1.4427788  1.6806337  1.7022027
#> [166,] 14.590905  3.4532862  1.5252716  1.0816360  0.9800538  0.9343520
#> [167,] 15.077494  2.4567691  1.6576822  0.5089035  0.7749459  1.1661315
#> [168,]  7.054348  3.3416303  1.3659754  0.8746534  0.4659566  1.2755453
#> [169,] 10.553833  1.6522881  1.1329624  0.9165794  0.4812230  0.9535848
#> [170,]  4.070697  2.7455985  0.6114832  1.5217125  0.6560821  1.3888568
#> [171,]  9.970322  5.9652320  0.6449357  1.4395065  1.4038726  1.6167242
#> [172,] 11.623936  1.6180316  1.8462001  1.6154504  0.9972039  1.0804375
#> [173,] 15.574987  1.9159770  0.7542732  0.7026695  1.6179518  1.0554678
#> [174,] 19.982530  1.3097517  0.7555247  1.2493757  0.6008817  0.5658692
#> [175,]  8.301642  2.4356964  3.3127552  2.8571296  1.1332666  1.8972187
#> [176,] 13.904659  1.4179917  0.5408621  0.8878594  0.5552848  0.8458524
#> [177,] 14.089671  4.4689856  0.8935714  1.3044318  1.0585956  1.0403568
#> [178,] 17.025821  1.9768468  0.9838364  1.0637728  0.8210092  0.8495320
#> [179,]  8.430138  2.8463007  0.6562466  1.0903199  0.9664730  0.7798142
#> [180,] 22.814138  4.8400174  2.6696041  1.2859971  0.6292883  2.0536596
#> [181,] 12.116904  2.3591135  1.3329290  0.7240977  1.5422192  0.9025513
#> [182,] 16.136801  1.6481337  3.5846618  1.3573042  1.0441731  1.5656765
#> [183,] 15.776290  2.7009898  1.0191143  1.6040363  0.7917626  0.8215748
#> [184,]  7.696011  4.1443535  1.4810908  1.2385264  1.6224019  4.7155940
#> [185,]  8.455886  2.9664477  3.2388735  1.8602747  1.6567569  1.2169591
#> [186,] 12.465246  2.0440372  2.1173285  1.1196756  1.0108987  1.1665126
#> [187,] 18.016501  3.7445008  1.8277814  1.2409573  1.7093656  3.3362138
#> [188,]  7.794437  3.0184541  0.6052254  0.8046564  0.7401193  0.6673931
#> [189,]  9.258612  3.0810081  1.4022197  0.6398565  1.0691322  0.9295848
#> [190,] 23.797205  1.3203019  0.8905412  0.9964186  1.5055595  0.9148742
#> [191,] 12.871755  6.9243873  1.7510352  1.5379683  1.7787964  3.1098262
#> [192,] 29.293114  3.5767378  0.8140480  0.9423634  2.0951668  1.3993372
#> [193,] 26.849517  2.3745701  2.8040054  0.6708822  1.5455465  2.1262358
#> [194,] 33.988146  4.4904323  1.1251138  0.7516761  1.2705437  1.7417922
#> [195,] 14.310470  3.3153270  1.4123956  2.7385595  1.3662545  1.1847897
#> [196,] 28.024223  6.5690804  1.7231079  1.4465763  1.0553050  1.3820684
#> [197,] 11.210957  3.5806845  1.0523434  1.0474763  2.6664974  2.5045972
#> [198,]  5.999006  6.1937734  3.1653758  1.6721701  1.9964044  1.6815703
#> [199,] 12.286317  5.7951812  4.5030404  1.8506530  2.9916760  4.9771277
#> [200,] 13.551795  8.8587625  1.6296339  1.1800246  2.2711006  2.2560678
#> [201,] 12.459007  5.0554523  1.2489626  1.3962471  2.1080146  4.9223994
#> [202,]  7.856007  4.3706996  1.8109134  1.3231130  1.1852985  3.6899734
#> [203,] 14.091782 13.3654205  4.1116932  1.4235794  2.0865428  1.3941174
#> [204,] 11.505041  4.5776696  1.5950648  1.7470982  0.5592987  1.3925744
#> [205,]  6.705141  5.6924542  2.6861034  1.7880367  1.9146838  1.3317150
#> [206,]  4.737832  3.5027155  1.6919733  1.2017173  3.1119901  1.0780540
#> [207,]  6.068668 12.3547502  1.1570489  1.1860284  0.7559395  2.4652921
#> [208,] 12.612050  3.9420821  1.7551227  1.0247282  1.6153673  1.3043597
#> [209,]  8.702938 12.4823678  3.2822205  1.5727156  1.8007779  3.6353874
#> [210,]  7.680585 20.2157181  1.6036398  1.4145868  2.8390031  3.8896493
#> [211,] 16.275916 11.0493164  2.6278416  2.5958062  2.7522098  1.3635464
#> [212,]  7.622727 17.2603530  2.2392221  2.2427297  1.9788728  1.3710135
#> [213,]  8.544610  6.8668396  8.3608628  2.9779823  4.9926703  4.5315178
#> [214,] 27.723917 12.3527040 11.2373844  5.0216110  8.8103050  8.1196406
#> [215,] 13.573929  8.8359124  1.7575658  0.9798225  2.1318144  1.7830023
#> [216,] 10.697602  2.2155543 10.0032820  8.4885199  2.8507687  8.0085907
#> [217,] 11.162993  3.5803685  0.6127942  1.2975608  1.5347295  1.4118725
#> [218,]  9.441007  2.5719046  0.8153171  2.1161478  1.0948158  0.9574826
#> [219,] 11.376632  2.0964088  1.4659923  1.5193639  5.2505800  3.3685601
#> [220,] 33.836508  5.3848367  1.8674064  1.5568136  2.4909633  1.5143967
#> [221,] 18.273121  2.5055801  0.9328180  1.4993087  1.1983670  1.7920783
#> [222,] 13.143701  4.3835900  2.5633666  4.9122762  7.6596040  2.8263009
#> [223,]  6.731599  5.2276737 11.3528810  1.1078820  2.2981421  1.9831395
#> [224,]  6.438355  4.5916940  2.2542113  0.5668997  1.0968051  1.2591679
#> [225,] 12.546155  3.7616558  3.3324792  1.2755670  0.7372611  0.9864705
#> [226,] 12.690049  7.5696849  0.8785735  1.6162878  0.7096890  0.7245077
#> [227,] 11.492904  2.7888725  1.2315884  1.1362678  1.0696516  1.8983492
#> [228,]  7.080245  8.9663942  2.5350734  1.7655873  0.6567670  1.4326087
#> [229,]  6.843110  2.3067000  1.6255372  1.4024426  1.0265634  0.8662931
#> [230,] 15.445377  3.8987613  0.6672065  1.0180870  0.9561317  0.7796176
#> [231,] 10.223895  1.9925810  1.6415200  1.9476843  0.9170028  1.1520455
#> [232,]  5.607070  4.6003680  0.9856552  0.6072470  2.7631769  1.0248083
#> [233,]  6.557319  1.9438687  0.6276190  0.8347809  1.0315632  0.8991436
#> [234,] 16.783205  1.4815146  1.0492836  2.1695530  1.4067635  2.7294338
#> [235,] 14.493818  4.3533407  1.0257342  0.8490944  1.3064849  1.3156960
#> [236,] 20.824848  2.8467433  0.4835100  1.0377550  0.9363307  0.8365157
#> [237,] 13.327295  5.9407871  0.7966679  1.7693865  0.6846187  1.5933351
#> [238,] 45.702438  1.4892700  1.1304010  1.1432845  0.7291241  2.3228245
#> [239,]  7.865513  2.6230660  1.1723184  1.1422005  1.1526076  3.1982035
#> [240,]  8.139406  2.9051808  3.0338209  1.5892381  1.4028795  1.9863275
#> [241,] 33.841571  1.9980292  0.8386563  0.9883841  1.2205237  0.6234650
#> [242,]  7.877181  2.3719641  0.7159560  1.1418577  1.1703495  2.9752264
#> [243,] 34.518081  4.2075760  0.5107220  1.7110694  2.4660194  1.3396781
#> [244,]  7.049992  8.7975732  3.3062100 14.4403768  6.5058285 12.6687905
#> [245,] 14.596360  3.2359573  3.1566990  2.4781898  1.4523825  2.2181949
#> [246,] 11.694690  8.8724148  1.7285766  2.6384134  1.2084101  2.5653994
#> [247,] 16.958142  3.0168009  2.0337386  0.7908931  0.7668796  1.0799866
#> [248,]  7.300832  6.0358139  9.9224474  1.2598955  1.1703391  0.6532920
#> [249,]  9.785336  3.5963549  4.1620404  1.6363680  0.6425639  1.0467114
#> [250,]  6.464027  4.5236841  2.2698886  2.0363290  1.2980892  0.8691579
#> [251,] 24.861002 12.8448845  1.6431191  1.2189950  1.4659819  1.3510612
#> [252,] 25.490662  4.1270294  8.5791812  2.3271322  2.1209870  3.3410362
#> [253,]  8.044127  6.1028703  1.5858509  1.1407268  1.4929815  0.7972294
#> [254,] 11.432010  4.7048134  1.7310733  2.5806700  0.5768187  1.4359153
#> [255,]  6.900453  4.3200010  2.5533453  2.2131880  2.3328991  1.6127142
#> [256,] 30.492812  7.6837370  2.5307141  1.3114868  0.3636010  0.7404239
#> [257,]  7.071157 29.1343916  2.3344909  0.9363145  1.1715483  1.0987713
#> [258,] 22.535503 13.6743905  1.9150669  0.6869084  3.9324239  1.8648515
#> [259,] 27.287553  5.0620234  1.4575379  1.8474363  1.5670085  0.4258231
#> [260,]  9.361614  8.3254136  2.2876097  2.4599530  3.7646320  4.8486600
#> [261,]  7.928296  4.8702764  0.7820080  0.5315818  1.0319651  0.6639533
#> [262,] 10.135769  2.0820646  1.5795357  1.7401117  5.2260698  6.2907226
#> [263,] 18.036941  3.9016022  2.5178045  0.9423386  1.4614015  1.2058884
#> [264,]  8.465348  2.7884803  1.4669543  0.7736328  0.6803737  0.8825995
#> [265,]  7.371724  3.1119846  1.4045077  0.8665924  2.8798804  1.7279143
#> [266,]  5.623151  2.8566227  0.5116365  3.1445997  0.6676107  1.4635050
#> [267,] 11.030698  2.7306651  2.5716765  1.1500295  4.0930019  1.1291046
#> [268,] 11.632822 14.5959775  1.9491051  1.1597868  1.0605107  1.5336165
#> [269,]  9.617194  4.5343430  1.1640685  2.2312759  0.6211973  0.6743037
#> [270,] 21.311071  2.8756254  3.0872984  1.3279282  2.7330065  0.6793148
#> [271,] 19.827785  8.1700759  2.4759917  0.5816863  1.4546233  0.9288868
#> [272,] 13.790665  3.1661432  2.6081304  1.8698979  1.0343870  1.9503750
#> [273,]  5.092950  1.8036468  1.1226178  0.8733292  1.3718511  1.2573651
#> [274,] 13.383787  3.6139340  1.4325825  0.7258748  1.0912872  3.3487739
#> [275,]  9.680752 25.0355397  1.0893394  2.8909270  0.9439439  2.2408804
#> [276,]  9.348128  8.6907953  2.2205405  1.9605053  1.2518703  1.6604365
#> [277,] 11.548269  4.2143582  1.3170625  1.5943387  1.1129936  1.0723894
#> [278,]  3.498828  3.7915563  1.2368569  2.5531797  2.8402172  1.2368832
#> [279,]  8.383400  5.7061231  0.9131377  1.5187030  1.3084634  1.4597256
#> [280,]  9.923532  4.1409347  2.7346273  1.9361992  1.2841480 13.0218153
#> [281,]  7.170634 10.3804289  2.5564527  1.0658036  2.9112343  0.9518997
#> [282,]  7.781511  5.0554864  1.2429138  1.5954924  0.8819095  1.3181190
#> [283,] 12.565663  2.0080711  0.3827508  0.6045080  3.1068558  2.9322296
#> [284,]  6.082016  4.4895736  1.1609210  1.0546232  1.8093433  1.0440106
#> [285,] 14.527048  2.3535249  1.2303942  2.8472310  1.2215833  1.0930888
#> [286,] 14.346147 12.9882187  1.2618931  0.8313383  1.0101203  0.9730883
#> [287,] 20.540305  3.0863408  0.9181828  0.8386827  1.0903786  1.3642875
#> [288,] 14.341562  4.4286815  0.5038864  1.4867926  2.2954771  2.1851638
#> [289,] 28.126582  9.9949202  6.0412320 10.9598001 38.4944267 32.0107009
#> [290,] 15.285148  6.1613640  5.2592104 18.5741677 34.1553800  4.7702033
#> [291,] 27.674692  3.1470114  1.0035980  2.8966516  0.7936055  1.7777464
#> [292,] 11.124259  2.9950221  0.6689712  1.0368090  3.3296520  1.2733332
#> [293,] 17.669778  4.1984650  1.8133362  1.1715157  2.9127458  0.7264497
#> [294,] 10.167746  4.1483272  1.0385899  1.0562339  3.7481888  1.6319245
#> [295,] 10.229226  2.8876873  0.5630367  0.9330024  1.8414765  0.8654532
#> [296,] 20.657450  4.6905771  0.6391928  2.2102174 10.1759932  4.8365533
#> [297,]  9.317777  5.5553528  3.1541230  2.4742957  1.9812537  2.9078193
#> [298,]  7.520641  4.4312326  1.0888178  1.2263447  4.6602723  2.8056464
#> [299,]  7.558410  8.8682857  2.8794137  2.0744087  2.9845447  2.6526225
#> [300,] 16.208887 10.7765910  4.3036179  8.6516184  3.1760895  5.7603555
#> [301,]  9.480941  3.0489597  1.3646864  1.3704365  1.3022678  0.9038213
#> [302,] 15.887328  4.7407421  0.9484192  3.2655869  0.8903489  1.5520064
#> [303,]  9.234537  3.2673539  1.2546769  1.7217035  0.7612764  0.8246793
#> [304,] 11.712939  6.7645029  1.6540452  1.3136646  0.7581616  1.2430837
#> [305,] 15.176340  3.5568382  0.7160915  1.3170843  1.4652158  0.9592310
#>              [,7]       [,8]       [,9]      [,10]
#>   [1,]  2.3415001  7.4893614 11.2110148  8.1234370
#>   [2,]  3.1826269  1.7968865  9.2510307  4.7526406
#>   [3,]  1.3387818  1.7198894  0.8320187  1.3742558
#>   [4,]  2.6970959  2.6950753  2.4178303  3.1034855
#>   [5,]  2.3373365  2.3527209  1.1624259  3.2690445
#>   [6,]  4.2467138  2.9668525  1.7323234  1.5268766
#>   [7,]  1.9062683  1.9316592  3.5170165  2.3755894
#>   [8,]  5.3980658 10.5075987  4.8318862 18.6471911
#>   [9,]  5.2591202  3.2472387  1.6556999  2.8426988
#>  [10,]  5.2456448  6.1563636  7.5336457  8.3418094
#>  [11,]  1.6008157  3.2011595  2.1279175  2.7039439
#>  [12,]  5.1033795 11.0791110 11.7782409 12.0864933
#>  [13,]  8.6871444  2.1046957  2.8252110  3.3578607
#>  [14,]  3.5106520  3.9010280  3.1103102  3.8687350
#>  [15,]  7.7858358  5.1778780  1.8425097  4.7019527
#>  [16,]  1.0187545  2.9827474  0.8589277  7.1146195
#>  [17,]  3.3027173  2.4920611  1.0473366  3.5725846
#>  [18,]  4.4314179  1.1058496  0.7680624  0.9076610
#>  [19,]  1.1821486  3.2940953  1.5570659  1.1420806
#>  [20,]  0.8535254  1.2498620  0.9825800  1.5404265
#>  [21,]  0.5558366  1.0264316  1.0226483  3.2152564
#>  [22,]  2.1713684  4.3007940  1.2432047  1.5338818
#>  [23,]  7.7064436  2.5887581  4.3418135  2.2728456
#>  [24,]  3.1290817  3.5186375  4.2338274  3.0273347
#>  [25,]  1.2770344  3.8488242  1.2960154  3.3262293
#>  [26,]  2.3389371  2.3611178  4.1700208  1.6383643
#>  [27,]  1.4811166  1.3391552  0.9554221  1.6180328
#>  [28,]  1.7869244  4.7882923  1.5388854  1.7464328
#>  [29,]  1.7480614  1.6965239  0.5741370  3.2093820
#>  [30,]  1.2505822  2.5026001  0.8175732  0.7862394
#>  [31,]  1.8648541  0.9536943  0.5752134  2.7936549
#>  [32,]  1.4621210  1.1396565  1.6284531  4.1434428
#>  [33,]  0.9838924  0.5351241  1.1358916  2.8266744
#>  [34,]  2.4526680  0.7574229  1.1817379  1.2900291
#>  [35,]  0.6761836  0.8412302  0.9567912  0.7223612
#>  [36,]  0.5301857  1.0059901  2.1699404  3.2552568
#>  [37,]  1.5479969  1.0382648  2.3490052  0.8676341
#>  [38,]  1.7264857  0.8182328  1.0103184  1.5037059
#>  [39,]  5.0379545  1.9504381  6.1257319  5.4418869
#>  [40,]  4.9763572  1.8715425  3.2681304  3.7578927
#>  [41,]  1.0200514  1.7705153  1.3651261  0.9945586
#>  [42,]  1.5104604  2.0926169  1.2861100  2.2378746
#>  [43,]  1.2760708  3.1879404  3.5520384  2.0495652
#>  [44,]  2.0648854  3.4023879  1.0847720  1.0729633
#>  [45,]  1.3166943  1.3967378  1.0752454  1.2689422
#>  [46,]  1.7280558  2.3131472  1.0846049  6.6596555
#>  [47,]  1.2023475  1.6154298  1.5955520  1.1482435
#>  [48,]  4.2514072  5.7015531  6.0579393  4.4860994
#>  [49,]  1.1353581  1.1253869  2.6358050  2.1016713
#>  [50,]  1.2539194  1.4202286  4.5122070  0.6983475
#>  [51,]  1.2049896  2.0385090  1.7730068  2.1722599
#>  [52,]  2.4535134  1.2674944  0.9019123  1.3409712
#>  [53,]  0.6046385  1.0355418  2.3298119  1.4491666
#>  [54,]  1.0547051  0.3853754  1.7172387  1.7248528
#>  [55,]  1.0564484  2.5420502  1.3886243  2.0321004
#>  [56,]  3.0523121  1.2342415  0.8074031  1.0008819
#>  [57,]  0.4569504  0.8095495  1.3438307  0.9925709
#>  [58,]  2.4173831  2.0998019  0.7381086  3.1126670
#>  [59,]  3.7953559  1.2597054  2.4062843  1.4250067
#>  [60,]  0.6584260  1.6977442  1.3299095  1.3950243
#>  [61,]  0.9393376  1.1238485  1.1766994  0.6706058
#>  [62,]  0.8513514  0.8710488  2.3234755  0.9276342
#>  [63,]  0.5197791  0.9923360  1.2701943  0.9429059
#>  [64,]  1.3312475  1.4042630  1.9649656  4.3836026
#>  [65,]  2.5433928  1.6590791  3.1843765  3.6737334
#>  [66,]  2.0219892  3.0146319  0.7675730  1.5317459
#>  [67,]  1.6177077  0.6999290  2.2429106  1.9998846
#>  [68,]  2.8982336  1.2162285  2.0639783  2.7537742
#>  [69,]  1.4520392  2.2050429  1.2754600  0.9140076
#>  [70,]  5.9851966  1.2596677  1.7137605  2.2302972
#>  [71,]  2.4830742  2.5379569  0.5748443  0.7414340
#>  [72,]  2.6111519  3.9093333  1.7455938  5.5303644
#>  [73,]  5.0285950  3.0291041  0.8066405  2.7975940
#>  [74,]  1.0745581  2.1871268  1.0359180  0.9884797
#>  [75,]  1.7309023  1.6966419  0.5399507  1.4677420
#>  [76,]  3.5267868  0.4946718  3.0902736  4.1369115
#>  [77,]  0.9468949  3.0476411  0.9777629  2.1342244
#>  [78,]  0.7757864  1.6636331  2.1253513  2.2378394
#>  [79,]  1.0913168  2.5960123  0.9238992  0.8961133
#>  [80,]  0.8308397  1.3075381  0.8723790  2.4163946
#>  [81,]  1.8121313  1.0907163  1.5453088  6.3139248
#>  [82,]  5.5282218  3.1234360  2.0828074  3.4933445
#>  [83,]  0.8479781  1.8555277  0.9048337  1.2635755
#>  [84,]  5.6378032  1.5871500  1.3167145  1.5114239
#>  [85,]  1.3507172  2.4366283  0.8608911  1.0438954
#>  [86,]  1.5756674  1.6717920  1.0720607  2.1706831
#>  [87,]  2.6715983  0.7608377  1.8279470  1.8981235
#>  [88,]  1.0119524  1.9171605  1.9186813  1.3551452
#>  [89,]  1.2614254  1.0422766  1.9255349  1.1894300
#>  [90,]  3.9775383  5.6707769  1.5302263  1.5559302
#>  [91,]  1.3947019  2.9705039  1.2494707  1.9907019
#>  [92,]  2.2775064  3.1983376  1.8731899  2.4826679
#>  [93,]  0.7291500  0.7812243  1.0081691  0.7468382
#>  [94,]  1.6251965  0.9033444  1.0837066  1.3128087
#>  [95,]  0.9704001  0.6313154  0.6279667  1.0978390
#>  [96,]  8.8933998  3.3987684  2.4526514 14.5865961
#>  [97,]  1.9133283  0.9233048  0.9152533  1.0445042
#>  [98,]  1.6424482  3.1246335  1.1717736  4.0441854
#>  [99,]  1.0083090  1.8416191  0.9295848  1.2675050
#> [100,]  1.0341828  0.9896857  6.1691612  1.9616850
#> [101,]  1.9117311  0.8128552  1.4675969  0.9356983
#> [102,]  0.6483174  0.4996688  0.7313799  2.0019785
#> [103,]  2.8060112  1.0004905  4.1916957  0.7664564
#> [104,]  8.8277872  3.7922915  7.1681352  2.5799079
#> [105,]  1.7408976  0.6587230  0.7298601  1.2361507
#> [106,]  0.6408769  0.7904658  5.2795407  1.4936904
#> [107,]  1.4088199  1.5187771  1.8935264  1.6004839
#> [108,]  1.5545670  1.7225405  3.6620708  1.3200793
#> [109,] 31.8355474 11.7109096 12.5661346 33.1866935
#> [110,] 20.1619277 29.5323268 11.9119785 20.0412863
#> [111,]  0.9493172  0.9677353  0.8192423  2.2208483
#> [112,]  2.0593043  1.1613892  4.8905724  0.9562187
#> [113,]  0.5797909  0.7143316  2.8866886  0.6059696
#> [114,]  1.2222758  1.1191588  0.8046230  1.8941858
#> [115,]  0.3698881  1.1445196  2.1176721  1.4248653
#> [116,]  0.8385862  0.8320785  2.6300400  0.7848599
#> [117,]  2.1056259  1.8267788  0.9623168  0.9682265
#> [118,]  2.3431584  1.7103484  0.8237216  0.6996918
#> [119,]  0.9971897  1.7899733  3.0713773  2.3921271
#> [120,]  3.1299861  1.0693976  0.5829329  1.0228911
#> [121,]  3.1176619  2.6092364  4.4504071  1.9650791
#> [122,]  2.2553523  1.9199587  0.8503235  1.0453288
#> [123,]  1.5070252  1.5759666  1.1535004  0.7284568
#> [124,]  2.8071958  1.3121944  3.5927111  0.8691488
#> [125,]  0.6906891  2.6435900  0.4705176  0.7556476
#> [126,]  1.0690299  0.6862684  2.2407591  1.2816125
#> [127,]  0.8165921  1.3716016  6.7417696  1.2043423
#> [128,]  0.7191528  2.4670535  1.1791429  0.7166517
#> [129,]  5.6551074  4.4345142  5.4874121  8.7774891
#> [130,]  1.0013925  0.8636381  0.7076679  1.0793104
#> [131,]  0.8055067  1.4645814  0.8551056  1.1355788
#> [132,]  1.1730206  1.3066400  2.8908994  2.8873994
#> [133,]  1.0955021  2.4866665  1.5649720  1.2797941
#> [134,]  3.0479542  1.4576714  0.8557142  0.8599238
#> [135,]  2.1843263  6.7836008  2.6829872  1.7013506
#> [136,]  1.5918263  4.1602587  1.5021943  0.9716938
#> [137,]  0.5296510  1.4760183  1.4004104  1.2702927
#> [138,]  5.3252703  2.1253650  1.5124304  1.6294767
#> [139,]  1.6743533  3.7155096  1.9164634  1.1662834
#> [140,]  0.9581860  3.3288749  0.8074472  0.6524526
#> [141,]  1.5080193  0.6546719  0.3716355  5.0580106
#> [142,]  3.4671387  3.0943603  5.9371174  1.8994516
#> [143,]  2.1365626  3.5648417  4.6759052  2.2240740
#> [144,]  0.6495214  2.6774581  3.2783741  1.5990783
#> [145,]  0.8787234  1.7936330  2.2737180  1.5080527
#> [146,]  1.2005082  1.1219380  1.2310323  0.9239552
#> [147,]  3.0455751  0.7167383  5.1078067  1.3368994
#> [148,]  5.0359723  3.2110582  2.7683537  1.6587287
#> [149,]  4.0237889  5.4168069  4.4743208  3.4807637
#> [150,]  1.9270523  2.8087582  1.4164444  1.4617337
#> [151,]  0.8157876  1.0619383  0.4601600  4.7000811
#> [152,]  1.1136071  0.9378582  0.9094329  1.3242560
#> [153,]  2.0861836  1.4085971  1.6588648  2.4488005
#> [154,]  2.9140149  0.9683215  2.4691364  0.9293384
#> [155,]  1.2195468  1.1644439  0.4719928  1.3402463
#> [156,]  1.2996439  1.3313741  1.9383597  1.0404663
#> [157,]  2.1972917  1.8706652  1.7005008  2.8782307
#> [158,]  1.5552319  1.5801760  0.7063645  2.1218941
#> [159,]  0.7351606  1.2875160  1.2781777  1.2866245
#> [160,]  1.7937663  1.3260785  0.4144033  0.7674795
#> [161,]  1.4478747  1.0848752  0.7801718  0.9308238
#> [162,]  2.0726101  0.6723774  0.7649978  1.5101436
#> [163,]  2.0039650  1.8266475  1.4672148  2.4161405
#> [164,]  0.7668633  1.1161346  1.0106481  1.5308692
#> [165,]  1.6408755  1.6521164  1.1107713  6.1949464
#> [166,]  0.6371549  1.0144322  1.8379791  0.5784724
#> [167,]  0.7415574  0.7855808  0.7227636  2.1447880
#> [168,]  1.5400760  0.4501778  2.3637598  0.6947130
#> [169,]  1.9970619  1.6346129  0.9833172  1.4821201
#> [170,]  0.4145906  1.3530231  1.5378027  2.3130331
#> [171,]  1.0930764  2.6695889  1.5498024  2.8121311
#> [172,]  1.4991412  1.7134883  2.1621933  1.2451817
#> [173,]  0.5372935  0.9197238  0.8333294  3.1110742
#> [174,]  0.7630561  1.6401851  1.5132498  1.1619072
#> [175,]  0.5531969  2.0745806  0.9006923  1.5443960
#> [176,]  1.3018247  0.9559260  1.5160713  1.2335986
#> [177,]  0.6369325  2.0732906  0.9847347  0.5702898
#> [178,]  0.7046688  0.5106173  0.6663404  1.5913334
#> [179,]  0.8385218  1.0575477  1.0443783  1.3983063
#> [180,]  2.4890672  2.0496295  0.5962440  1.7092036
#> [181,]  1.4521847  2.6934231  1.4908962  1.3177870
#> [182,]  1.0377939  0.6219170  0.6792214  0.7271690
#> [183,]  0.7639578  1.3529518  6.4261734  1.0368138
#> [184,]  2.1291530  4.4576256  0.7010442  0.8404893
#> [185,]  1.5287573  0.8625785  1.2706959  0.6034819
#> [186,]  0.9267296  0.7152325  1.6362001  1.1203894
#> [187,]  1.7719937  1.6666938  2.0428624  1.8310549
#> [188,]  1.3018628  1.1646188  1.0853115  1.2506457
#> [189,]  1.1683074  1.3835158  0.7969504  0.5734844
#> [190,]  2.3456568  1.1699989  1.1167540  1.2413569
#> [191,]  2.0718439  1.3919508  0.8502048  1.6032460
#> [192,]  0.8436236  1.3838807  4.0629369  1.2433345
#> [193,]  2.3554549  1.1647091  1.2185362  2.3940837
#> [194,]  1.5099853  1.3860817  3.0166990  0.8515235
#> [195,]  0.6659609  1.6342542  1.7924782  3.5465629
#> [196,]  1.7823393  1.7305359  1.2023177  3.0875538
#> [197,]  1.8953411  0.8330563  1.8062405  3.1876673
#> [198,]  2.2775879  0.8130346  2.2706503  1.3783272
#> [199,]  1.0497419  2.6810538  1.1386936  2.4868409
#> [200,]  2.7483070  2.6269119  1.2900284  1.2215116
#> [201,]  2.6318615  1.5323330  1.1723786  2.8203156
#> [202,]  1.5002320  4.7930295  3.0763837  1.9087832
#> [203,]  1.2003423  2.7615985  1.2886081 12.9061553
#> [204,]  0.9924778  1.6903053  1.2549087  1.5761897
#> [205,]  1.6506259  2.1124954  1.2191333  2.8450907
#> [206,]  1.7970673  2.9131803  1.6512740  2.7395048
#> [207,]  1.6221480  1.3501281  2.2718036  0.9824223
#> [208,]  0.9589544  1.6693703  3.0588762  1.7865265
#> [209,]  2.1194652  7.2382104  2.2020465  1.4249685
#> [210,]  2.1353208  2.5289008  2.1393723  1.1801777
#> [211,]  1.1277402  1.1337312  0.7545932  1.5808938
#> [212,]  1.5739135  8.8428791  1.8698844  1.2205826
#> [213,]  1.3803468  1.7541838  1.3255235  1.3403912
#> [214,]  6.2790664 11.3521304  8.9087103 34.7234209
#> [215,]  4.9281613  8.1399449  2.5329383  6.2721877
#> [216,]  4.3345653  8.5762499  9.0588660  5.3224423
#> [217,]  1.7964400  1.3229962  3.7552081  1.5344761
#> [218,]  1.1398929  2.2773786  1.6807184  1.0072043
#> [219,]  2.4819721  1.5442125  1.8979645  7.7568175
#> [220,]  1.8447024  1.0493193  1.3670005  0.8205486
#> [221,]  4.6789095  2.8036385  1.4484284  0.7867616
#> [222,] 14.1839185  2.9270890  6.7098536  7.0600694
#> [223,]  0.7441756  0.6075243  4.2276576  1.5502457
#> [224,]  0.9176352  1.3032329  0.8822288  0.9779032
#> [225,]  3.1586917  1.3500266  2.1552448  0.8851910
#> [226,]  0.7654312  2.3568056  0.7175080  0.7577741
#> [227,]  0.9678811  1.4983960  4.6597513  1.3166176
#> [228,]  2.1875623  2.7609190  1.0923611  1.2008864
#> [229,]  0.9373141  1.1060812  0.9523887  1.2979901
#> [230,]  0.7614265  0.7119504  3.2385836  1.9104954
#> [231,]  0.8229277  0.7313150  1.1575819  0.8245097
#> [232,]  1.7833843  1.0070167  1.5205773  2.3768661
#> [233,]  0.7472971  0.5521115  0.5720601  1.3774872
#> [234,]  0.6270519  2.2093122  0.5116519  1.6416116
#> [235,]  2.2764545  3.4139194  0.8085193  1.3731734
#> [236,]  1.4151235  1.6251976  0.9602400  1.0496642
#> [237,]  0.9841697  2.7440725  1.6268727  0.8421290
#> [238,]  1.2500820  0.8773787  1.8099017  1.4510969
#> [239,]  2.3729390  1.7705449  6.0856355  0.9099196
#> [240,]  3.0290946  1.4619842  0.7428085  1.7975612
#> [241,]  1.6742828  1.0436127  2.4315470  1.7082266
#> [242,]  2.2142448  2.0556654  2.2198099  1.3662998
#> [243,]  3.8787536  3.9613308  2.9476879  5.2309219
#> [244,] 15.8678559  8.2627815 24.2074001 27.7616716
#> [245,]  3.1865849  5.7876636  1.8015688  3.4382206
#> [246,]  2.1623149  2.3142099  1.9454723  0.6110268
#> [247,]  2.0296914  1.1995543  0.8353126  1.0075106
#> [248,]  0.6150199  3.1937608  1.8271973  0.7528552
#> [249,]  0.8791020  0.8603991  1.1624473  0.9408011
#> [250,]  0.7286660  1.1069041  1.0523171  0.4404371
#> [251,]  0.6640559  0.8340013  0.6850973  1.1956067
#> [252,]  1.7518932  2.4716101  1.5422812  1.5534447
#> [253,]  1.9994698  1.4528572  0.7705521  1.8535544
#> [254,]  1.4332337  1.8559056  0.8623139  1.3928204
#> [255,]  0.4859360  0.9941720  1.3214766  1.1975831
#> [256,]  1.0322958  4.1301929  0.4986110  0.8528642
#> [257,]  0.4669130  1.1962713  0.8140313  0.6380277
#> [258,]  1.3375552  0.6238969  2.2438175  1.6736400
#> [259,]  1.9438840  1.1595410  1.5409537  1.6832978
#> [260,]  2.8709043  4.5895939  1.3425327  3.0568136
#> [261,]  2.0430483  0.7811925  1.4247175  1.1711416
#> [262,]  1.8474347  0.9487345  2.3538983  1.1244110
#> [263,]  1.3785498  0.7117512  2.1155760  2.4294004
#> [264,]  1.2173727  1.3815548  0.7587186  4.7582095
#> [265,]  3.7241153  2.2234262  3.2632347  3.3342052
#> [266,]  0.7288202  0.7746363  0.7748841  1.1816217
#> [267,]  0.7003731  0.9031714  0.9061938  1.0413277
#> [268,]  2.0210024  2.1863767  1.5242586  1.3555153
#> [269,]  0.7569566  2.9607167  1.2377341  1.1253220
#> [270,]  0.8292548  2.5339409  1.6423977  4.1832278
#> [271,]  1.7106662  1.5307210  1.3458098  1.5623383
#> [272,]  2.8363021  1.2615194  1.2195306  1.6053048
#> [273,]  1.5385013  0.9861371  1.5769241  1.1893402
#> [274,]  1.6915673  1.0129580  1.5013262  1.8644129
#> [275,]  1.5353378  1.1506190  1.3891384  1.6814040
#> [276,]  0.7099133  4.0893365  0.9511181  1.8470765
#> [277,]  2.0974409  0.9043379  1.6216094  0.9613077
#> [278,]  1.4952030  3.1681290  0.9149304  2.7692301
#> [279,]  1.1216580  1.6546944  2.0350060  1.5276622
#> [280,]  0.5813503  2.4286158  0.9134888  1.6558409
#> [281,]  3.1802511  1.2152296  2.3518065  1.7793500
#> [282,]  1.1567983  2.3017859  2.0790534  1.2012739
#> [283,]  2.1787578  2.1415029  0.9948561  1.4077223
#> [284,]  3.1351118  1.7052525  1.0696248  3.5179789
#> [285,]  1.4927539  1.7827065  0.9698161  0.7592638
#> [286,]  1.9184784  1.5858361  1.9545726  3.6829838
#> [287,]  2.1834221  1.2298935  1.0885195  1.9548456
#> [288,]  1.4585959  2.7977852  1.6033126  1.8018685
#> [289,] 17.2573496 21.1558685 45.4310531 16.8445820
#> [290,]  8.3201035 22.8222024 33.0440867  7.6550862
#> [291,]  1.1739540  4.0103705  6.2315360  1.0459680
#> [292,]  4.0954270  0.6843976  1.2717089  2.4923381
#> [293,]  1.4855277  8.1160536  1.0433661  1.4002874
#> [294,]  1.1491367  1.5889719  2.1476266  1.9908983
#> [295,]  3.5684956  1.2145161  3.4440027  1.2045218
#> [296,]  3.0781088  7.9613945  4.1040867  4.1226806
#> [297,]  2.5261773  2.1548747  6.0270674  2.3309869
#> [298,]  2.8824453  1.2904444  2.8383046  1.5587757
#> [299,]  1.7668434  4.1983119  2.0883382  1.6597510
#> [300,] 14.0969407  3.3741700  3.7918649  9.5738466
#> [301,]  1.5632897  0.9095542  1.0876947  3.0443921
#> [302,]  2.0614846  1.4496520  2.3482101  1.9878968
#> [303,]  1.9468233  0.5486979  2.8865077  0.9001356
#> [304,]  0.9799658  1.9909902  3.3504475  1.3343700
#> [305,]  2.6290442  1.4088652  1.3510104  2.2565492
#> 
#> $df
#>           [,1]
#>  [1,] 6.000000
#>  [2,] 5.689735
#>  [3,] 5.689735
#>  [4,] 5.689735
#>  [5,] 5.589529
#>  [6,] 5.589529
#>  [7,] 5.589529
#>  [8,] 5.482226
#>  [9,] 5.236727
#> [10,] 5.338465
#> 

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$get_last_draw`
## ------------------------------------------------

data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
set.seed(123)
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$is_normalised`
## ------------------------------------------------

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.

# estimate the model
set.seed(123)
posterior      = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# check normalisation status beforehand
posterior$is_normalised()
#> [1] TRUE

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) %*% BB         # set negative diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()
#> [1] TRUE

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$set_normalised`
## ------------------------------------------------

# This is an internal function that is run while executing normalise_posterior()
# Observe its working by analysing the workflow:

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# check normalisation status beforehand
posterior$is_normalised()
#> [1] TRUE

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) %*% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()
#> [1] TRUE