Provides posterior summary of historical decompositions

Provides posterior means of the historical decompositions variable by variable.

Usage

# S3 method for class 'PosteriorHD'
summary(object, ...)

Arguments

object

an object of class PosteriorHD obtained using the compute_historical_decompositions() function containing posterior draws of historical decompositions.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior means of historical decompositions for each of the variables.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

specification  = specify_bsvar$new(diff(us_fiscal_lsuw))
#> The identification is set to the default option of lower-triangular structural matrix.
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# compute historical decompositions
hds            = compute_historical_decompositions(posterior)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Computing historical decomposition               |
#> **************************************************|
#>  This might take a little while :)                
#> **************************************************|
hds_summary    = summary(hds)
head(hds_summary$gdp) # browse the contributions
#>       shock 1    shock 2     shock 3
#> 1 -0.46028004 -2.0855778 -0.01336049
#> 2  0.75343854 -2.1906277  1.15052584
#> 3 -0.07668000 -1.6888570  1.09637179
#> 4 -0.23923108 -1.5991496  0.41607436
#> 5 -0.08556427 -0.8840088 -1.43091984
#> 6 -0.06797145 -1.0653542 -1.12912439

# workflow with the pipe |>
############################################################
diff(us_fiscal_lsuw) |>
  specify_bsvar$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_historical_decompositions() |>
  summary() -> hds_summary
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Computing historical decomposition               |
#> **************************************************|
#>  This might take a little while :)                
#> **************************************************|
hds_summary$gdp # browse the contributions
#>           shock 1      shock 2       shock 3
#> 1   -1.879205e+00 1.568105e+00  7.622971e-01
#> 2   -1.537728e+00 1.594907e+00  1.044380e+00
#> 3   -1.742927e+00 2.685976e+00  3.415494e-01
#> 4   -1.848000e+00 3.520329e+00  6.086634e-01
#> 5   -3.120549e+00 5.312802e+00 -6.227198e-01
#> 6   -5.764692e+00 7.107249e+00 -4.587101e-01
#> 7   -2.781037e+00 8.655068e+00  1.796655e-01
#> 8   -2.519092e+00 1.076667e+01  1.026693e+00
#> 9   -7.063116e-01 1.373134e+01 -1.068081e+00
#> 10   7.885005e-01 1.652230e+01  3.984702e+00
#> 11   3.002484e+00 2.106304e+01  3.339569e+00
#> 12   3.437788e+00 2.740084e+01  3.053580e+00
#> 13   4.598142e+00 3.543749e+01  3.319056e+00
#> 14   6.367971e+00 4.626440e+01 -6.494917e-01
#> 15   8.797088e+00 5.992611e+01  3.107514e-01
#> 16   1.004189e+01 7.830266e+01  4.953978e-01
#> 17   1.040925e+01 1.030253e+02  1.752803e-01
#> 18   1.166673e+01 1.354735e+02  2.422396e+00
#> 19   1.068583e+01 1.793892e+02  8.694885e-01
#> 20   9.609288e+00 2.375754e+02  3.005813e+00
#> 21   6.344687e+00 3.164308e+02  3.594153e-01
#> 22   1.317877e+00 4.214615e+02  4.061405e+00
#> 23  -8.294234e+00 5.639882e+02  4.143537e+00
#> 24  -2.016903e+01 7.551949e+02  7.988110e+00
#> 25  -3.871722e+01 1.013895e+03  1.131083e+01
#> 26  -6.528003e+01 1.363511e+03  1.732325e+01
#> 27  -1.032902e+02 1.837351e+03  2.262900e+01
#> 28  -1.592306e+02 2.479889e+03  3.126896e+01
#> 29  -2.395255e+02 3.352475e+03  4.294550e+01
#> 30  -3.540826e+02 4.538717e+03  5.662076e+01
#> 31  -5.150375e+02 6.151663e+03  7.885205e+01
#> 32  -7.418243e+02 8.347783e+03  1.079567e+02
#> 33  -1.060652e+03 1.134010e+04  1.430390e+02
#> 34  -1.505258e+03 1.541817e+04  1.987157e+02
#> 35  -2.127086e+03 2.098198e+04  2.645969e+02
#> 36  -2.992895e+03 2.857467e+04  3.580822e+02
#> 37  -4.195906e+03 3.894191e+04  4.852436e+02
#> 38  -5.866522e+03 5.310424e+04  6.543368e+02
#> 39  -8.181347e+03 7.245841e+04  8.821587e+02
#> 40  -1.138453e+04 9.891665e+04  1.197970e+03
#> 41  -1.581488e+04 1.351012e+05  1.614780e+03
#> 42  -2.193881e+04 1.846008e+05  2.186772e+03
#> 43  -3.039387e+04 2.523360e+05  2.953378e+03
#> 44  -4.206566e+04 3.450481e+05  4.001625e+03
#> 45  -5.816725e+04 4.719799e+05  5.414559e+03
#> 46  -8.036759e+04 6.457983e+05  7.337783e+03
#> 47  -1.109633e+05 8.838724e+05  9.945609e+03
#> 48  -1.531213e+05 1.210017e+06  1.349242e+04
#> 49  -2.111851e+05 1.656891e+06  1.831236e+04
#> 50  -2.911361e+05 2.269282e+06  2.486443e+04
#> 51  -4.011936e+05 3.108619e+06  3.378534e+04
#> 52  -5.526632e+05 4.259165e+06  4.592921e+04
#> 53  -7.610843e+05 5.836510e+06  6.247220e+04
#> 54  -1.047817e+06 7.999226e+06  8.501654e+04
#> 55  -1.442221e+06 1.096487e+07  1.157561e+05
#> 56  -1.984650e+06 1.503195e+07  1.576842e+05
#> 57  -2.730559e+06 2.061005e+07  2.149049e+05
#> 58  -3.756157e+06 2.826120e+07  2.930148e+05
#> 59  -5.166159e+06 3.875666e+07  3.996908e+05
#> 60  -7.104455e+06 5.315486e+07  5.454237e+05
#> 61  -9.768741e+06 7.290836e+07  7.445826e+05
#> 62  -1.343064e+07 1.000108e+08  1.016828e+06
#> 63  -1.846330e+07 1.371982e+08  1.389108e+06
#> 64  -2.537940e+07 1.882262e+08  1.898302e+06
#> 65  -3.488317e+07 2.582495e+08  2.594956e+06
#> 66  -4.794206e+07 3.543435e+08  3.548307e+06
#> 67  -6.588500e+07 4.862204e+08  4.853246e+06
#> 68  -9.053745e+07 6.672123e+08  6.639810e+06
#> 69  -1.244068e+08 9.156199e+08  9.086255e+06
#> 70  -1.709372e+08 1.256566e+09  1.243693e+07
#> 71  -2.348591e+08 1.724539e+09  1.702686e+07
#> 72  -3.226698e+08 2.366884e+09  2.331541e+07
#> 73  -4.432932e+08 3.248598e+09  3.193253e+07
#> 74  -6.089858e+08 4.458912e+09  4.374215e+07
#> 75  -8.365808e+08 6.120329e+09  5.992919e+07
#> 76  -1.149197e+09 8.401034e+09  8.211899e+07
#> 77  -1.578585e+09 1.153193e+10  1.125411e+08
#> 78  -2.168351e+09 1.583001e+10  1.542543e+08
#> 79  -2.978381e+09 2.173053e+10  2.114549e+08
#> 80  -4.090915e+09 2.983103e+10  2.899006e+08
#> 81  -5.618901e+09 4.095194e+10  3.974914e+08
#> 82  -7.717446e+09 5.621967e+10  5.450681e+08
#> 83  -1.059956e+10 7.718080e+10  7.475064e+08
#> 84  -1.455776e+10 1.059588e+11  1.025221e+09
#> 85  -1.999375e+10 1.454690e+11  1.406228e+09
#> 86  -2.745920e+10 1.997147e+11  1.928979e+09
#> 87  -3.771165e+10 2.741919e+11  2.646245e+09
#> 88  -5.179141e+10 3.764474e+11  3.630460e+09
#> 89  -7.112706e+10 5.168427e+11  4.981041e+09
#> 90  -9.768039e+10 7.096051e+11  6.834451e+09
#> 91  -1.341453e+11 9.742693e+11  9.378003e+09
#> 92  -1.842213e+11 1.337657e+12  1.286882e+10
#> 93  -2.529882e+11 1.836598e+12  1.765986e+10
#> 94  -3.474220e+11 2.521660e+12  2.423565e+10
#> 95  -4.771019e+11 3.462278e+12  3.326131e+10
#> 96  -6.551822e+11 4.753789e+12  4.564994e+10
#> 97  -8.997259e+11 6.527103e+12  6.265506e+10
#> 98  -1.235537e+12 8.961967e+12  8.599755e+10
#> 99  -1.696677e+12 1.230519e+13  1.180400e+11
#> 100 -2.329916e+12 1.689567e+13  1.620258e+11
#> 101 -3.199480e+12 2.319873e+13  2.224080e+11
#> 102 -4.393560e+12 3.185333e+13  3.053002e+11
#> 103 -6.033260e+12 4.373681e+13  4.190960e+11
#> 104 -8.284874e+12 6.005384e+13  5.753193e+11
#> 105 -1.137675e+13 8.245859e+13  7.897921e+11
#> 106 -1.562245e+13 1.132224e+14  1.084237e+12
#> 107 -2.145255e+13 1.554640e+14  1.488481e+12
#> 108 -2.945828e+13 2.134659e+14  2.043472e+12
#> 109 -4.045152e+13 2.931084e+14  2.805437e+12
#> 110 -5.554707e+13 4.024657e+14  3.851573e+12
#> 111 -7.627577e+13 5.526249e+14  5.287874e+12
#> 112 -1.047396e+14 7.588095e+14  7.259874e+12
#> 113 -1.438251e+14 1.041924e+15  9.967396e+12
#> 114 -1.974957e+14 1.430671e+15  1.368481e+13
#> 115 -2.711939e+14 1.964465e+15  1.878882e+13
#> 116 -3.723930e+14 2.697425e+15  2.579671e+13
#> 117 -5.113549e+14 3.703864e+15  3.541868e+13
#> 118 -7.021708e+14 5.085823e+15  4.862994e+13
#> 119 -9.641900e+14 6.983416e+15  6.676948e+13
#> 120 -1.323982e+15 9.589038e+15  9.167588e+13
#> 121 -1.818029e+15 1.316687e+16  1.258736e+14
#> 122 -2.496430e+15 1.807967e+16  1.728291e+14
#> 123 -3.427974e+15 2.482555e+16  2.373018e+14
#> 124 -4.707119e+15 3.408847e+16  3.258272e+14
#> 125 -6.463572e+15 4.680762e+16  4.473789e+14
#> 126 -8.875436e+15 6.427259e+16  6.142787e+14
#> 127 -1.218727e+16 8.825419e+16  8.434457e+14
#> 128 -1.673490e+16 1.211839e+17  1.158111e+15
#> 129 -2.297943e+16 1.664007e+17  1.590175e+15
#> 130 -3.155407e+16 2.284890e+17  2.183438e+15
#> 131 -4.332825e+16 3.137442e+17  2.998043e+15
#> 132 -5.949586e+16 4.308106e+17  4.116576e+15
#> 133 -8.169627e+16 5.915578e+17  5.652434e+15
#> 134 -1.121805e+17 8.122844e+17  7.761324e+15
#> 135 -1.540397e+17 1.115371e+18  1.065705e+16
#> 136 -2.115181e+17 1.531547e+18  1.463319e+16
#> 137 -2.904440e+17 2.103012e+18  2.009287e+16
#> 138 -3.988201e+17 2.887708e+18  2.758961e+16
#> 139 -5.476356e+17 3.965198e+18  3.788347e+16
#> 140 -7.519797e+17 5.444733e+18  5.201812e+16
#> 141 -1.032572e+18 7.476329e+18  7.142662e+16
#> 142 -1.417865e+18 1.026598e+19  9.807676e+16
#> 143 -1.946924e+18 1.409653e+19  1.346705e+17
#> 144 -2.673394e+18 1.935639e+19  1.849182e+17
#> 145 -3.670938e+18 2.657886e+19  2.539142e+17
#> 146 -5.040702e+18 3.649628e+19  3.486541e+17
#> 147 -6.921574e+18 5.011420e+19  4.787436e+17
#> 148 -9.504268e+18 6.881341e+19  6.573725e+17
#> 149 -1.305066e+19 9.448992e+19  9.026520e+17
#> 150 -1.792034e+19 1.297472e+20  1.239451e+18
#> 151 -2.460707e+19 1.781600e+20  1.701919e+18
#> 152 -3.378885e+19 2.446374e+20  2.336946e+18
#> 153 -4.639669e+19 3.359196e+20  3.208918e+18
#> 154 -6.370897e+19 4.612622e+20  4.406246e+18
#> 155 -8.748107e+19 6.333743e+20  6.050330e+18
#> 156 -1.201234e+20 8.697072e+20  8.307868e+18
#> 157 -1.649457e+20 1.194224e+21  1.140776e+19
#> 158 -2.264928e+20 1.639828e+21  1.566431e+19
#> 159 -3.110053e+20 2.251703e+21  2.150910e+19
#> 160 -4.270524e+20 3.091889e+21  2.953476e+19
#> 161 -5.864007e+20 4.245576e+21  4.055503e+19
#> 162 -8.052074e+20 5.829743e+21  5.568731e+19
#> 163 -1.105659e+21 8.005016e+21  7.646589e+19
#> 164 -1.518219e+21 1.099196e+22  1.049976e+20
#> 165 -2.084719e+21 1.509343e+22  1.441754e+20
#> 166 -2.862601e+21 2.072529e+22  1.979717e+20
#> 167 -3.930738e+21 2.845860e+22  2.718410e+20
#> 168 -5.397433e+21 3.907746e+22  3.732733e+20
#> 169 -7.411404e+21 5.365859e+22  5.125533e+20
#> 170 -1.017686e+22 7.368043e+22  7.038032e+20
#> 171 -1.397420e+22 1.011731e+23  9.664145e+20
#> 172 -1.918845e+22 1.389242e+23  1.327015e+21
#> 173 -2.634833e+22 1.907616e+23  1.822166e+21
#> 174 -3.617981e+22 2.619412e+23  2.502075e+21
#> 175 -4.967975e+22 3.596804e+23  3.435681e+21
#> 176 -6.821699e+22 4.938893e+23  4.717645e+21
#> 177 -9.367112e+22 6.781763e+23  6.477953e+21
#> 178 -1.286231e+23 9.312270e+23  8.895090e+21
#> 179 -1.766168e+23 1.278700e+24  1.221414e+22
#> 180 -2.425186e+23 1.755826e+24  1.677164e+22
#> 181 -3.330107e+23 2.410985e+24  2.302970e+22
#> 182 -4.572684e+23 3.310605e+24  3.162284e+22
#> 183 -6.278911e+23 4.545905e+24  4.342237e+22
#> 184 -8.621789e+23 6.242138e+24  5.962471e+22
#> 185 -1.183888e+24 8.571293e+24  8.187268e+22
#> 186 -1.625637e+24 1.176954e+25  1.124221e+23
#> 187 -2.232218e+24 1.616115e+25  1.543706e+23
#> 188 -3.065136e+24 2.219143e+25  2.119715e+23
#> 189 -4.208844e+24 3.047182e+25  2.910651e+23
#> 190 -5.779308e+24 4.184189e+25  3.996714e+23
#> 191 -7.935767e+24 5.745454e+25  5.488022e+23
#> 192 -1.089688e+25 7.889280e+25  7.535789e+23
#> 193 -1.496287e+25 1.083304e+26  1.034765e+24
#> 194 -2.054604e+25 1.487522e+26  1.420871e+24
#> 195 -2.821247e+25 2.042567e+26  1.951045e+24
#> 196 -3.873952e+25 2.804719e+26  2.679047e+24
#> 197 -5.319457e+25 3.851256e+26  3.678690e+24
#> 198 -7.304329e+25 5.288292e+26  5.051335e+24
#> 199 -1.002983e+26 7.261535e+26  6.936159e+24
#> 200 -1.377230e+26 9.971063e+26  9.524277e+24
#> 201 -1.891122e+26 1.369161e+27  1.307811e+25
#> 202 -2.596764e+26 1.880042e+27  1.795800e+25
#> 203 -3.565706e+26 2.581550e+27  2.465874e+25
#> 204 -4.896194e+26 3.544815e+27  3.385976e+25
#> 205 -6.723132e+26 4.867506e+27  4.649399e+25
#> 206 -9.231764e+26 6.683740e+27  6.384248e+25
#> 207 -1.267645e+27 9.177673e+27  8.766429e+25
#> 208 -1.740647e+27 1.260218e+28  1.203748e+26
#> 209 -2.390143e+27 1.730448e+28  1.652908e+26
#> 210 -3.281987e+27 2.376138e+28  2.269665e+26
#> 211 -4.506609e+27 3.262757e+28  3.116554e+26
#> 212 -6.188180e+27 4.480203e+28  4.279447e+26
#> 213 -8.497203e+27 6.151921e+28  5.876255e+26
#> 214 -1.166780e+28 8.447414e+28  8.068888e+26
#> 215 -1.602146e+28 1.159943e+29  1.107967e+27
#> 216 -2.199962e+28 1.592758e+29  1.521387e+27
#> 217 -3.020843e+28 2.187071e+29  2.089068e+27
#> 218 -4.148023e+28 3.003142e+29  2.868571e+27
#> 219 -5.695793e+28 4.123717e+29  3.938934e+27
#> 220 -7.821089e+28 5.662418e+29  5.408685e+27
#> 221 -1.073941e+29 7.775261e+29  7.426851e+27
#> 222 -1.474665e+29 1.067648e+30  1.019806e+28
#> 223 -2.024913e+29 1.466024e+30  1.400331e+28
#> 224 -2.780477e+29 2.013048e+30  1.922843e+28
#> 225 -3.817969e+29 2.764185e+30  2.640321e+28
#> 226 -5.242585e+29 3.795597e+30  3.625516e+28
#> 227 -7.198773e+29 5.211865e+30  4.978320e+28
#> 228 -9.884882e+29 7.156591e+30  6.835902e+28
#> 229 -1.357327e+30 9.826960e+30  9.386612e+28
#> 230 -1.863793e+30 1.349374e+31  1.288908e+29
#> 231 -2.559238e+30 1.852871e+31  1.769844e+29
#> 232 -3.514177e+30 2.544241e+31  2.430233e+29
#> 233 -4.825437e+30 3.493585e+31  3.337036e+29
#> 234 -6.625974e+30 4.797162e+31  4.582199e+29
#> 235 -9.098352e+30 6.587147e+31  6.291975e+29
#> 236 -1.249326e+31 9.045038e+31  8.639727e+29
#> 237 -1.715492e+31 1.242005e+32  1.186351e+30
#> 238 -2.355602e+31 1.705440e+32  1.629019e+30
#> 239 -3.234558e+31 2.341798e+32  2.236862e+30
#> 240 -4.441482e+31 3.215604e+32  3.071511e+30
#> 241 -6.098752e+31 4.415456e+32  4.217598e+30
#> 242 -8.374405e+31 6.063014e+32  5.791328e+30
#> 243 -1.149918e+32 8.325333e+32  7.952272e+30
#> 244 -1.578992e+32 1.143180e+33  1.091954e+31
#> 245 -2.168169e+32 1.569740e+33  1.499399e+31
#> 246 -2.977187e+32 2.155464e+33  2.058877e+31
#> 247 -4.088078e+32 2.959741e+33  2.827114e+31
#> 248 -5.613480e+32 4.064123e+33  3.882007e+31
#> 249 -7.708062e+32 5.580586e+33  5.330518e+31
#> 250 -1.058421e+33 7.662895e+33  7.319517e+31
#> 251 -1.453354e+33 1.052218e+34  1.005068e+32
#> 252 -1.995650e+33 1.444837e+34  1.380093e+32
#> 253 -2.740295e+33 1.983956e+34  1.895054e+32
#> 254 -3.762793e+33 2.724238e+34  2.602163e+32
#> 255 -5.166821e+33 3.740744e+34  3.573120e+32
#> 256 -7.094739e+33 5.136544e+34  4.906373e+32
#> 257 -9.742030e+33 7.053166e+34  6.737110e+32
#> 258 -1.337712e+34 9.684944e+34  9.250957e+32
#> 259 -1.836858e+34 1.329873e+35  1.270281e+33
#> 260 -2.522253e+34 1.826094e+35  1.744266e+33
#> 261 -3.463392e+34 2.507473e+35  2.395112e+33
#> 262 -4.755702e+34 3.443097e+35  3.288810e+33
#> 263 -6.530218e+34 4.727835e+35  4.515978e+33
#> 264 -8.966866e+34 6.491952e+35  6.201045e+33
#> 265 -1.231271e+35 8.914322e+35  8.514867e+33
#> 266 -1.690701e+35 1.224056e+36  1.169206e+34
#> 267 -2.321559e+35 1.680794e+36  1.605476e+34
#> 268 -3.187813e+35 2.307956e+36  2.204535e+34
#> 269 -4.377295e+35 3.169133e+36  3.027123e+34
#> 270 -6.010615e+35 4.351645e+36  4.156646e+34
#> 271 -8.253382e+35 5.975394e+36  5.707633e+34
#> 272 -1.133300e+36 8.205019e+36  7.837348e+34
#> 273 -1.556173e+36 1.126659e+37  1.076173e+35
#> 274 -2.136835e+36 1.547055e+37  1.477730e+35
#> 275 -2.934162e+36 2.124314e+37  2.029122e+35
#> 276 -4.028998e+36 2.916968e+37  2.786257e+35
#> 277 -5.532356e+36 4.005389e+37  3.825906e+35
#> 278 -7.596668e+36 5.499938e+37  5.253483e+35
#> 279 -1.043125e+37 7.552153e+37  7.213737e+35
#> 280 -1.432350e+37 1.037012e+38  9.905431e+35
#> 281 -1.966809e+37 1.423957e+38  1.360149e+36
#> 282 -2.700693e+37 1.955284e+38  1.867667e+36
#> 283 -3.708415e+37 2.684868e+38  2.564558e+36
#> 284 -5.092152e+37 3.686684e+38  3.521482e+36
#> 285 -6.992209e+37 5.062313e+38  4.835468e+36
#> 286 -9.601242e+37 6.951236e+38  6.639747e+36
#> 287 -1.318380e+38 9.544981e+38  9.117265e+36
#> 288 -1.810312e+38 1.310654e+39  1.251923e+37
#> 289 -2.485802e+38 1.799704e+39  1.719058e+37
#> 290 -3.413340e+38 2.471236e+39  2.360498e+37
#> 291 -4.686974e+38 3.393338e+39  3.241281e+37
#> 292 -6.435845e+38 4.659510e+39  4.450715e+37
#> 293 -8.837280e+38 6.398133e+39  6.111429e+37
#> 294 -1.213477e+39 8.785496e+39  8.391813e+37
#> 295 -1.666267e+39 1.206367e+40  1.152309e+38
#> 296 -2.288009e+39 1.656503e+40  1.582275e+38
#> 297 -3.141744e+39 2.274602e+40  2.172676e+38
#> 298 -4.314036e+39 3.123334e+40  2.983376e+38
#> 299 -5.923752e+39 4.288757e+40  4.096576e+38
#> 300 -8.134107e+39 5.889040e+40  5.625149e+38
#> 301 -1.116922e+40 8.086442e+40  7.724085e+38
#> 302 -1.533684e+40 1.110377e+41  1.060621e+39
#> 303 -2.105954e+40 1.524697e+41  1.456375e+39
#> 304 -2.891758e+40 2.093614e+41  1.999798e+39
#> 305 -3.970773e+40 2.874813e+41  2.745991e+39
#> 306 -5.452404e+40 3.947505e+41  3.770615e+39
#> 307 -7.486884e+40 5.420454e+41  5.177561e+39
#> 308 -1.028050e+41 7.443012e+41  7.109487e+39
#> 309 -1.411650e+41 1.022026e+42  9.762281e+39
#> 310 -1.938386e+41 1.403378e+42  1.340492e+40
#> 311 -2.661664e+41 1.927027e+42  1.840676e+40