R6 Class Representing PosteriorBSVAR

The class PosteriorBSVAR contains posterior output and the specification including the last MCMC draw for the homoskedastic bsvar model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BSVAR with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output collected in elements an NxNxS array B, an NxKxS array A, and a 5xS matrix hyper.

Methods

Public methods

specify_posterior_bsvar$new()

specify_posterior_bsvar$get_posterior()

specify_posterior_bsvar$get_last_draw()

specify_posterior_bsvar$is_normalised()

specify_posterior_bsvar$set_normalised()

specify_posterior_bsvar$clone()

Method new()

Create a new posterior output PosteriorBSVAR.

Usage

specify_posterior_bsvar$new(specification_bsvar, posterior_bsvar)

Arguments

specification_bsvar

an object of class BSVAR with the last draw of the current MCMC run as the starting value.

posterior_bsvar

a list containing Bayesian estimation output collected in elements an NxNxS array B, an NxKxS array A, and a 5xS matrix hyper.

Returns

A posterior output PosteriorBSVAR.

Method get_posterior()

Returns a list containing Bayesian estimation output collected in elements an NxNxS array B, an NxKxS array A, and a 5xS matrix hyper.

Usage

specify_posterior_bsvar$get_posterior()

Examples

data(us_fiscal_lsuw)
specification  = specify_bsvar$new(us_fiscal_lsuw)
set.seed(123)
estimate       = estimate(specification, 50)
estimate$get_posterior()

Method get_last_draw()

Returns an object of class BSVAR with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bsvar$get_last_draw()

Examples

data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# run the burn-in
burn_in        = estimate(specification, 10)

# estimate the model
posterior      = estimate(burn_in, 10)

Method is_normalised()

Returns TRUE if the posterior has been normalised using normalise_posterior() and FALSE otherwise.

Usage

specify_posterior_bsvar$is_normalised()

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10, thin = 1)

# check normalisation status beforehand
posterior$is_normalised()

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) %*% BB         # set negative diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()

Method set_normalised()

Sets the private indicator normalised to TRUE.

Usage

specify_posterior_bsvar$set_normalised(value)

Arguments

value

(optional) a logical value to be passed to indicator normalised.

Examples

# This is an internal function that is run while executing normalise_posterior()
# Observe its working by analysing the workflow:

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10, thin = 1)

# check normalisation status beforehand
posterior$is_normalised()

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) %*% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bsvar$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# This is a function that is used within estimate()
data(us_fiscal_lsuw)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)
estimate       = estimate(specification, 50)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 50 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(estimate)
#> [1] "PosteriorBSVAR" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bsvar$get_posterior`
## ------------------------------------------------

data(us_fiscal_lsuw)
specification  = specify_bsvar$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)
estimate       = estimate(specification, 50)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 50 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
estimate$get_posterior()
#> $B
#> , , 1
#> 
#>            [,1]       [,2]     [,3]
#> [1,]  8.6261202  0.0000000 0.000000
#> [2,]  0.8378946 12.0267273 0.000000
#> [3,] -8.5095386  0.4928987 1.285509
#> 
#> , , 2
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  33.359735  0.000000  0.00000
#> [2,]   1.923635 37.547790  0.00000
#> [3,] -13.540951 -3.727666 93.37954
#> 
#> , , 3
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  32.852230  0.000000  0.00000
#> [2,]   3.195413 37.190639  0.00000
#> [3,] -12.607721  1.211786 96.77188
#> 
#> , , 4
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  34.994010  0.000000  0.00000
#> [2,]   5.753199 37.856870  0.00000
#> [3,] -14.526017  1.629372 99.94835
#> 
#> , , 5
#> 
#>             [,1]       [,2]     [,3]
#> [1,]  32.2710050  0.0000000   0.0000
#> [2,]   0.3780487 39.1320006   0.0000
#> [3,] -16.8021397 -0.2978158 103.2576
#> 
#> 
#> $A
#> , , 1
#> 
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 1.2375822  0.1792126 -0.6236299 -0.6915682
#> [2,] 0.8617537  0.7761186 -1.0160909 -1.7170776
#> [3,] 0.3298017 -0.1537224  0.7477115 -0.2736510
#> 
#> , , 2
#> 
#>              [,1]         [,2]       [,3]        [,4]
#> [1,]  0.900789869  0.002042344 0.11966744 -0.01093535
#> [2,] -0.006672419  0.959253393 0.01248304 -0.37258433
#> [3,] -0.009642604 -0.006964845 1.00891841 -0.08986496
#> 
#> , , 3
#> 
#>             [,1]         [,2]       [,3]         [,4]
#> [1,]  0.90976752  0.006308909 0.10555534  0.008862357
#> [2,] -0.03623198  0.936741976 0.05275208 -0.574674488
#> [3,] -0.01538934 -0.008387779 1.01746430 -0.092980614
#> 
#> , , 4
#> 
#>             [,1]         [,2]       [,3]       [,4]
#> [1,]  0.92168543 -0.015910427 0.09907609 -0.1485495
#> [2,] -0.05439032  0.941661621 0.07953510 -0.4949069
#> [3,] -0.01286130 -0.006110601 1.01800134 -0.0431329
#> 
#> , , 5
#> 
#>              [,1]         [,2]       [,3]        [,4]
#> [1,]  0.940547546 -0.004144984 0.07781093 -0.01087784
#> [2,] -0.052144064  0.950114898 0.07457366 -0.42588207
#> [3,] -0.009912825 -0.005117454 1.01039847 -0.06258951
#> 
#> 
#> $hyper
#> , , 1
#> 
#>           [,1]       [,2]
#> [1,]  2.234132  1.1461110
#> [2,]  6.153157  1.0956512
#> [3,]  3.382887  0.5103117
#> [4,] 22.651755 17.4748071
#> [5,] 26.013064 15.6841633
#> [6,] 41.449674 10.0039639
#> [7,]  2.962011  0.8364195
#> 
#> , , 2
#> 
#>           [,1]      [,2]
#> [1,] 274.55911 0.0930118
#> [2,] 291.47071 0.1616454
#> [3,] 554.04374 0.2168844
#> [4,] 673.58983 2.2452704
#> [5,] 358.36251 2.9235495
#> [6,] 701.33690 4.5903124
#> [7,]  54.69388 0.3843997
#> 
#> , , 3
#> 
#>           [,1]      [,2]
#> [1,]  509.0258 0.5561562
#> [2,]  451.7650 0.4256890
#> [3,]  834.0089 0.4519654
#> [4,] 3082.5552 5.4816032
#> [5,] 3511.8930 3.7464007
#> [6,] 6992.9929 6.8380845
#> [7,]  324.3035 0.5994299
#> 
#> , , 4
#> 
#>           [,1]      [,2]
#> [1,]  2076.602 0.4971159
#> [2,]  4905.775 0.2388233
#> [3,]  1850.122 0.6692453
#> [4,] 22288.747 5.2163116
#> [5,] 17366.085 3.4855224
#> [6,] 16605.217 4.2634562
#> [7,]  2233.914 0.4785924
#> 
#> , , 5
#> 
#>           [,1]      [,2]
#> [1,]  6711.711 0.3123227
#> [2,]  4654.280 0.3959182
#> [3,]  2623.426 0.6935315
#> [4,] 31510.247 2.5153773
#> [5,] 29683.746 4.7403916
#> [6,] 33080.734 6.6575112
#> [7,]  2365.514 0.7492982
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bsvar$get_last_draw`
## ------------------------------------------------

data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

## ------------------------------------------------
## Method `specify_posterior_bsvar$is_normalised`
## ------------------------------------------------

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10, thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# check normalisation status beforehand
posterior$is_normalised()
#> [1] TRUE

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) %*% BB         # set negative diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()
#> [1] TRUE

## ------------------------------------------------
## Method `specify_posterior_bsvar$set_normalised`
## ------------------------------------------------

# This is an internal function that is run while executing normalise_posterior()
# Observe its working by analysing the workflow:

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10, thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# check normalisation status beforehand
posterior$is_normalised()
#> [1] TRUE

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) %*% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()
#> [1] TRUE