R6 Class Representing PosteriorBSVART

The class PosteriorBSVART contains posterior output and the specification including the last MCMC draw for the bsvar model with t-distributed structural shocks. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BSVART with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bsvar_t$new()

specify_posterior_bsvar_t$get_posterior()

specify_posterior_bsvar_t$get_last_draw()

specify_posterior_bsvar_t$is_normalised()

specify_posterior_bsvar_t$set_normalised()

specify_posterior_bsvar_t$clone()

Method new()

Create a new posterior output PosteriorBSVART.

Usage

specify_posterior_bsvar_t$new(specification_bsvar, posterior_bsvar)

Arguments

specification_bsvar

an object of class BSVART with the last draw of the current MCMC run as the starting value.

posterior_bsvar

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBSVART.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bsvar_t$get_posterior()

Examples

data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
set.seed(123)
estimate       = estimate(specification, 10)
estimate$get_posterior()

Method get_last_draw()

Returns an object of class BSVART with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bsvar_t$get_last_draw()

Examples

data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)

# run the burn-in
set.seed(123)
burn_in        = estimate(specification, 10)

# estimate the model
posterior      = estimate(burn_in, 10)

Method is_normalised()

Returns TRUE if the posterior has been normalised using normalise_posterior() and FALSE otherwise.

Usage

specify_posterior_bsvar_t$is_normalised()

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)

# estimate the model
set.seed(123)
posterior      = estimate(specification, 10)

# check normalisation status beforehand
posterior$is_normalised()

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) %*% BB         # set negative diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()

Method set_normalised()

Sets the private indicator normalised to TRUE.

Usage

specify_posterior_bsvar_t$set_normalised(value)

Arguments

value

(optional) a logical value to be passed to indicator normalised.

Examples

# This is an internal function that is run while executing normalise_posterior()
# Observe its working by analysing the workflow:

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10)

# check normalisation status beforehand
posterior$is_normalised()

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) %*% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bsvar_t$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# This is a function that is used within estimate()
data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)
estimate       = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(estimate)
#> [1] "PosteriorBSVART" "R6"             

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$get_posterior`
## ------------------------------------------------

data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)
estimate       = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
estimate$get_posterior()
#> $B
#> , , 1
#> 
#>            [,1]      [,2]     [,3]
#> [1,] 39.3278839  0.000000 0.000000
#> [2,] 16.2877941 16.592296 0.000000
#> [3,] -0.6580358  5.169431 9.618209
#> 
#> , , 2
#> 
#>           [,1]     [,2]     [,3]
#> [1,] 45.746900  0.00000  0.00000
#> [2,] 33.497415 31.86115  0.00000
#> [3,] -3.432359 16.75727 49.71248
#> 
#> , , 3
#> 
#>           [,1]     [,2]     [,3]
#> [1,]  42.42143  0.00000   0.0000
#> [2,]  27.67203 42.93766   0.0000
#> [3,] -29.12918 15.50224 107.3847
#> 
#> , , 4
#> 
#>           [,1]     [,2]     [,3]
#> [1,]  38.38618  0.00000   0.0000
#> [2,]  23.90500 55.12055   0.0000
#> [3,] -31.38716 15.12797 121.3633
#> 
#> , , 5
#> 
#>           [,1]       [,2]     [,3]
#> [1,]  41.85448  0.0000000   0.0000
#> [2,]  13.75762 61.5526984   0.0000
#> [3,] -29.87929 -0.7783589 142.2431
#> 
#> , , 6
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  51.599344  0.000000   0.0000
#> [2,]  -3.277018 63.530042   0.0000
#> [3,] -33.120346 -1.577987 150.7041
#> 
#> , , 7
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  58.668958  0.000000   0.0000
#> [2,]   4.532166 65.821311   0.0000
#> [3,] -22.505677 -9.139299 149.1685
#> 
#> , , 8
#> 
#>            [,1]        [,2]    [,3]
#> [1,]  64.525503  0.00000000   0.000
#> [2,]   2.670176 67.59607379   0.000
#> [3,] -22.589864 -0.06630228 154.524
#> 
#> , , 9
#> 
#>           [,1]      [,2]     [,3]
#> [1,]  69.25239   0.00000   0.0000
#> [2,]  -5.53338  63.01956   0.0000
#> [3,] -18.08221 -11.65500 170.5143
#> 
#> , , 10
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  67.851891   0.00000   0.0000
#> [2,]  -3.820607  68.54862   0.0000
#> [3,] -30.615519 -10.33229 178.0307
#> 
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.1533909 -0.1773105 -0.0604695 -0.7881146
#> [2,] -0.3516446  1.5552609 -0.1526818  1.1840918
#> [3,] -1.0020651  0.2755269  2.0807280  1.0451583
#> 
#> , , 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  1.03246476 -0.06537434  0.03308503 -0.10851591
#> [2,]  0.04533813  1.04606391 -0.11367738  0.09961653
#> [3,] -0.10048112  0.03888222  1.10135449  0.17437260
#> 
#> , , 3
#> 
#>             [,1]         [,2]       [,3]       [,4]
#> [1,] 0.913978848 -0.036570957  0.1308530 -0.2061772
#> [2,] 0.074877012  0.978263732 -0.0967264 -0.2019719
#> [3,] 0.009244557 -0.006606585  0.9910046 -0.0558321
#> 
#> , , 4
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.93612947 -0.02416502 0.09016149 -0.17996405
#> [2,] -0.01372968  0.99782914 0.01435346 -0.02149363
#> [3,] -0.01147374 -0.01424598 1.01381458 -0.15140317
#> 
#> , , 5
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.961678467 -0.01622346  0.049792956 -0.16056080
#> [2,]  0.003446681  0.98980388 -0.002634331 -0.07867621
#> [3,] -0.018376681 -0.01167032  1.023020675 -0.11740366
#> 
#> , , 6
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.914709917 -0.026862156 0.106426128 -0.27535085
#> [2,] -0.008532071  1.003191968 0.009528175  0.02645740
#> [3,] -0.025572271 -0.006122004 1.028755478 -0.08379931
#> 
#> , , 7
#> 
#>             [,1]         [,2]       [,3]        [,4]
#> [1,]  0.94568859  0.009018638 0.05922836  0.02400686
#> [2,] -0.01727631  0.982361333 0.02846490 -0.12323897
#> [3,] -0.01058652 -0.007622513 1.01098989 -0.09119439
#> 
#> , , 8
#> 
#>              [,1]         [,2]       [,3]        [,4]
#> [1,]  0.956077391 -0.014723097 0.05569930 -0.13904482
#> [2,] -0.024342439  0.974402168 0.02931844 -0.26299419
#> [3,] -0.005639281 -0.004945347 1.00868831 -0.03332264
#> 
#> , , 9
#> 
#>               [,1]         [,2]       [,3]        [,4]
#> [1,]  0.9685901479 -0.006196329 0.03334303 -0.10261842
#> [2,] -0.0358614429  0.974665010 0.04249011 -0.26927060
#> [3,]  0.0004113735 -0.010643545 0.99985415 -0.09950745
#> 
#> , , 10
#> 
#>              [,1]         [,2]       [,3]        [,4]
#> [1,]  0.976041609 -0.019099484 0.03418676 -0.15344144
#> [2,] -0.017616487  0.955358857 0.02641442 -0.41179225
#> [3,]  0.001539218 -0.007229232 0.99906529 -0.06065964
#> 
#> 
#> $hyper
#> , , 1
#> 
#>           [,1]       [,2]
#> [1,]  7.413741  1.4936374
#> [2,]  5.498240  0.4553124
#> [3,]  2.175053  1.1248598
#> [4,] 37.230551 11.5263567
#> [5,] 58.054742  8.4424441
#> [6,] 27.578742 15.5970321
#> [7,]  2.969482  0.9962926
#> 
#> , , 2
#> 
#>           [,1]       [,2]
#> [1,] 240.84769  0.9724895
#> [2,]  96.90712  0.9926595
#> [3,]  11.63167  0.7344653
#> [4,]  63.69760 13.9895262
#> [5,]  20.70084  8.8519567
#> [6,]  50.39444 12.1605699
#> [7,]   4.96400  1.1961222
#> 
#> , , 3
#> 
#>            [,1]       [,2]
#> [1,] 228.333978  0.4981657
#> [2,] 191.184834  0.7110807
#> [3,] 271.751714  2.5650867
#> [4,]  89.661153  9.3755023
#> [5,]  49.445931 12.9433084
#> [6,]  79.415008 15.3171479
#> [7,]   5.510131  1.2049897
#> 
#> , , 4
#> 
#>            [,1]      [,2]
#> [1,] 205.661972 1.0467303
#> [2,] 286.599596 0.6661651
#> [3,] 744.314686 0.5646045
#> [4,] 145.153593 7.6840031
#> [5,] 102.765696 8.7234146
#> [6,] 117.196272 6.8343969
#> [7,]   8.148901 0.9863598
#> 
#> , , 5
#> 
#>           [,1]       [,2]
#> [1,] 169.24382  0.4339616
#> [2,] 456.24460  0.7668927
#> [3,] 859.27454  0.8022372
#> [4,] 158.67059  5.4393917
#> [5,] 156.18460 10.5630243
#> [6,] 342.71710  9.9518733
#> [7,]  17.20923  0.9288462
#> 
#> , , 6
#> 
#>            [,1]      [,2]
#> [1,]  223.22300 0.4967632
#> [2,]  430.07275 0.5313865
#> [3,] 1009.68287 0.6846026
#> [4,]  352.12924 6.0771745
#> [5,]  205.87501 5.3903189
#> [6,]  286.36060 6.4183688
#> [7,]   21.07281 0.8010212
#> 
#> , , 7
#> 
#>            [,1]      [,2]
#> [1,]  686.25656 0.6478217
#> [2,]  390.62435 0.5477689
#> [3,] 1362.00526 0.5225953
#> [4,]  790.15960 5.9569710
#> [5,]  295.29466 4.8230982
#> [6,]  542.10719 8.4166793
#> [7,]   30.35847 0.6040122
#> 
#> , , 8
#> 
#>           [,1]      [,2]
#> [1,]  452.0606 0.2493868
#> [2,]  638.7109 0.4823189
#> [3,] 2644.0417 0.2729260
#> [4,]  945.3882 5.1879800
#> [5,]  800.4810 3.9736545
#> [6,]  249.6736 4.6946398
#> [7,]   48.9423 0.6068556
#> 
#> , , 9
#> 
#>            [,1]      [,2]
#> [1,]  686.91753 0.6611756
#> [2,]  362.66266 0.3753999
#> [3,] 1961.63493 0.3812172
#> [4,]  520.58058 5.3973117
#> [5,]  927.82798 5.3901962
#> [6,]  634.75081 4.7006811
#> [7,]   47.44313 0.5832104
#> 
#> , , 10
#> 
#>            [,1]      [,2]
#> [1,]  566.08099 0.3390585
#> [2,]  840.78669 0.1574949
#> [3,] 1593.21099 0.2071879
#> [4,] 1029.91223 5.8641571
#> [5,]  678.87963 4.0736058
#> [6,]  902.16440 5.2235823
#> [7,]   58.66812 0.6125942
#> 
#> 
#> $lambda
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>   [1,]  8.807952  3.9144184  2.6450033  6.0850890  1.9633977  3.2585344
#>   [2,] 44.404426  2.9868856  2.3240254  2.0169544  1.1081477  2.1530347
#>   [3,] 13.826366  9.0545711  0.9610479  1.0392420  3.8699154  1.1327414
#>   [4,]  7.186022  6.6751310  1.0870514  1.5869042  2.1487399  2.8940727
#>   [5,] 11.755645  2.0375232  1.0541595  0.8232262  0.3581479  1.4380136
#>   [6,] 31.386016  4.6277091  1.3620688  3.8950031  3.6338701  1.6060049
#>   [7,] 45.666855  4.3395245  5.4666337  3.3657983  2.2695865  2.2944776
#>   [8,]  8.637378 12.6777214  2.0756659  1.7523950  2.4270277  3.5243112
#>   [9,] 12.343676  2.5185845  2.0889275  2.9312245  1.2737639  0.7744255
#>  [10,] 12.142848  1.7942851 12.0921537  1.7503624  1.7469865  4.1147654
#>  [11,] 13.689296  6.0173156  6.7835884  7.1730521  6.0698800  9.7523321
#>  [12,] 19.034471  3.8307369  3.8993419  6.3640755 13.1868284 16.4916125
#>  [13,]  7.918713  2.8212537  4.1489303  6.1507298 37.2496300 26.3904651
#>  [14,]  9.362781  1.8696993  6.9685689  7.7160406 14.4276070 11.2191810
#>  [15,] 13.940470 13.0344579  3.5506277  2.4368583  5.6505977  4.3112358
#>  [16,] 17.392385 19.0174632  6.1047984  3.3873268  2.0975036  2.1901003
#>  [17,] 15.027218  9.0209162  8.9374977  1.6661793  1.4515843  0.9336725
#>  [18,] 24.308314 11.4352335  1.4623780  0.7765638  0.7362741  1.5842425
#>  [19,] 13.941594 21.2508602  2.6521978  2.4550565  3.2663478  3.2131318
#>  [20,] 39.872408  6.9265288  2.0832590  1.9814239  2.7929825  0.7541353
#>  [21,] 26.454872 10.2348545  1.1501296  4.0814750  2.2496362  2.6744182
#>  [22,] 26.139590  9.8243316  7.3035618  2.1955383  1.9143682  3.3642834
#>  [23,] 15.852703 11.4034756  2.7680825  1.1965593  3.7520854  2.8964751
#>  [24,] 15.759582  9.1676769  5.1632263  4.1745161  5.0515159  3.1558456
#>  [25,] 14.504847 18.7504447  1.5691065  1.9764111  4.5442746  4.7435127
#>  [26,]  9.374929  7.9747666  1.9226573  3.1696387  4.6908225  1.9847168
#>  [27,]  9.893213 14.2570920  4.2979491  2.2480008  2.3564434  2.8526110
#>  [28,] 10.476098 14.5019029  5.6319371  1.0249355  1.2883438  2.0037734
#>  [29,] 14.003353  6.2279017  5.1640416  0.9586257  1.8535653  1.0198329
#>  [30,] 11.795728  3.7342566  2.1091167  3.7647504  1.3442992  0.5323031
#>  [31,] 28.199166 11.1935410  5.2544950  7.4670871  1.4979730  3.3075619
#>  [32,] 15.154949 27.0755117  9.6956205  1.6559599  1.5243525  1.1287039
#>  [33,] 14.729227  6.1795118  2.8783126  0.9271649  2.4925994  0.6314967
#>  [34,]  8.035368  8.6316114  1.9559371  0.6360089  1.8685192  0.7158891
#>  [35,] 26.141419  4.7045989  6.6757805  0.7993058  0.5844359  3.9569336
#>  [36,] 11.131273  6.1851444  1.8456211  0.9376907  0.8370993  1.6050394
#>  [37,] 22.355668  6.8816135  2.7428115  1.0607720  1.9370308  1.1611440
#>  [38,] 14.100575 10.7485582  1.9023297 11.5760897  1.4453461  1.0777802
#>  [39,] 16.986257 10.9380245  1.6657369  1.2941655  1.5321579  1.2847052
#>  [40,] 13.928862  9.2461176  2.4735630  1.7809431  4.2472841  5.5115401
#>  [41,] 11.839394  7.6918434  4.2996287  1.9764327  1.5165487  0.9705819
#>  [42,] 24.830192 13.9618845  2.7590207  2.6705092  1.4993383  1.1250486
#>  [43,] 35.828069  7.3016482  1.7354153  0.9340933  1.1661841  0.8050931
#>  [44,] 25.135858 10.4566893  6.3808206  3.1688711  2.1801632  2.2075477
#>  [45,] 12.093251  5.5621096  2.5716136  2.5536409  1.9080829  1.6875391
#>  [46,] 11.215450  5.6552692  3.1091883  1.3994415  1.0952236  1.9063312
#>  [47,] 17.626811 21.2086252  2.1842479  4.8019662  1.7143906  0.6240573
#>  [48,]  5.722921  9.0357919  2.4484687  3.9523917  3.4035795  2.4717118
#>  [49,] 30.496757  8.6991761  3.1794121  1.8731441  2.5902148  1.4312407
#>  [50,] 11.634614 13.2661515  0.7940888  1.0459394  0.8937926  0.7768456
#>  [51,] 10.914996 17.1602381  2.0513472  2.9169322  2.2759868  2.2554224
#>  [52,] 13.148691  4.8531967  1.9171808  1.6114530  1.0830249  0.7416038
#>  [53,]  8.890706  5.8892873  1.9244173  2.0530533  2.3172134  0.7341282
#>  [54,]  5.861604  3.3755828  3.9670368  1.5266706  0.8571892  9.2708453
#>  [55,] 17.358817  7.9064575  1.3604762  1.6593899  1.3690698  1.4953312
#>  [56,] 22.113932  9.8769457  1.8241018  1.4627496  2.5692907  2.8618254
#>  [57,] 19.627241  3.7872544  1.2636489  1.3901983  2.8022023  1.1212186
#>  [58,]  9.915892  4.3936162  2.7443708  1.9759332  2.3885378  1.8587006
#>  [59,] 13.287098  7.0276098  1.1712882  4.3108225  1.6240350  1.2033584
#>  [60,] 27.506940  5.2385115  2.5298825  2.5102513  3.4235345  0.9289478
#>  [61,] 20.534651 26.5413694  2.3888323  0.4229365  0.5896960  0.8257187
#>  [62,] 13.138929 11.3929214  2.0402293  1.4881134  1.3096456  3.3479337
#>  [63,] 11.004311  8.4012003  1.3616156  1.7445357  1.3322529  1.3120052
#>  [64,] 16.041010 10.2333232  1.9536724  1.6828452  3.8806313  1.7723080
#>  [65,] 20.253823 13.1872591  2.6652219  2.7646207  2.6872783  2.4202884
#>  [66,] 11.243498  5.7971102  2.3004118  1.9462235  0.9800183  0.5827608
#>  [67,]  8.145870  9.0511870  2.1168483  2.0474067  1.2153748  2.3204260
#>  [68,] 10.742978  7.9006362 12.0391773  2.3472410  1.0903561  1.6070375
#>  [69,] 15.655037  3.0801260  1.3142509  4.0972180  1.0361150  2.6927947
#>  [70,] 11.856562  2.7949846  0.9492038  2.7851801  1.6130240  4.6915407
#>  [71,] 16.916310  5.1898887  1.5676092  3.5737911  3.4419383  1.2005425
#>  [72,] 11.555175  8.3228124  4.8224671 14.7967072  4.6536759  0.9119039
#>  [73,] 18.245064  3.1604959  2.7777090  1.9321948  1.5717790  1.7704358
#>  [74,]  7.363116  2.1531151  2.3578483  1.4725553  0.9321876  1.3734263
#>  [75,] 16.504562  3.5669474  1.8700260  1.0101506  0.9879776  0.5208163
#>  [76,]  6.690380  3.3058725  3.5396767  4.1130369  2.8615201  4.2044784
#>  [77,] 14.507387  8.4966997  2.1574279  2.8484893  1.6863444  2.0924845
#>  [78,] 11.711077  6.0804416  0.5348212  1.3954121  1.2948687  0.4834504
#>  [79,]  6.137846  9.0100799  1.0047272  2.2411839  1.5679859  1.1339288
#>  [80,] 12.526559 16.8144445  1.6715124  1.1042395  2.1225928  1.8910227
#>  [81,]  7.624849  3.8808501  3.6623359  0.9212536  1.4358599  0.9366082
#>  [82,] 17.235068  8.0356223  4.7755989  2.2015344  2.4133579  1.4462236
#>  [83,] 13.081892  7.9286466  3.8731116  0.5762740  1.3075293  0.9114784
#>  [84,] 10.322282 22.2560390  3.8792850  1.8978626  1.7949750  3.6803772
#>  [85,] 10.011413  2.5272028  3.1310797  1.3709357  2.7883504  2.2532622
#>  [86,] 14.415025  4.9144999  1.3526219  0.5403760  1.1392595  1.0788341
#>  [87,] 17.560765  2.9075363  4.7906986  2.7255520  3.9881931  1.7815962
#>  [88,] 10.245607  4.6382309  2.0543948  1.7570389  2.1322041  0.9520845
#>  [89,] 14.850835  8.7991277  3.3732894  1.0670980  1.1928575  1.4876556
#>  [90,] 10.142986  4.1858615  2.4052750  2.4640494  1.0722600  1.2846750
#>  [91,] 11.230225  6.6538159  3.1624725  1.7603326  1.4745890  1.2417624
#>  [92,] 12.323248  5.2852093  2.5111183  1.7191242  1.8946896  1.4451952
#>  [93,] 18.434690  5.6371806  1.9175468  1.6437281  1.2736640  0.8601045
#>  [94,]  8.450835  2.4954929  2.0166577  1.8083957  0.6444040  1.1437989
#>  [95,]  8.858530  2.4321049  1.4858737  1.6563964  1.1406823  1.4396641
#>  [96,] 10.621938  4.5185610  5.1666220  2.5770181  3.1251089  3.1574056
#>  [97,] 10.999944  6.2981904  1.4199975  1.6907979  1.0934854  0.9124677
#>  [98,]  9.887718  2.4903183  1.7932445  2.6719681  1.7369253  2.9799181
#>  [99,]  6.033867  3.1961160  1.2295142  2.3925999  1.5087378  1.1168586
#> [100,]  9.774233  4.7276091  1.7771109  4.7922859  1.7266619  5.6209277
#> [101,] 19.882884  2.3814328  2.5406486  1.3733295  1.3870329  2.6352098
#> [102,] 12.585239  2.9511975  1.5892442  3.6959496  3.5957598  1.1558254
#> [103,] 22.626481  2.9032063  1.2531432  0.9933034  0.9448969  1.3786668
#> [104,]  7.974877  2.9207318  1.0354566  4.2967661  1.6277765  1.6048518
#> [105,]  7.788878  2.7519172  2.5671306  1.4633768  0.9022139  1.0737824
#> [106,]  6.575061  5.1321498  1.0313435  1.6200432  3.4066784  1.6273194
#> [107,] 11.920716  5.3441332  2.4417865  1.0933428  1.2256981  1.8997130
#> [108,] 16.186311  6.1627235  1.8303899  1.1217871  3.9602025  1.7083350
#> [109,]  8.032336  8.8745267 20.1481183 11.1108147 17.1461890 20.0408635
#> [110,] 32.477775 16.9601611 11.8937610  7.2937895  7.8545828  5.2872812
#> [111,]  9.786715  1.7497123  1.7462622  0.6075215  0.8235597  0.7632888
#> [112,] 13.671930  6.1703016  1.0183918  1.2744247  1.4629213  1.7308524
#> [113,]  9.493305  5.5227993  1.8749345  1.0904719  0.8818108  2.0083148
#> [114,] 15.026983  3.7190098  3.8390446  0.8133955  1.2339763  1.8272193
#> [115,] 15.022976  2.1553959  1.0454184  1.1456088  1.5718645  0.6751563
#> [116,] 36.417357  4.3347092  0.6478305  0.9961024  1.5473510  1.7522247
#> [117,] 28.202332  8.3600044  2.0425508  1.7063442  1.6256004  0.7970628
#> [118,]  9.508291  4.9861981  1.5563486  0.9597893  2.1437202  1.0969783
#> [119,]  5.622124  2.2418401  1.2106395  1.2386989  2.3106684  0.7325499
#> [120,] 28.445411  1.7524962  1.2067341  2.1215517  2.2330902  1.3558330
#> [121,] 76.235396  4.7652508  2.7703236  3.5663937  2.7158915  4.1491708
#> [122,] 11.086405  2.2822269  2.2522947  0.6005618  1.0988448  1.9052053
#> [123,] 23.230972  1.8757346  2.9740538  1.4040557  0.9038868  3.1745990
#> [124,] 13.140065  4.9946024  1.4771730  1.8859311  1.2070952  0.6610680
#> [125,] 10.007383  4.7990555  1.2901616  2.1770860  1.1542834  1.4562019
#> [126,] 15.826348  0.9829187  2.1353422  1.2152826  1.2812254  0.5792636
#> [127,] 19.290918  2.3149341  0.9313527  0.4733514  1.6480178  1.7856336
#> [128,]  8.139267  2.8300985  3.1422192  1.6463038  0.7362928  0.8533688
#> [129,] 23.837072  1.4280134  4.6968093  1.5021615  3.5438559  5.6523581
#> [130,]  7.093648  3.1580950  2.3069799  1.7907046  1.8766977  1.6918975
#> [131,]  9.666534  2.9229901  1.3419597  2.4680108  3.6401191  1.0614354
#> [132,] 16.750938  6.4002157  4.4158037  2.4110433  1.3055592  2.3023764
#> [133,] 10.872255  5.2405240  2.1924710  1.0465433  2.3856651  1.2326996
#> [134,]  5.727551  3.2972648  2.5464475  1.1934387  0.7302910  0.8347639
#> [135,] 25.986527  2.8024497  2.6369499  9.1914025  1.1528414  0.4831819
#> [136,] 13.186502  7.4514048  2.1740117 10.6215220  1.1053584  1.7550351
#> [137,] 14.988340  9.3062245  3.0627375  1.4350086  0.9248532  1.6834542
#> [138,] 24.558226  3.9433670  6.9134914  1.7037815  1.5540776  0.5407219
#> [139,] 22.659187  3.3585342  0.9091036  1.3909454  2.1756213  1.0771404
#> [140,] 10.471032  5.0226706  1.2888189  1.0884435  2.3390165  1.4346138
#> [141,] 16.355944  3.8221123  1.0867425  2.0152348  1.9240235  1.7726540
#> [142,] 20.839011  4.3509496  2.6640897  2.6236327  2.3021812  3.6395115
#> [143,] 15.401640  4.5036979  2.4520215  3.4003047  1.5305130  3.8882402
#> [144,] 56.771601  1.8635534  3.0978163  1.2709903  0.7921598  0.5288758
#> [145,] 14.064973  2.1027624  2.3148683  1.3776720  1.5404819  1.8781123
#> [146,] 13.462043  4.7151244  1.4446693  1.3609218  4.2729847  1.6695827
#> [147,]  8.953501  1.2006066  1.6789147  2.0887363  1.1364732  3.3781968
#> [148,] 16.522650  1.6279909  2.1871989  2.6953571  1.8464808  4.0610433
#> [149,]  9.119398  6.8695845  3.1787632  2.7977372  4.2689135  3.6736149
#> [150,]  8.954013  6.0773323  2.0603570  1.5288509  3.2389940  4.9091912
#> [151,]  6.987858  5.3445363  1.4220410  0.8400251  1.2242363  0.8305563
#> [152,] 17.198966 11.8093209  1.6095744  4.3674715  2.6416894  4.3031954
#> [153,]  5.582069  0.9436890  1.6150432  1.3476131  0.8534351  0.5486392
#> [154,]  9.443051  3.8507648  0.9968224  1.7105939  0.7055298  3.1143219
#> [155,] 26.551820  6.2091471  1.2638230  2.7487176  2.4508966  0.8789366
#> [156,] 27.032377  4.9893773  3.1808246  1.0090567  0.2574732  0.9413294
#> [157,]  5.202395  6.2107863  5.5421570  2.0376300  2.2877844  3.2363609
#> [158,]  7.097950  5.6740604  1.7638764  1.4169473  1.5123493  2.0549273
#> [159,] 14.306892  2.5742140  1.6538470  3.1481238  1.3208104  1.2136557
#> [160,] 29.391470  4.8434349  1.4177147  0.9532986  1.0075387  1.6356612
#> [161,] 16.137477  2.5644810  3.2522229  1.1821750  0.5965215  1.0969988
#> [162,] 12.574895  1.7085654  1.4494855  1.1623380  1.2720062  1.6507641
#> [163,]  6.249722  2.4416342  1.0615503  0.7798103  0.7066689  1.1475260
#> [164,] 12.861594  2.0560070  1.6278141  2.9626104  1.7546063  1.1685594
#> [165,] 10.456326  2.2499427  2.2530567  1.3079804  0.7908095  2.1246006
#> [166,] 15.944113  2.4346348  2.6384364  5.8283648  2.0867335  3.2746124
#> [167,] 16.470353  2.5021804  1.7859339  2.0035211  1.6578537  1.2096830
#> [168,]  7.702334  1.7559584  6.5917695  1.2994283  1.0623732  1.4794360
#> [169,] 11.527359  5.2090510  2.3886834  0.8050111  3.0482676  0.7405106
#> [170,]  4.444610  4.4291575  2.5864574  1.7171105  0.8083349  1.9008398
#> [171,] 10.892454  3.0511497  2.2939983  1.2513296  1.0521153  0.9617159
#> [172,] 12.707645  3.8370117  2.6373906  2.2706948  1.3158816  0.4048326
#> [173,] 17.027693  2.1181921  1.7054202  1.5380417  1.7570857  2.7679901
#> [174,] 21.831327  3.4403110  1.8626637  1.4141319  1.8717173  2.6574190
#> [175,]  9.060255  7.3754063  1.1273146  0.8039447  1.0077347  0.7471427
#> [176,] 15.158573  1.8518564  0.8532447  0.8062379  1.4829720  2.4515438
#> [177,] 15.347873  2.3439551  2.1822760  1.1042676  0.9050777  0.4984136
#> [178,] 18.544808  1.4838995  0.9960646  1.7023373  1.7301413  0.9107521
#> [179,]  9.178773  2.7812100  1.0959422  1.1010320  1.0969158  1.3043261
#> [180,] 24.795238  1.6756590  7.9213368  2.0140323  1.7440032  0.7875262
#> [181,] 13.158749  4.9798941  0.7873626  0.9407279  0.5321187  2.7676150
#> [182,] 17.512202  2.2967096  1.2518126  1.4652257  0.5120158  0.8126772
#> [183,] 17.112425  3.3821185  1.4161407  3.7482707  0.7808456  0.7331932
#> [184,]  8.330510  4.8882059  1.3116748  1.4889310  1.8236458  1.1066151
#> [185,]  9.144208  3.0205303  3.5222076  1.7659612  0.8768058  1.1032191
#> [186,] 13.477887  2.2234938  1.7953563  1.1391931  1.6506379  1.5377110
#> [187,] 19.460303  3.1739394  6.2229170  1.8369740  0.9373122  1.4476148
#> [188,]  8.414428  4.5858063  1.3089594  1.2087411  1.2326317  1.1805998
#> [189,]  9.995293  3.3065724  1.5629218  0.9980623  0.5990024  1.3222038
#> [190,] 25.660394  2.2484296  4.9102702  1.9711600  1.2064946  1.0274973
#> [191,] 13.862125  4.2376665  2.9174791  2.8618048  1.2913977  2.2345810
#> [192,] 31.546333  3.7670199  1.8848086  1.0303477  1.0583446  1.0329793
#> [193,] 28.897510  3.9274410  0.8615036  2.6808628  0.6756006  0.9129020
#> [194,] 36.529548  1.5853534  2.4218136  1.0341905  1.8290644  1.3138948
#> [195,] 15.367359  8.6654800  0.9579801  1.3114187  1.2025244  0.6808994
#> [196,] 30.066301  3.9973458  2.1032073  1.2702103  1.0220535  2.9849618
#> [197,] 12.024400  3.2496022  1.2161724  0.8495460  1.5520886  1.1363570
#> [198,]  6.429244  5.2369286  3.7594819  0.9465781  1.4061621  1.2339169
#> [199,] 13.161580  3.9536459  1.0316345  1.6487418  0.8190051  0.9066500
#> [200,] 14.492872  7.6215477  1.9210625  1.7196062  2.0724272  1.4179421
#> [201,] 13.322416  4.0154210  1.9262848  0.6813771  0.7538060  2.0582206
#> [202,]  8.395974  7.2670483  1.2633319  0.6776626  0.9951132  1.0236488
#> [203,] 15.048447  6.8356798  3.3194159  3.0044290  1.3641506  1.9378415
#> [204,] 12.277104 10.3533627  5.6128024  1.7621804  1.4942487  0.7812549
#> [205,]  7.152306  4.7575200  1.4728113  1.0573106  1.5901553  3.3740866
#> [206,]  5.053864  4.5498362  1.2707282  1.3870983  1.0776398  1.0339509
#> [207,]  6.475197 12.7280365  1.5860613  1.9065236  1.2603761  1.2042653
#> [208,] 13.449566  3.5507669  4.0729231  2.3748557  2.8451856  6.8164383
#> [209,]  9.272792  6.6142312  1.4983986  1.2489188  1.1036584  0.8990926
#> [210,]  8.182209  2.9717834  2.2477045  1.6046511  2.4486637  1.1123424
#> [211,] 17.328198 12.3008004  1.6370705  1.1898949  1.2738405  0.7082906
#> [212,]  8.124222  4.7570494  0.9144670  1.2604830  2.4885749  3.2978641
#> [213,]  9.107001 11.1992636  2.0744985  1.2566494  1.1835769  0.9604573
#> [214,] 29.522475 32.9779829  9.3735526  3.7872929  4.4814640  4.3687764
#> [215,] 14.498208  6.9301348  1.2572926  2.0008521  1.8043999  3.7124902
#> [216,] 11.409134 20.1313979  5.7836377  3.4359633  4.2516890  2.9988218
#> [217,] 11.929447  3.2108003  1.4496802  0.7222339  0.3174470  0.9924316
#> [218,] 10.093599  4.7545561  3.7157222  1.0056997  1.1062959  1.3810797
#> [219,] 12.177171  5.5018501  1.6020667  1.2266214  1.5883606  1.0821846
#> [220,] 36.242723  0.8345537  2.4107290  1.3022915  0.3883929  3.0613329
#> [221,] 19.599276  2.4312981  1.6731475  1.2350434  1.3140758  1.6311867
#> [222,] 14.072766  2.4410401  1.9088483  2.2332866  3.3042516  9.3404490
#> [223,]  7.218482  1.3376698  0.8871181  1.0003251  2.3201383  3.2977488
#> [224,]  6.901438  2.9510627  1.5962180  1.3305585  1.5869344  0.7110362
#> [225,] 13.450296  1.4809444  1.8571600  0.9282579  0.8816689  0.8161165
#> [226,] 13.605987  1.6394935  0.9642176  1.4882136  2.6104960  0.8142880
#> [227,] 12.314762  2.7131307  1.4197318  1.1643671  1.7577435  0.6076500
#> [228,]  7.590103  2.7208674  5.4797986  1.3242143  0.8818348  1.8950272
#> [229,]  7.332747  1.9624676  2.0823909  1.4823271  0.5800950  0.8681233
#> [230,] 16.548811  4.2653772  2.4189971  1.4093413  1.0625268  1.7878525
#> [231,] 10.951102  1.4415141  0.6829715  0.9713386  0.8316789  2.4432634
#> [232,]  6.009355  6.3053116  1.0956704  0.7984131  4.7384425  1.1691629
#> [233,]  7.023834  1.3651084  1.6495219  1.6874512  2.1739164  0.6082057
#> [234,] 17.969903  2.3737642  1.8181882  2.1375368  1.1190302  2.7184715
#> [235,] 15.523069  1.3332228  0.6258552  1.2461992  2.3609637  1.5397290
#> [236,] 22.298436  3.2907758  1.8553057  1.8984851  2.1980437  0.6117105
#> [237,] 14.270224  1.3909678  0.7789639  1.5884025  0.9987676  1.4532067
#> [238,] 48.948190  1.0284210  0.7692430  1.1322345  0.7390677  0.9172141
#> [239,]  8.429283  3.3089556  1.1448586  2.0776554  0.7081237  0.7018629
#> [240,]  8.733203  2.0289910  0.9468416  0.5907016  0.9977054  2.4031543
#> [241,] 36.336563  4.1385683  0.6012494  0.8639574  0.5588937  1.1191114
#> [242,]  8.466918  0.9157443  1.1434522  2.2094760  0.9899949  1.3573553
#> [243,] 37.173698  1.5221111  1.8836531  1.3591384  1.6068950  1.3250950
#> [244,]  7.585370  4.7434959 10.3208451  7.5198452  5.1924942  7.4606921
#> [245,] 15.744467  0.8966191  4.0813071  1.9684485  2.4881793  1.1323411
#> [246,] 12.623149  1.1055391  1.0593543  1.2736522  1.1265451  0.9137842
#> [247,] 18.311717  2.1426704  0.7711548  1.4911026  1.6195632  0.7791169
#> [248,]  7.891069  2.5529729  0.6597032  1.4752408  1.2612363  2.2151465
#> [249,] 10.577262  1.2779711  0.9025514  0.8489029  0.9714456  1.3504664
#> [250,]  6.983317  3.0038349  3.4018424  0.7795972  0.6436638  0.8859199
#> [251,] 26.847043  0.9516289  1.2426072  0.7852727  0.6816919  0.7212922
#> [252,] 27.530204  2.2884949  1.1897144  1.9448581  1.2780089  1.5069558
#> [253,]  8.680302  1.2064085  3.1289222  2.3273242  1.0823280  0.8406826
#> [254,] 12.325762  1.4888855  2.3360896  3.7924995  1.5190388  1.2103658
#> [255,]  7.431142  3.7456192  1.1988914  1.2180100  0.9869380  3.0861586
#> [256,] 32.822026  1.2825988  1.0574800  1.0722129  1.1795036  1.2051209
#> [257,]  7.606320  2.0376778  4.1413707  1.8208063  1.0574269  0.8531926
#> [258,] 24.234033  1.3975417  1.1464612  2.2197164  0.9115854  1.2716000
#> [259,] 29.331580  1.7128804  1.5086974  1.6880203  1.7151794  3.3723996
#> [260,] 10.053029  4.2622263  4.0613345  4.7668177  1.9750405  7.2008362
#> [261,]  8.499931 10.8775962  2.0085944  1.3901718  1.2851080  1.4112597
#> [262,] 10.851902  5.6263319  1.9753500  3.9307335  0.9776880  0.5549713
#> [263,] 19.293587  2.2437679  1.4860931  2.6235013  1.6863014  1.9277814
#> [264,]  9.056283  4.7748524  0.7181936  1.7927780  2.2650295  0.6430500
#> [265,]  7.875179  3.6725097  0.9177727  1.3846062  2.9350097  1.8333611
#> [266,]  6.004867  1.4865913  0.9132668  0.9056954  1.7963282  1.2572394
#> [267,] 11.772597  2.6324408  1.9280117  2.0294321  0.8118344  2.3565676
#> [268,] 12.411254  2.3575703  2.6556386  1.2086582  2.5790586  1.2460659
#> [269,] 10.253400  2.5068868  1.9025568  0.6262518  0.3797283  0.6462802
#> [270,] 22.712024  2.3856408  1.7113813  1.6614627  1.3541465  1.1606372
#> [271,] 21.135182  2.3415363  0.5995459  0.9037564  4.2172438  0.8683411
#> [272,] 14.693395 13.1413414  3.8877674  1.0127424  1.8410028  1.3599777
#> [273,]  5.424191  3.4967094  2.8280272  0.7819454  1.9840778  3.7588111
#> [274,] 14.252208  2.2378318  1.3809292  3.2484609  1.0494356  0.7461209
#> [275,] 10.306709  6.2710399  3.6601638  1.2579114  4.9087139  1.4581533
#> [276,]  9.946645  2.3392327  3.4320415  1.5124113  1.5010624  1.3110424
#> [277,] 12.286725  1.3784674  2.5297318  1.9039024  0.5634405  0.6452318
#> [278,]  3.720747  2.8300997  1.3033818  1.1694250  2.3770584  1.0808330
#> [279,]  8.913228 20.5487958  1.7023216  0.5244450  0.6126809  1.9524298
#> [280,] 10.548186  7.0216989  1.7027471  1.5719089  4.5378500  0.4450227
#> [281,]  7.623836  2.9844695  2.4895556  0.6686943  0.8286736  2.0590665
#> [282,]  8.275864  2.5977642  1.4409184  0.6077349  0.4883897  0.6436412
#> [283,] 13.368892  3.6860753  1.2846304  2.4437645  2.2267416  5.0920421
#> [284,]  6.477014  2.7388195  0.9171393  1.7032966  1.2765683  1.2882243
#> [285,] 15.446161  7.1712003  3.0867770  1.7284679  0.8764545  0.7605169
#> [286,] 15.247879  3.5631764  3.0352575  2.6091997  1.2622017  1.2992411
#> [287,] 21.846241  1.3159314  1.4183205  1.3439006  0.9087290  1.2303949
#> [288,] 15.263625  3.0912414  0.3874362  1.2446740  1.0632624  1.8439214
#> [289,] 30.160732  2.1829246  3.8299777  4.6121742  7.1126551 17.7758018
#> [290,] 16.345482 10.8274162  3.1071194  9.4332732  7.2199168  6.7229946
#> [291,] 29.573806  2.1490680  1.4989893  0.7069137  3.2422638  0.8098563
#> [292,] 11.886073  4.0618693  1.2839473  2.0089311  2.0617638  1.5598878
#> [293,] 18.861706  5.9024520  0.8071390  3.1318633  1.4144284  2.6609808
#> [294,] 10.832342  4.8892219  1.6414754  0.9277039  3.4960976  1.4254677
#> [295,] 10.890201  3.2441056  2.2138471  1.0924818  0.9982205  4.0806545
#> [296,] 21.990054  2.7041256  1.0588398  2.4965225  7.0218231  5.2378410
#> [297,]  9.903199  5.0695856  1.5190768  0.8984383  1.9829024  1.6122094
#> [298,]  7.985402  4.6187998  2.7511383  2.3499466  0.9935670  0.8781684
#> [299,]  8.026003  3.1616197  1.9241266  3.8817379  0.9820600  1.0990698
#> [300,] 17.193879  6.6338466  2.5429485  4.1596868  2.9978880  2.8980835
#> [301,] 10.064425  2.4227781  0.7495110  0.9660444  1.2856366  8.5908981
#> [302,] 16.849357  2.1121730  1.5320194  1.1429373  2.0583645  0.8866873
#> [303,]  9.784175  3.5644477  0.7337240  0.7934247  3.0493786  1.0827110
#> [304,] 12.405713  3.8613171  1.4839500  2.4321155  0.7728216  3.7848956
#>               [,7]       [,8]       [,9]      [,10]
#>   [1,]  16.6705559  3.9393549 12.6836125  4.6662333
#>   [2,]   1.4051127  1.8892754  8.5101089  5.6601275
#>   [3,]   2.5254366  1.6354947  1.2771725  2.7107924
#>   [4,]   8.5031652  8.1057414  4.9310778  5.9196308
#>   [5,]   4.8006298  6.0614600  2.3696945  1.0142679
#>   [6,]   3.2866876  4.0335227  0.9154458  6.8099960
#>   [7,]   4.7114144  5.1953358  9.9467333  5.0476933
#>   [8,]   2.3344404  5.5532272 16.8223913 14.2478127
#>   [9,]   1.1203445  2.4565590  3.1162624  2.9405877
#>  [10,]   4.0309393  3.1274407  5.7643120 83.4384340
#>  [11,]   8.0181894  7.4095041  2.0106173  5.6369094
#>  [12,] 368.5507101 26.6922365 11.9312358  7.6632068
#>  [13,]   9.0580295 20.6886207 16.2058814  8.8388537
#>  [14,]   6.7998273  7.7192644 14.4200574 12.1705400
#>  [15,]   6.2542205  2.9299396 14.3922853  3.9787614
#>  [16,]   0.8281706  2.5768047  2.6267384  1.0460868
#>  [17,]   4.5007904  2.5280852  1.5627216  2.5763861
#>  [18,]   1.2608716  1.9517381  1.0241169  4.2870238
#>  [19,]   1.8233862  1.5792147  3.1518876  3.7573696
#>  [20,]   2.6324882  1.1804780  2.0271923  0.8242291
#>  [21,]   0.8384505  1.0250859  1.0330641  1.7799533
#>  [22,]   1.0420648  3.4059951 10.4888038  2.2982588
#>  [23,]   0.9739303  1.4007300  1.7130947  2.5445478
#>  [24,]   3.9407724  9.8464313  5.5848392  4.7497312
#>  [25,]   1.1400281  0.9477386  2.9214234  4.6732517
#>  [26,]   1.1078167  3.0412812  2.6516146  3.0061431
#>  [27,]   0.7382834  2.5069466  2.3232570  1.8371983
#>  [28,]   1.3041764  6.7405437  2.2608109  2.4816234
#>  [29,]   1.5006683  1.3664387  1.9152105  0.7192621
#>  [30,]   1.1087059  1.2770772  2.2526971  1.6946991
#>  [31,]   1.6203348  2.7287176  2.1136366  1.2464276
#>  [32,]   2.4207521  2.5473503  1.1913323  2.3534327
#>  [33,]   2.0148614  1.0631508  1.3791786  0.7814143
#>  [34,]   1.1397103  2.1365329  1.9852146  1.8669735
#>  [35,]   1.1622410  1.0088926  1.8893461  1.0099410
#>  [36,]   2.5655760  1.8499642  0.8749807  1.1790696
#>  [37,]   1.1456203  2.9345294  2.2027585  0.7328388
#>  [38,]   0.5907987  0.8344928  2.9031173  1.0417202
#>  [39,]   1.3554204  1.5271277  1.6536140  1.2075389
#>  [40,]  10.9794985 14.7783236  7.0596635 17.0708180
#>  [41,]   1.5884563  3.3689829  2.1177104  1.7019985
#>  [42,]   2.8977776  1.1041121  4.1752957  2.3416587
#>  [43,]   0.8086889  1.2723800  4.4363163  1.6753684
#>  [44,]   3.1740453  2.9390955  2.7781380  6.1399654
#>  [45,]   4.6182197  2.0921868  2.9847589  1.2048174
#>  [46,]   0.8387095  0.9221359  1.0350076  1.3041059
#>  [47,]   0.8999723  1.4643506  4.5642632  2.7062459
#>  [48,]   1.4076029  2.5887572  2.6939086  8.6232566
#>  [49,]   1.8565533  1.6244444  1.0547256  1.4755674
#>  [50,]   5.0794306  2.0299273  0.8727374  2.0247002
#>  [51,]   1.7286515  1.4315172  1.2512681  2.1693939
#>  [52,]   0.6036299  1.4420181  1.1424424  0.6969644
#>  [53,]   1.0801131  1.4424752  0.8720462  1.2396423
#>  [54,]   1.1679771  1.3458452  0.9705004  1.8432886
#>  [55,]   0.9174524  1.4892983  1.7815921  1.8093415
#>  [56,]   2.1269833  2.9437952  1.7579666  3.3835114
#>  [57,]   1.1226109  0.5417174  1.8205823  0.6075720
#>  [58,]   1.5758804  1.5602671  1.2691091  1.2850823
#>  [59,]   1.2451525  1.3423232  1.1028535  2.2806710
#>  [60,]   1.4160825  2.1473467  0.8074355  0.9735020
#>  [61,]   2.1988102  1.0731842  0.5016253  0.4712096
#>  [62,]   1.3129769  1.9520431  1.1733157  1.8681728
#>  [63,]   0.4869174  2.1242025  0.6071644  1.3041592
#>  [64,]   2.6103975  2.2587657  2.1217191  2.7868928
#>  [65,]   3.7193669  2.0445902  2.3481899  5.0523303
#>  [66,]   2.2457691  2.5160161  0.7673997  0.7246077
#>  [67,]   2.8815416  1.1872295  0.8836625  4.2282394
#>  [68,]   1.2411510  1.6594699  3.3694255  4.4974496
#>  [69,]   0.8662612  1.9081322  1.1033233  0.7937061
#>  [70,]   3.1830910  3.0047369  2.0906207  6.1845926
#>  [71,]   1.4962466  1.4743201  3.4573741  2.1996709
#>  [72,]   2.6223836  5.4684035  1.9448429  3.6503368
#>  [73,]   2.4257604  1.7814882  0.9066258  2.6211831
#>  [74,]   1.8309729  4.3705671  0.7200329  0.7936998
#>  [75,]   1.4114337  0.5898974  0.6871229  0.7409764
#>  [76,]   1.9897396  1.9260251  0.8855145  5.2247195
#>  [77,]   1.1624190  0.4411183  1.1803448  1.0462132
#>  [78,]   0.6419822  0.6481436  1.0476347  1.1129594
#>  [79,]   2.4342115  0.7066147  0.8681109  0.9040288
#>  [80,]   0.7134450  1.2242715  1.3899894  2.5043774
#>  [81,]   1.2461170  1.0524755  2.3300186  1.3592313
#>  [82,]   3.4879125  2.1062042  6.6403488  3.8695841
#>  [83,]   3.2812032  0.7496716  0.9332412  1.0725203
#>  [84,]   1.0594116  1.2700979  5.7268691  9.0812523
#>  [85,]   3.1750284  1.2294206  1.2322179  1.4287317
#>  [86,]   3.5887455  1.5874104  1.1367136  0.9756226
#>  [87,]   0.8849195  3.7379573  1.2457945  1.1247276
#>  [88,]   0.9290039  1.6386598  3.4872551  1.9044865
#>  [89,]   1.4654372  1.5801928  0.6634364  3.1969237
#>  [90,]   0.7577426  3.6378850  1.6899921  1.9620758
#>  [91,]   1.1760440  2.5755545  1.7237769  1.7549496
#>  [92,]   2.0749383  4.2221420  3.3725851  2.1547702
#>  [93,]   1.8332339  0.9426340  1.9157374  1.8589861
#>  [94,]   1.3522857  1.0228562  2.1017887  0.6751169
#>  [95,]   2.0335491  1.1099383  4.7376521  1.2109074
#>  [96,]   4.8262420  4.9155516  3.3745536  2.6813799
#>  [97,]   4.7647333  1.4201611  1.2697267  3.4818550
#>  [98,]   6.1547026  0.9183324  5.5693458  1.7772055
#>  [99,]   1.0006710  2.6947161  1.7300871  2.0824354
#> [100,]   6.0198981  2.1950139  3.0195906  6.6369392
#> [101,]   0.9264394  0.9109918  0.8090348  2.8727129
#> [102,]   3.9903807  4.1405604  3.2626475  2.4668789
#> [103,]   2.2486480  1.0238910  0.6477315  0.6356282
#> [104,]   1.8342175  2.8318972  4.4386494  6.4941045
#> [105,]   0.8425292  1.1767709  1.1984779  0.8052026
#> [106,]   1.7281643  1.5389821  1.9559182  2.8925597
#> [107,]   0.8532851  1.2153057  3.3638045  3.6540235
#> [108,]   2.8582312  1.3050575  2.7000785  3.0676140
#> [109,]  29.1107743  8.9841158 18.0845060 14.5051275
#> [110,]   6.2551120 21.9020076  9.1956047 13.2196536
#> [111,]   2.9514450  0.5766155  0.6095989  1.4695925
#> [112,]   2.0214526  0.8407347  2.8877720  0.7536819
#> [113,]   1.4221998  2.1104167  1.0443680  1.1001056
#> [114,]   2.8789641  0.6545129  1.1604225  5.2192962
#> [115,]   0.7202241  2.2503215  0.7411419  0.9311687
#> [116,]   1.2892264  1.2976455  0.5432786  1.3639763
#> [117,]   2.3391412  2.2481836  0.6467917  1.8457628
#> [118,]   1.0434392  1.9635195  0.6206779  0.6268907
#> [119,]   1.1512540  1.9176588  1.0011301  1.1085310
#> [120,]   0.8266483  0.3574074  0.5852640  1.5096701
#> [121,]   4.6977486  8.3159268  7.2763702  4.5320421
#> [122,]   1.2450997  1.9264418  0.9979196  1.8650287
#> [123,]   2.9217401  2.7637575  2.0973101  1.9669065
#> [124,]   2.0802780  1.3565704  1.0436502  1.0048605
#> [125,]   1.0453048  0.9362715  0.8397855  1.0885926
#> [126,]   0.9753486  0.9783575  1.1756320  2.6232246
#> [127,]   0.6738434  2.0666726  0.7238789  1.3385399
#> [128,]   1.8967089  0.9359841  1.0755757  1.5494365
#> [129,]   2.5903508 10.2166764  2.4590111  3.0895549
#> [130,]   5.2774510  1.7776168  1.4940386  1.4752785
#> [131,]   6.6637702  0.9855687  0.9863801  1.0848073
#> [132,]   4.4777016  3.4998041  4.3848323  4.6979863
#> [133,]   0.8033159  2.0970108  2.6147359  2.5683944
#> [134,]   1.0944129  2.7798323  1.8342996  1.3088974
#> [135,]   1.5578847  3.0192100  1.2688386  4.6490962
#> [136,]   3.1766937  2.3652267  1.5442048  3.7481095
#> [137,]   2.5744875  0.6934788  0.7011715  0.8250330
#> [138,]   1.0472552  1.1666459  1.4091007  2.0055508
#> [139,]   1.6090868  0.8661234  1.3497871  0.9969890
#> [140,]   0.7073333  1.0510479  0.7388663  2.1341761
#> [141,]   2.3321844  1.4640443  1.2336450  2.6756421
#> [142,]   1.0357680  4.9303700  2.2616929  2.0476444
#> [143,]   1.9203931  4.3157425  1.5662661  3.5480371
#> [144,]   2.3667420  0.6844740  1.2109742  3.3270236
#> [145,]   2.7317561  1.2280648  5.3207392  2.5983212
#> [146,]   1.1107592  0.5736345  1.4027698  1.9980263
#> [147,]   4.9353659  1.7541951  1.2976586  1.2024945
#> [148,]   3.4152866  4.2009481 11.7678895  5.9820838
#> [149,]   6.7995222  2.6244370 15.7293291 27.4933459
#> [150,]   3.0738073  1.1464096  2.1168831  2.4880728
#> [151,]   2.8558696  0.8651970  3.0564269  1.0506589
#> [152,]   0.7399610  1.2406490  1.1873902  0.8122534
#> [153,]   1.1659094  0.7163840  2.9333222  1.2573451
#> [154,]   1.0221231  5.0614317  0.9330481  2.1579442
#> [155,]   0.7851793  1.1903170  1.6050458  1.9358653
#> [156,]   1.4513875  0.7511918  0.8468665  0.9628406
#> [157,]   3.3021850  2.2268503 10.6905390  3.9142610
#> [158,]   0.6959352  1.5499569  3.7650084  0.9393897
#> [159,]   0.7307317  1.7411144  1.2271078  1.4340831
#> [160,]   0.7698097  1.0070985  2.8349977  1.0376326
#> [161,]   3.2437548  2.0471281  2.9865997  0.8653533
#> [162,]   1.1011269  1.3886215  0.8088601  1.2743775
#> [163,]   1.1095294  2.4824110  0.9112088  1.5091244
#> [164,]   0.8983774  1.7026125  5.0375639  6.1558244
#> [165,]   3.3671350  1.1320475  0.6424852  0.9007259
#> [166,]   1.2554469  1.5486532  1.6868997  2.1740687
#> [167,]   1.0018079  1.4053283  0.7480867  1.2124463
#> [168,]   1.1200618  1.2221040  1.0511186  1.0956512
#> [169,]   1.6328778  2.3569572  3.2827599  0.7918840
#> [170,]   1.1510139  0.6707605  0.4892949  2.2115092
#> [171,]   0.7713033  1.6503413  3.3818818  1.4235456
#> [172,]   0.8993278  2.0217741  7.9893141  1.5874066
#> [173,]   1.1970436  0.9896094  0.7683983  1.2438986
#> [174,]   2.6470827  0.8641395  1.3196129  0.7805261
#> [175,]   3.4611091  1.3182769  0.9185939  1.1363390
#> [176,]   1.1876407  0.5803122  0.6814881  0.9115942
#> [177,]   1.1612223  1.1002415  1.2069912  1.5657530
#> [178,]   0.9125479  1.3335237  0.7605506  2.2077235
#> [179,]   1.0682931  2.8369611  0.7045924  2.3017188
#> [180,]   3.3201076  3.5414624  2.0814603  0.9652711
#> [181,]   3.5322413  1.5163664  0.8513314  0.7464365
#> [182,]   1.2541963  1.0794834  1.1817319  1.8857258
#> [183,]   1.3814732  2.9609088  0.8445586  1.6615940
#> [184,]   1.5330106  5.8436312  2.5225037  1.3918914
#> [185,]   1.5054677  0.6990026  0.9353517  1.1054137
#> [186,]   0.7524654  1.6317375  0.3565149  0.6345854
#> [187,]   2.7141754  2.3738252  3.5632717  3.2371441
#> [188,]   1.1286317  2.2044506  1.9994151  0.9628832
#> [189,]   2.5088295  2.0734188  3.2059349  0.8097090
#> [190,]   2.1180140  1.1955730  2.3508851  0.9042625
#> [191,]   0.8623377  0.8430057  1.8045946  0.9314512
#> [192,]   1.7894823  1.6847443  4.8299843  0.8708731
#> [193,]   0.9008372  1.9683618  1.6801163  1.0654551
#> [194,]   0.8427978  1.6282543  1.0634246  1.3296969
#> [195,]   1.5199700  1.3035774  0.7855847  1.1404865
#> [196,]   2.9068273  1.2282432  0.7050596  0.7291352
#> [197,]   0.6432683  1.9080530  1.2657524  2.4396550
#> [198,]   0.8380043  1.0602194  1.5808945  1.2396994
#> [199,]   1.1424438  1.3329458  2.4229046  0.8881170
#> [200,]   0.9551900  2.2464994  0.8085853  1.9882865
#> [201,]   0.9568339  0.8312702  2.1435044  0.5824724
#> [202,]   0.6550088  1.3199360  0.6064673  0.8014219
#> [203,]   1.4351381  1.4880347  0.5520116  0.8035649
#> [204,]   0.5902710  1.3445277  1.0722295  0.3963742
#> [205,]   0.6832366  0.8547403  0.3862520  1.1031136
#> [206,]   1.0242406  0.8115036  0.7094366  1.0121175
#> [207,]   3.0253780  2.7828873  1.0840771  1.0654325
#> [208,]   4.1554860  2.1472783  4.3663930  1.6978918
#> [209,]   1.7459930  2.4468305  1.1134989  5.7237347
#> [210,]   0.9438306  1.4345245  1.4969448  1.1049229
#> [211,]   1.3403231  0.7721625  1.7019494  0.8741009
#> [212,]   1.6285898  0.5861596  1.1404782  2.3871309
#> [213,]   0.6575461  1.5343448  2.0084897  1.5952861
#> [214,]   5.1833854  6.9128547  6.2294899  6.8118567
#> [215,]   3.7839089  5.8125355  5.6024219  3.8032235
#> [216,]   5.2246829  5.4272386  6.3309654  4.8141404
#> [217,]   1.9592573  1.0193063  1.6828129  0.9397762
#> [218,]   1.5446783  1.6600821  0.9142957  0.8117865
#> [219,]   0.7822741  2.6865173  2.1378869  7.6498852
#> [220,]   1.4925769  0.8877588  1.1278061  1.6352361
#> [221,]   1.4621031  2.7730210  1.9532176  2.7953254
#> [222,]   1.8667962  2.0606650  4.0360244  3.9268433
#> [223,]   1.7738540  1.5263778  0.9751785  2.1076745
#> [224,]   1.1527890  1.6908949  1.0415619  1.0121212
#> [225,]   1.2895747  2.3469021  0.4466263  0.8163223
#> [226,]   0.6651211  0.5970158  1.3247610  1.4625763
#> [227,]   1.4015631  1.1815958  0.6351696  1.1815951
#> [228,]   2.0293419  4.0250346  7.7302836  1.1729587
#> [229,]   0.8191790  0.5399269  0.4210856  0.6723243
#> [230,]   0.5878229  0.4720985  0.7837111  1.3971337
#> [231,]   0.5536093  1.0560328  1.0615731  2.1994459
#> [232,]   1.3577753  1.6121775  2.1952222  1.6906823
#> [233,]   1.0584649  1.1528130  0.9938408  6.0039082
#> [234,]   0.9776020  0.9124297  0.6996661  0.6038954
#> [235,]   0.7339878  1.1694379  0.9391862  2.4474790
#> [236,]   1.2312629  1.0818754  0.7177458  0.6596831
#> [237,]   1.3288392  0.6577237  3.0952814  1.4644940
#> [238,]   0.5772635  0.9208642  0.6439062  2.2817094
#> [239,]   0.7590522  1.2139370  2.3617136  1.7114530
#> [240,]   0.8369591  1.4786551  1.6739177  1.0664286
#> [241,]   1.0640529  0.6210917  1.1624700  4.4082442
#> [242,]   1.7660561  0.5965770  1.4971078  1.8283028
#> [243,]   1.8223727  4.4336875  4.2790704  5.6843129
#> [244,]   7.5200943 11.2917997  7.0072878 11.9667779
#> [245,]   1.0675273  1.4142309  1.2084081  1.0104560
#> [246,]   1.5503222  3.3350515  0.7954600  1.9193953
#> [247,]   1.2470300  1.7579354  1.9917070  1.6582833
#> [248,]   4.3017163  1.1522249  0.8470168  1.7278311
#> [249,]   3.1870602  1.3673859  0.8230706  1.3535340
#> [250,]   0.5642690  1.6507311  0.6603577  0.7193715
#> [251,]   0.7288964  0.6814204  0.8171885  0.8143404
#> [252,]   4.8737442  1.2506752  1.5606508  0.6426032
#> [253,]   0.7129284  3.6861896  1.8720532  0.3998600
#> [254,]   1.3476743  2.1640919  1.4003867  1.0357903
#> [255,]   1.9550383  0.8255818  1.4864942  1.5410550
#> [256,]   0.9653408  4.3404826  0.4391243  1.1594963
#> [257,]   0.7597211  0.9758199  0.8495326  0.4632282
#> [258,]   0.7716940  1.5287226  0.6635938  0.8866220
#> [259,]   2.7195593  3.4218473  1.9228661  1.9522033
#> [260,]   2.4262240  4.1942821  2.5477718  3.7246547
#> [261,]   0.6607998  0.8979611  1.3538157  1.3465829
#> [262,]   2.3724907  1.9684693  2.6802032  0.7003049
#> [263,]   1.4094752  0.9792052  1.0697656  2.5398265
#> [264,]   1.2484687  0.9031132  1.6978259  3.2199674
#> [265,]   1.2812601  2.7871380  0.9747324  2.3514344
#> [266,]   2.0439217  0.6546685  1.1242032  0.8854107
#> [267,]   2.4841241  0.5135270  1.5455644  1.8709193
#> [268,]   1.4954553  1.3698468  3.5040372  0.7761593
#> [269,]   1.2628517  1.2830548  0.6758946  1.4973112
#> [270,]   1.2921394  2.2329105  1.9394040  0.9942180
#> [271,]   1.7441615  1.2556484  0.6750624  6.7617183
#> [272,]   1.5493823  2.0950658  3.9707992  0.9641265
#> [273,]   1.5947394  1.0952820  0.7727470  1.7253339
#> [274,]   1.8480060  0.6306016  0.8299733  1.4202786
#> [275,]   0.5036256  0.6465700  2.7185628  0.4778146
#> [276,]   0.9107255  0.7842229  0.9979202  1.1235513
#> [277,]   0.7289274  0.4756666  1.0321411  0.6339907
#> [278,]   0.6594218  2.1303068  0.4822837  0.7142435
#> [279,]   1.7606558  3.1888119  0.5016786  0.9792865
#> [280,]   2.2450354  1.7089414  0.7695891  0.6417611
#> [281,]   1.4112084  2.0933665  0.8099488  0.6855990
#> [282,]   0.3829889  0.7150678  1.2559307  4.6297885
#> [283,]   0.9390250  1.3720374  1.5150417  1.5480676
#> [284,]   0.5313861  1.1580233  5.9998160  0.9478869
#> [285,]   1.2775443  0.7622524  0.4860300  1.0320273
#> [286,]   2.1188732  1.3444651  1.7827517  0.7362549
#> [287,]   1.1617134  1.2982274  1.2474841  4.2771967
#> [288,]   4.0649106  1.3848321  2.2172880  1.1377948
#> [289,]  28.2714094 64.6884547 55.7075298 18.5791101
#> [290,]  16.5557694 26.3541590 13.3611943 26.2180315
#> [291,]   1.6324255  1.7794593  2.3840131  1.8191813
#> [292,]   5.2549257  7.1917431  2.0645373  1.8795170
#> [293,]   1.2319785  1.9527396  2.8975341  1.8114408
#> [294,]   0.9446471  1.6330187  1.5479733  0.9397171
#> [295,]   2.2880572  1.3073515  1.4008479  1.7009269
#> [296,]   2.1624745  7.1449794  2.4477001  2.7401197
#> [297,]   0.7278942  1.6955389  1.8213959  1.4677326
#> [298,]   1.2995342  1.9019177  0.7192067  1.6452383
#> [299,]   2.6067137  1.0032701  1.3680686  1.1659600
#> [300,]   3.8546842 14.7346774  4.6083764  8.4467226
#> [301,]   1.2366020  1.1165065  0.4524267  1.2012111
#> [302,]   1.4605243  0.7481171  1.0054356  1.9164425
#> [303,]   0.5185713  1.1693612  2.2991042  1.1112323
#> [304,]   2.7303024  2.4602090  2.3507284  1.6202401
#> 
#> $df
#>           [,1]
#>  [1,] 6.000000
#>  [2,] 5.776183
#>  [3,] 5.776183
#>  [4,] 5.188033
#>  [5,] 5.139921
#>  [6,] 5.271451
#>  [7,] 5.271440
#>  [8,] 5.237644
#>  [9,] 5.237644
#> [10,] 5.240770
#> 

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$get_last_draw`
## ------------------------------------------------

data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
set.seed(123)
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$is_normalised`
## ------------------------------------------------

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.

# estimate the model
set.seed(123)
posterior      = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# check normalisation status beforehand
posterior$is_normalised()
#> [1] TRUE

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) %*% BB         # set negative diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()
#> [1] TRUE

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$set_normalised`
## ------------------------------------------------

# This is an internal function that is run while executing normalise_posterior()
# Observe its working by analysing the workflow:

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# check normalisation status beforehand
posterior$is_normalised()
#> [1] TRUE

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) %*% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()
#> [1] TRUE