R6 Class Representing PosteriorBSVART

The class PosteriorBSVART contains posterior output and the specification including the last MCMC draw for the bsvar model with t-distributed structural shocks. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BSVART with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bsvar_t$new()

specify_posterior_bsvar_t$get_posterior()

specify_posterior_bsvar_t$get_last_draw()

specify_posterior_bsvar_t$is_normalised()

specify_posterior_bsvar_t$set_normalised()

specify_posterior_bsvar_t$clone()

Method new()

Create a new posterior output PosteriorBSVART.

Usage

specify_posterior_bsvar_t$new(specification_bsvar, posterior_bsvar)

Arguments

specification_bsvar

an object of class BSVART with the last draw of the current MCMC run as the starting value.

posterior_bsvar

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBSVART.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bsvar_t$get_posterior()

Examples

data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
set.seed(123)
estimate       = estimate(specification, 10)
estimate$get_posterior()

Method get_last_draw()

Returns an object of class BSVART with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bsvar_t$get_last_draw()

Examples

data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)

# run the burn-in
set.seed(123)
burn_in        = estimate(specification, 10)

# estimate the model
posterior      = estimate(burn_in, 10)

Method is_normalised()

Returns TRUE if the posterior has been normalised using normalise_posterior() and FALSE otherwise.

Usage

specify_posterior_bsvar_t$is_normalised()

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)

# estimate the model
set.seed(123)
posterior      = estimate(specification, 10)

# check normalisation status beforehand
posterior$is_normalised()

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) %*% BB         # set negative diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()

Method set_normalised()

Sets the private indicator normalised to TRUE.

Usage

specify_posterior_bsvar_t$set_normalised(value)

Arguments

value

(optional) a logical value to be passed to indicator normalised.

Examples

# This is an internal function that is run while executing normalise_posterior()
# Observe its working by analysing the workflow:

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10)

# check normalisation status beforehand
posterior$is_normalised()

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) %*% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bsvar_t$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# This is a function that is used within estimate()
data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)
estimate       = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(estimate)
#> [1] "PosteriorBSVART" "R6"             

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$get_posterior`
## ------------------------------------------------

data(us_fiscal_lsuw)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)
estimate       = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
estimate$get_posterior()
#> $B
#> , , 1
#> 
#>            [,1]      [,2]     [,3]
#> [1,] 39.3278839  0.000000 0.000000
#> [2,] 16.2877941 16.592296 0.000000
#> [3,] -0.6580358  5.169431 9.618209
#> 
#> , , 2
#> 
#>          [,1]     [,2]     [,3]
#> [1,] 46.24334  0.00000  0.00000
#> [2,] 33.74912 32.19528  0.00000
#> [3,] -3.41989 16.68778 49.81872
#> 
#> , , 3
#> 
#>           [,1]     [,2]     [,3]
#> [1,]  43.01484  0.00000   0.0000
#> [2,]  27.91174 43.70968   0.0000
#> [3,] -29.31973 15.32813 108.5227
#> 
#> , , 4
#> 
#>           [,1]      [,2]     [,3]
#> [1,]  45.71726  0.000000   0.0000
#> [2,]  25.62723 53.253836   0.0000
#> [3,] -37.11298  2.956566 140.0355
#> 
#> , , 5
#> 
#>           [,1]      [,2]     [,3]
#> [1,]  49.21011  0.000000   0.0000
#> [2,]  17.74305 63.494266   0.0000
#> [3,] -39.84100 -1.072434 168.9216
#> 
#> , , 6
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  57.102178   0.00000   0.0000
#> [2,]  -9.675433  68.14665   0.0000
#> [3,] -33.616695 -11.39240 158.0512
#> 
#> , , 7
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  64.750939  0.000000   0.0000
#> [2,]  -2.830093 67.064885   0.0000
#> [3,] -30.208826 -3.112429 174.3969
#> 
#> , , 8
#> 
#>             [,1]      [,2]    [,3]
#> [1,]  62.8125401   0.00000   0.000
#> [2,]  -0.6695388  69.50007   0.000
#> [3,] -27.5947567 -13.75499 176.579
#> 
#> , , 9
#> 
#>            [,1]      [,2]    [,3]
#> [1,]  75.099725   0.00000   0.000
#> [2,]  -6.089077  69.99302   0.000
#> [3,] -30.419405 -15.16653 161.269
#> 
#> , , 10
#> 
#>            [,1]      [,2]     [,3]
#> [1,]  78.133167   0.00000   0.0000
#> [2,]  -6.915948  71.32445   0.0000
#> [3,] -30.616835 -15.92941 177.6508
#> 
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.1533909 -0.1773105 -0.0604695 -0.7881146
#> [2,] -0.3516446  1.5552609 -0.1526818  1.1840918
#> [3,] -1.0020651  0.2755269  2.0807280  1.0451583
#> 
#> , , 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  1.03309877 -0.06510465  0.03217039 -0.1067230
#> [2,]  0.04520958  1.04681015 -0.11369714  0.1055582
#> [3,] -0.10068165  0.03958292  1.10099053  0.1768488
#> 
#> , , 3
#> 
#>            [,1]         [,2]        [,3]        [,4]
#> [1,] 0.91483023 -0.036660045  0.12992560 -0.20591560
#> [2,] 0.07413581  0.978744111 -0.09581856 -0.19771732
#> [3,] 0.01031459 -0.006563508  0.98964651 -0.05530988
#> 
#> , , 4
#> 
#>              [,1]          [,2]       [,3]       [,4]
#> [1,]  0.939161603 -0.0005415391 0.07491541 -0.0280586
#> [2,] -0.048585850  0.9729870381 0.06935002 -0.1860023
#> [3,] -0.005414629 -0.0146762160 1.00528819 -0.1606658
#> 
#> , , 5
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.96109295 -0.02573731 0.04609051 -0.2867948
#> [2,] -0.01709441  0.96552775 0.02675456 -0.2920487
#> [3,] -0.00959742 -0.01660219 1.01262384 -0.1617306
#> 
#> , , 6
#> 
#>             [,1]         [,2]       [,3]        [,4]
#> [1,]  0.94030470 -0.017784850 0.07464228 -0.18366961
#> [2,] -0.04068988  0.967161235 0.05447752 -0.29648629
#> [3,] -0.01578156 -0.008698563 1.02143093 -0.07317296
#> 
#> , , 7
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.95627560 -0.02095900 0.05980373 -0.1724243
#> [2,] -0.01490755  0.96706519 0.02153141 -0.3088362
#> [3,] -0.01110167 -0.01280193 1.01553226 -0.1122727
#> 
#> , , 8
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.92923585 -0.01792009 0.08983027 -0.1678449
#> [2,] -0.02939004  0.98357203 0.03397876 -0.1850050
#> [3,] -0.02193683 -0.01109149 1.02750074 -0.1068470
#> 
#> , , 9
#> 
#>            [,1]         [,2]       [,3]        [,4]
#> [1,]  0.9620141 -0.020438250 0.04939848 -0.18401878
#> [2,] -0.0398042  0.956769431 0.05434646 -0.39753934
#> [3,] -0.0149524 -0.007537312 1.01786585 -0.07768651
#> 
#> , , 10
#> 
#>              [,1]         [,2]       [,3]        [,4]
#> [1,]  0.958563401 -0.014712697 0.05214279 -0.14060851
#> [2,] -0.031862249  0.960230702 0.04291961 -0.37430964
#> [3,] -0.009384833 -0.005048006 1.00808383 -0.07390716
#> 
#> 
#> $hyper
#> , , 1
#> 
#>           [,1]       [,2]
#> [1,]  7.413741  1.4936374
#> [2,]  5.498240  0.4553124
#> [3,]  2.175053  1.1248598
#> [4,] 37.230551 11.5263567
#> [5,] 58.054742  8.4424441
#> [6,] 27.578742 15.5970321
#> [7,]  2.969482  0.9962926
#> 
#> , , 2
#> 
#>           [,1]       [,2]
#> [1,] 240.84769  0.9724895
#> [2,]  96.90712  0.9926595
#> [3,]  11.63167  0.7344653
#> [4,]  63.69760 13.9895262
#> [5,]  20.70084  8.8519567
#> [6,]  50.39444 12.1605699
#> [7,]   4.96400  1.1961222
#> 
#> , , 3
#> 
#>            [,1]       [,2]
#> [1,] 233.111913  0.4981424
#> [2,] 194.536070  0.7111509
#> [3,] 272.532050  2.5652354
#> [4,]  89.661153  9.3755023
#> [5,]  49.445931 12.9433084
#> [6,]  79.415008 15.3171479
#> [7,]   5.510131  1.2049897
#> 
#> , , 4
#> 
#>            [,1]       [,2]
#> [1,]  87.498532  0.3603473
#> [2,] 687.312560  2.7883117
#> [3,] 848.847910  0.7645959
#> [4,]  98.749135  7.8255039
#> [5,]  95.705693 10.7231452
#> [6,]  99.162193 15.0192972
#> [7,]   7.154731  1.4945665
#> 
#> , , 5
#> 
#>            [,1]       [,2]
#> [1,]  159.86540  0.5816910
#> [2,]  414.12620  2.8598495
#> [3,] 2345.41188  0.7570182
#> [4,]  196.71779  5.1686966
#> [5,]  208.13785 15.5972597
#> [6,]  226.70791  6.8061672
#> [7,]   12.52435  1.1723254
#> 
#> , , 6
#> 
#>            [,1]       [,2]
#> [1,]  314.80626  0.4654010
#> [2,]  216.05496  0.7037555
#> [3,] 2433.84162  0.9532637
#> [4,]  297.69168  6.9965654
#> [5,]  235.35690 10.4834699
#> [6,]  249.69889 11.3217098
#> [7,]   19.48551  0.9507550
#> 
#> , , 7
#> 
#>            [,1]       [,2]
#> [1,]  717.13541  1.3200478
#> [2,]  518.69540  0.4562101
#> [3,] 1694.50371  0.8708359
#> [4,]  459.40740  8.1333548
#> [5,]  244.55341  8.2395462
#> [6,]  529.46320 10.6100118
#> [7,]   33.31646  0.9180586
#> 
#> , , 8
#> 
#>           [,1]      [,2]
#> [1,]  299.4573 0.6199924
#> [2,]  549.1073 0.7527425
#> [3,] 4935.6769 0.3934605
#> [4,]  706.9299 8.8799910
#> [5,]  461.9361 4.1026810
#> [6,]  689.3890 8.9231542
#> [7,]   43.7731 0.7738855
#> 
#> , , 9
#> 
#>            [,1]      [,2]
#> [1,]  399.66122 0.5567372
#> [2,]  360.06573 0.8672957
#> [3,] 1240.11308 0.3022840
#> [4,] 1011.04977 7.2244609
#> [5,]  984.05714 8.2858283
#> [6,]  849.80399 4.2317218
#> [7,]   66.29617 0.7885460
#> 
#> , , 10
#> 
#>            [,1]      [,2]
#> [1,]  983.76478 0.7896129
#> [2,]  443.73160 0.6233231
#> [3,] 1845.31569 0.3977367
#> [4,] 1260.29043 8.6063866
#> [5,] 1274.95355 8.1208398
#> [6,] 1076.22454 6.0172879
#> [7,]   94.68655 0.7872110
#> 
#> 
#> $lambda
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>   [1,]  8.807952  3.9866811  2.7123524  6.9716889  1.9615892  7.2204391
#>   [2,] 44.404426  3.0415160  2.3700824  2.1891744  4.0827428 10.1816946
#>   [3,] 13.826366  9.4513187  0.9695234  1.0501325  2.2923493  1.4475428
#>   [4,]  7.186022  6.9173321  1.0980436  1.6595250  2.0211140  3.1638718
#>   [5,] 11.755645  2.0670626  1.0598345  0.8203320  1.9040499  4.9552532
#>   [6,] 31.386016  4.7318014  1.3838658  4.4355925  3.9297599  8.6076164
#>   [7,] 45.666855  4.4307721  5.6688648  3.7584436  3.4157215  2.6298885
#>   [8,]  8.637378 13.1828305  2.1451034  1.9152381  3.1067123  6.8538299
#>   [9,] 12.343676  2.5610572  2.1604659  3.3197775  3.9971608  1.3087868
#>  [10,] 12.142848  1.8122923 12.9152702  1.9111053  3.1297776  3.8888629
#>  [11,] 13.689296  6.2198645  7.1109154  8.5457743  2.3090232  5.6354758
#>  [12,] 19.034471  3.9223316  4.0209827  7.1926530 12.2278432 37.0503388
#>  [13,]  7.918713  2.8773280  4.2945377  7.0317450 11.9119118 10.6265430
#>  [14,]  9.362781  1.8946860  7.2636149  8.9302194 14.6842470 21.0263129
#>  [15,] 13.940470 13.5271474  3.6810841  2.7046082  6.3744126  4.0535337
#>  [16,] 17.392385 19.8181436  6.3879907  3.8616408  2.3320851  1.5819638
#>  [17,] 15.027218  9.2851695  9.4486614  1.8056091  3.7808777  3.0923521
#>  [18,] 24.308314 11.7997762  1.4963439  0.7802450  1.0987330  1.8873450
#>  [19,] 13.941594 22.1791240  2.7349693  2.7115769  2.3462471  6.3719992
#>  [20,] 39.872408  7.0874494  2.1436028  2.1426978  1.3358934  0.8911324
#>  [21,] 26.454872 10.5358880  1.1739396  4.7796446  1.7843744  2.3053709
#>  [22,] 26.139590 10.0783896  7.6819136  2.4145793  1.8128135  1.2140409
#>  [23,] 15.852703 11.7409417  2.8605739  1.2615375  2.1215862  1.7155855
#>  [24,] 15.759582  9.3912562  5.3828241  4.7655516  5.6253833  2.6134352
#>  [25,] 14.504847 19.4406459  1.6111026  2.1595006  5.0984026  3.4427459
#>  [26,]  9.374929  8.1659548  1.9786528  3.5965239  2.0010030  4.3728579
#>  [27,]  9.893213 14.7664887  4.4810988  2.4734908  2.3925680  2.7552344
#>  [28,] 10.476098 15.0157442  5.8987133  1.0791971  1.9007015  7.7523089
#>  [29,] 14.003353  6.3728213  5.4015358  0.9929274  2.7795454  0.6779039
#>  [30,] 11.795728  3.8016122  2.1709827  4.3672017  1.3170329  1.4182051
#>  [31,] 28.199166 11.5623566  5.4906021  9.4305586  2.0660973  1.1662607
#>  [32,] 15.154949 28.6620365 10.2952016  1.7699488  3.8426499  3.0329953
#>  [33,] 14.729227  6.3456985  2.9737103  0.9468377  0.8852668  1.0524441
#>  [34,]  8.035368  8.8858348  2.0096109  0.6486717  2.6246148  3.7881381
#>  [35,] 26.141419  4.8101244  7.0235301  0.8231817  1.2860470  1.5477392
#>  [36,] 11.131273  6.3482792  1.8933274  0.9569986  2.5863207  0.8366325
#>  [37,] 22.355668  7.0623875  2.8314635  1.1028391  0.9987918  1.7177970
#>  [38,] 14.100575 11.1319927  1.9507101 16.5426720  0.9202916  0.4794630
#>  [39,] 16.986257 11.3426815  1.7016879  1.3374227  0.8466980  0.6285971
#>  [40,] 13.928862  9.5195227  2.5441931  1.9230071  9.2357513  5.8708589
#>  [41,] 11.839394  7.9221438  4.4709879  2.1175655  1.9149465  2.2985703
#>  [42,] 24.830192 14.4883175  2.8522379  2.9837831  1.6291451  1.0310749
#>  [43,] 35.828069  7.5015649  1.7781249  0.9681650  0.9797740  1.9449741
#>  [44,] 25.135858 10.7676335  6.6874135  3.5962111  1.7435558  2.5954805
#>  [45,] 12.093251  5.6945016  2.6542985  2.8409028  2.4974690  1.6015622
#>  [46,] 11.215450  5.7930274  3.2171071  1.4627279  1.5399155  1.6765392
#>  [47,] 17.626811 22.2828000  2.2482064  5.7310672  1.6097566  2.4444226
#>  [48,]  5.722921  9.2941969  2.5282031  4.5325995  4.3333666  1.2335427
#>  [49,] 30.496757  8.9976900  3.2868586  2.0220654  2.2720099  1.8729944
#>  [50,] 11.634614 13.8891933  0.8012739  0.6639659  1.3198183  1.8846305
#>  [51,] 10.914996 18.0310087  2.1026893  3.2798371  1.3350179  4.8502706
#>  [52,] 13.148691  4.9636904  1.9668002  1.7138250  1.0535995  2.2285359
#>  [53,]  8.890706  6.0435575  1.9722344  2.2271778  1.0790507  1.8779420
#>  [54,]  5.861604  3.4388921  4.1168666  1.6182763  3.0110035  3.9347273
#>  [55,] 17.358817  8.1552760  1.3882132  1.7773399  1.6344479  1.3459932
#>  [56,] 22.113932 10.2284565  1.8732547  1.5680172  4.7363059  9.0030372
#>  [57,] 19.627241  3.8603992  1.2906853  1.4636474  2.0174222  0.7238330
#>  [58,]  9.915892  4.4900867  2.8311524  2.1448332  0.8340731  1.6934680
#>  [59,] 13.287098  7.2248821  1.1950215  5.0827827  0.4586522  0.9537814
#>  [60,] 27.506940  5.3563141  2.6116625  2.7895759  1.4926601  2.0075848
#>  [61,] 20.534651 28.1926957  2.4606059  0.9029228  0.4182504  1.9334443
#>  [62,] 13.138929 11.8717524  2.0827223  1.9793078  1.4481294  2.0917774
#>  [63,] 11.004311  8.6644895  1.3919925  0.6264308  1.0527756  1.3726826
#>  [64,] 16.041010 10.6037667  2.0033819  2.3738853  1.9474761  2.2948480
#>  [65,] 20.253823 13.7271735  2.7507396  4.0762522  1.9346811  4.9100838
#>  [66,] 11.243498  5.9409904  2.3709512  1.5593434  2.1454333  0.9850571
#>  [67,]  8.145870  9.3419635  2.1780354  2.5717142  2.4947451  1.0872794
#>  [68,] 10.742978  8.2499554 12.8647361  1.9756405  1.1317187  4.3886775
#>  [69,] 15.655037  3.1358862  1.3392261  1.2302884  2.8105530  4.2350759
#>  [70,] 11.856562  2.8452505  0.9524421  3.1607171  3.5179701  3.8665395
#>  [71,] 16.916310  5.3215740  1.6016080 11.7861861  3.6104866  3.5274616
#>  [72,] 11.555175  8.5541743  5.0213702  4.4940661  4.2269479  2.7106354
#>  [73,] 18.245064  3.2221538  2.8725534  1.6726819  1.9986732  3.3888944
#>  [74,]  7.363116  2.1839256  2.4251230  1.2995060  1.2348142  1.5339384
#>  [75,] 16.504562  3.6339560  1.9203247  3.4658950  1.1734052  0.4637642
#>  [76,]  6.690380  3.3696316  3.6650426  4.5758827  1.0461871  7.1735695
#>  [77,] 14.507387  8.7621345  2.2202118  1.4243763  0.8088935  2.7868528
#>  [78,] 11.711077  6.2395591  0.5415154  2.6449252  1.5484813  0.6170692
#>  [79,]  6.137846  9.3066353  1.0231437  0.8045738  1.8178604  2.2490922
#>  [80,] 12.526559 17.6058601  1.7128813  0.8334381  2.9990739  0.7780018
#>  [81,]  7.624849  3.9557120  3.8046831  1.8326596  2.5122810  1.3119927
#>  [82,] 17.235068  8.3580776  4.9694514  4.1502701  4.0224595  4.8618216
#>  [83,] 13.081892  8.1823452  4.0264328  3.2136745  0.9812110  0.2956606
#>  [84,] 10.322282 23.4820039  4.0293851  6.6101792  1.7178970  2.5449620
#>  [85,] 10.011413  2.5670986  3.2413975  1.6638021  1.1042970  0.9117289
#>  [86,] 14.415025  5.0396166  1.3796349  1.7741108  1.0701806  0.8487375
#>  [87,] 17.560765  2.9577615  5.0013372  1.8132705  1.9039019  1.5809928
#>  [88,] 10.245607  4.7508116  2.1123604  4.0580245  1.5290003  0.7081484
#>  [89,] 14.850835  9.1069831  3.4980070  1.9728022  1.3011202  1.9198730
#>  [90,] 10.142986  4.2796768  2.4808576  2.0465778  3.3436857  1.2761162
#>  [91,] 11.230225  6.8620748  3.2692620  1.5646018  1.7058700  4.2846822
#>  [92,] 12.323248  5.4234461  2.5886284  1.0334184  8.7095395  2.7315546
#>  [93,] 18.434690  5.7988551  1.9648388  0.9610423  0.7123123  1.1903314
#>  [94,]  8.450835  2.5369867  2.0663953  1.2637369  1.8331190  0.8683748
#>  [95,]  8.858530  2.4712877  1.5259594  0.7763356  0.9326834  1.1979112
#>  [96,] 10.621938  4.6216279  5.3864294  3.4749391  5.2756333  8.8944644
#>  [97,] 10.999944  6.4991738  1.4442712  1.2047998  1.0775761  1.5640371
#>  [98,]  9.887718  2.5301907  1.8415008  1.7248441  2.5516916  5.2360985
#>  [99,]  6.033867  3.2630417  1.2468735  1.3364752  1.9051977  0.9012743
#> [100,]  9.774233  4.8508691  1.8248829  1.8275012  5.0476635  8.7966757
#> [101,] 19.882884  2.4208035  2.6155149  2.1198823  2.9356355  1.2871677
#> [102,] 12.585239  3.0084233  1.6252722  1.8581683  3.0212060  2.5483776
#> [103,] 22.626481  2.9605699  1.2721506  1.1643668  1.1061922  2.7479594
#> [104,]  7.974877  2.9797250  1.0476931  2.9466419  1.8860378  1.5679956
#> [105,]  7.788878  2.8029611  2.6406562  2.7044025  0.7875032  0.6756041
#> [106,]  6.575061  5.2815715  1.0440662  0.7361022  1.7615845  1.3717082
#> [107,] 11.920716  5.4986688  2.5073820  1.4466987  4.4136556  3.0377725
#> [108,] 16.186311  6.3547066  1.8755387  3.4306552  1.7560579  2.5477872
#> [109,]  8.032336  9.0825231 21.1672516 29.3362351 16.7224654 29.0779967
#> [110,] 32.477775 17.6803725 12.4986966  7.8592626 15.9503008 13.4144937
#> [111,]  9.786715  1.7727279  1.7809740  0.5806364  1.5713261  0.5272823
#> [112,] 13.671930  6.3728046  1.0332993  1.1803542  1.7800272  3.5216496
#> [113,]  9.493305  5.6959280  1.9176023  1.8826035  2.8186374  4.9560288
#> [114,] 15.026983  3.8082565  3.9791783  2.4444940  1.0143701  1.5244918
#> [115,] 15.022976  2.1894925  1.0566161  1.1431926  0.5076839  2.9464884
#> [116,] 36.417357  4.4499247  0.6531982  0.5677296  2.1456071  2.6481494
#> [117,] 28.202332  8.7023509  2.0883469  1.1137337  0.5191324  0.9925340
#> [118,]  9.508291  5.1313336  1.5843812 12.3896152  0.9920721  1.2603587
#> [119,]  5.622124  2.2788188  1.2276291  8.9992998  2.9845121  1.1629397
#> [120,] 28.445411  1.7757633  1.2246251  1.7144757  1.9233245  0.7959907
#> [121,] 76.235396  4.8977619  2.8492089  3.2077681  3.1184454  5.1054154
#> [122,] 11.086405  2.3201544  2.3109709  1.7510209  4.9072386  1.1702373
#> [123,] 23.230972  1.9020295  3.0646808  1.1766596  0.7597938  2.2675238
#> [124,] 13.140065  5.1415034  1.5024751  2.0183443  1.2503921  3.8934146
#> [125,] 10.007383  4.9341652  1.3092511  1.9647228  1.7790010  0.9522023
#> [126,] 15.826348  0.9851784  2.1845845  2.5074767  1.9249477  0.5766348
#> [127,] 19.290918  2.3529585  0.9437981  1.3342062  2.2016144  3.4120533
#> [128,]  8.139267  2.8830398  3.2345431  0.9914882  0.5992404  1.2601004
#> [129,] 23.837072  1.4431855  4.8785962  2.3320968  4.5520136  2.2210095
#> [130,]  7.093648  3.2215370  2.3716484  4.1998148  1.3593662  3.1063458
#> [131,]  9.666534  2.9779815  1.3642773  1.5932701  2.4498303  1.4228420
#> [132,] 16.750938  6.5992574  4.5941446  2.6259192  2.8942221  1.6534535
#> [133,] 10.872255  5.3859746  2.2375685  1.0585658  1.6682642  4.9156541
#> [134,]  5.727551  3.3621508  2.6191744  1.0452564  0.4461855  1.2373659
#> [135,] 25.986527  2.8512190  2.7156843  5.1515960  0.8673547  1.9766350
#> [136,] 13.186502  7.7062174  2.2317136  2.9352379  2.5379753  1.2503404
#> [137,] 14.988340  9.6687970  3.1589217  1.7487513  0.6576933  1.5401895
#> [138,] 24.558226  4.0301300  7.2796849  2.6808720  1.6979332  1.1989265
#> [139,] 22.659187  3.4294371  0.9160517  0.9723625  0.7313492  1.1695952
#> [140,] 10.471032  5.1586101  1.3097745  0.8060784  1.4195250  0.9569967
#> [141,] 16.355944  3.9097334  1.1008267  1.6852518  0.9140754  3.8476937
#> [142,] 20.839011  4.4574468  2.7397829  5.0855526  2.7229623  4.2245372
#> [143,] 15.401640  4.6181599  2.5279148  0.8585591  1.6620369  2.3026504
#> [144,] 56.771601  1.8933874  3.1952486  1.2796231  1.4029116  1.1389866
#> [145,] 14.064973  2.1353341  2.3664597  9.0410083  0.8368957  2.0389355
#> [146,] 13.462043  4.8473640  1.4706317  2.0366265  1.8395070  0.9917244
#> [147,]  8.953501  1.2092271  1.7062882  1.5276617  1.5529643  1.3518595
#> [148,] 16.522650  1.6474161  2.2516265  2.0248025  1.7810360  8.1520949
#> [149,]  9.119398  7.0815908  3.3159903  3.9414118  4.0267557  4.7936009
#> [150,]  8.954013  6.2682936  2.1171451  2.5249318  3.3952543  2.3205410
#> [151,]  6.987858  5.5060413  1.4604108  2.7777886  0.7681246  1.4155797
#> [152,] 17.198966 12.3940901  1.6462057  1.9734677  1.5623280  2.7567859
#> [153,]  5.582069  0.9462118  1.6393011  1.1628185  0.6709664  3.4122570
#> [154,]  9.443051  3.9457755  1.0058630  0.6640517  1.2698542  2.1054461
#> [155,] 26.551820  6.4106104  1.2887009  1.0438580  1.5537093  0.5762039
#> [156,] 27.032377  5.1330144  3.2891197  1.1178513  1.6042385  2.0370694
#> [157,]  5.202395  6.3980752  5.7916387  5.6778361  5.7662034  1.5818462
#> [158,]  7.097950  5.8461657  1.8082145  1.3567912  1.3236528  2.2029792
#> [159,] 14.306892  2.6202765  1.6856173  1.4508615  1.4282311  2.0103613
#> [160,] 29.391470  4.9788896  1.4484791  1.1913730  1.3123263  2.9320642
#> [161,] 16.137477  2.6102608  3.3639693  1.8117116  1.0589009  1.5859944
#> [162,] 12.574895  1.7306591  1.4787890  4.8036477  1.9590938  0.7338343
#> [163,]  6.249722  2.4847708  1.0705759  0.9997907  1.8202464  1.8422268
#> [164,] 12.861594  2.0868762  1.6657645  2.4958621  3.4135508  1.5664218
#> [165,] 10.456326  2.2869549  2.3033229  1.1836177  1.3278589  1.5269576
#> [166,] 15.944113  2.4773795  2.7102285  1.3682525  1.2577423  4.3817078
#> [167,] 16.470353  2.5470268  1.8251381  1.4955459  0.7385429  0.8881521
#> [168,]  7.702334  1.7792670  6.9366547  0.8831878  1.1372992  1.5262689
#> [169,] 11.527359  5.3709175  2.4480110  1.6188227  0.2994830  1.2692175
#> [170,]  4.444610  4.5511124  2.6617427  2.5830361  1.3267399  0.7707237
#> [171,] 10.892454  3.1157472  2.3512813  1.2465490  2.1878279  1.4906081
#> [172,] 12.707645  3.9318288  2.7129570  2.2776641  0.9134755  1.8141276
#> [173,] 17.027693  2.1513661  1.7392936  0.8472881  1.3694465  0.4513835
#> [174,] 21.831327  3.5190793  1.9057329  1.3155387  4.4034316  2.6208294
#> [175,]  9.060255  7.6604427  1.1502204  1.2653214  1.5296362  0.9273489
#> [176,] 15.158573  1.8777388  0.8621401  0.8438752  1.2140920  8.3767633
#> [177,] 15.347873  2.3842012  2.2356558  0.8961547  1.0884887  1.3234673
#> [178,] 18.544808  1.5008871  1.0040027  1.5097899  2.5344226  0.8083837
#> [179,]  9.178773  2.8364622  1.1097237  1.1085104  2.1154272  1.3813218
#> [180,] 24.795238  1.6971041  8.3610273  1.3628696  2.1602019  2.7949886
#> [181,] 13.158749  5.1327881  0.7952869  0.7143857  0.9943329  1.4175431
#> [182,] 17.512202  2.3356744  1.2694404  3.1722855  0.7574329  1.3241928
#> [183,] 17.112425  3.4591517  1.4379997  1.7276186  1.2866163  0.6540157
#> [184,]  8.330510  5.0320688  1.3381299  1.7679801  0.9700024  1.1829554
#> [185,]  9.144208  3.0827440  3.6439589  1.7949204  1.4953093  1.0139170
#> [186,] 13.477887  2.2587929  1.8283090  1.7268792  0.8901337  1.4469151
#> [187,] 19.460303  3.2423643  6.5336085  1.3234645  1.4450106  1.7479361
#> [188,]  8.414428  4.7141521  1.3266997  1.1526571  2.6799362  3.2935621
#> [189,]  9.995293  3.3795410  1.5888158  1.1840622  1.3394743  2.7153581
#> [190,] 25.660394  2.2842674  5.1183523  1.3639891  1.7739516  1.6801713
#> [191,] 13.862125  4.3477136  3.0077550  1.7669837  1.8617274  1.5587948
#> [192,] 31.546333  3.8538230  1.9199681  1.2217008  1.0914340  1.1458975
#> [193,] 28.897510  4.0199827  0.8709110  1.0437838  0.8122248  1.7098875
#> [194,] 36.529548  1.6008261  2.4873871  1.1883049  0.7710870 16.7005141
#> [195,] 15.367359  9.0001441  0.9694744  1.0505939  1.1937048  0.4505811
#> [196,] 30.066301  4.0908651  2.1536102  1.1515248  1.5190807  4.6481587
#> [197,] 12.024400  3.3106886  1.2330657  1.1243063  1.3899242  1.2104047
#> [198,]  6.429244  5.3763504  3.8937170  1.7534075  0.9076883  1.0317828
#> [199,] 13.161580  4.0402612  1.0440652  1.4242716  1.1504649  0.8879816
#> [200,] 14.492872  7.8751816  1.9684604  2.1628072  1.4582757  2.5284129
#> [201,] 13.322416  4.0976536  1.9663426  0.5472431  1.0738321  1.3257561
#> [202,]  8.395974  7.4867870  1.2844560  0.9933761  1.2429774  1.5126959
#> [203,] 15.048447  7.0299760  3.4273980  1.4359762  1.0134514  0.4837776
#> [204,] 12.277104 10.7307364  5.8690341  1.1646768  1.2309893  0.7442924
#> [205,]  7.152306  4.8651462  1.5003996  1.7448925  0.7562631  3.1038348
#> [206,]  5.053864  4.6477822  1.2907120  1.0738820  1.8149893  1.2408628
#> [207,]  6.475197 13.2392574  1.6153146  1.2734147  2.5048661  2.1208196
#> [208,] 13.449566  3.6222537  4.2263417  4.8819704  7.3112636  4.1635395
#> [209,]  9.272792  6.7804438  1.5275495  1.8099381  2.4670325  1.0584439
#> [210,]  8.182209  3.0217963  2.3071548  1.7053054  0.9869191  2.1151336
#> [211,] 17.328198 12.7627884  1.6726032  0.7262464  2.4734292  0.6720802
#> [212,]  8.124222  4.8592771  0.9236545  1.9583643  0.5773677  1.8826123
#> [213,]  9.107001 11.5825642  2.1278744  2.3084113  1.1915924  0.6875430
#> [214,] 29.522475 34.7053816  9.7924328  6.7885949 24.2125832  6.7992104
#> [215,] 14.498208  7.1453288  1.2840066  4.6967278  6.8763520  4.0642102
#> [216,] 11.409134 21.0001221  6.0078126  8.0541058  8.2428198  5.7910520
#> [217,] 11.929447  3.2758491  1.4709474  1.1977360  0.9946587  2.5314479
#> [218,] 10.093599  4.8852793  3.8416478  2.6321861  5.5421599  1.2296822
#> [219,] 12.177171  5.6728310  1.6261640  0.6260012  1.2267135  1.2163345
#> [220,] 36.242723  0.8394792  2.4703925  0.9399510  0.5679739  0.6974155
#> [221,] 19.599276  2.4748717  1.6950979  2.8489014  2.6993778  2.7191280
#> [222,] 14.072766  2.4821632  1.9590397  4.5317944  2.2226899 10.0579523
#> [223,]  7.218482  1.3512169  0.8983718  1.5605326  1.5047177  1.2515265
#> [224,]  6.901438  3.0167112  1.6172918  1.9279206  0.9655940  1.3752482
#> [225,] 13.450296  1.4973321  1.8871396  1.0922396  2.9042716  0.7225870
#> [226,] 13.605987  1.6600579  0.9690323  1.1764225  0.7553585  1.2125437
#> [227,] 12.314762  2.7693823  1.4384587  1.5692752  3.7375830  1.0539190
#> [228,]  7.590103  2.7758615  5.7421277  0.9039388  2.0934891  2.1267876
#> [229,]  7.332747  1.9922935  2.1230912  0.7193250  1.0423173  1.0463164
#> [230,] 16.548811  4.3909996  2.4719249  0.7643942  0.8481980  2.6746347
#> [231,] 10.951102  1.4572430  0.6856952  2.1272170  1.0782861  0.8038468
#> [232,]  6.009355  6.5392175  1.1151699  3.2198356  1.8814141  1.4575402
#> [233,]  7.023834  1.3791575  1.6799501  4.1156677  2.1661782  1.0414765
#> [234,] 17.969903  2.4167509  1.8535805  1.0181185  3.8428475  1.6640229
#> [235,] 15.523069  1.3468199  0.6263907  1.0099826  0.6339184  1.1754255
#> [236,] 22.298436  3.3680218  1.8909159  1.9421095  3.0777232  3.0560994
#> [237,] 14.270224  1.4058809  0.7822474  2.3130784  1.7656834  0.9488820
#> [238,] 48.948190  1.0361421  0.7732184  1.4340739  3.1667578  1.8026718
#> [239,]  8.429283  3.3856548  1.1534553  2.3225556  1.2005826  1.5160357
#> [240,]  8.733203  2.0604749  0.9503527  1.1706003  4.5867947  0.7953636
#> [241,] 36.336563  4.2526033  0.6023436  3.6019391  1.3088082  0.3921828
#> [242,]  8.466918  0.9211658  1.1583046  2.3984244  3.2901370  0.8054182
#> [243,] 37.173698  1.5396719  1.9241307  2.3425105  3.2306214  5.7049982
#> [244,]  7.585370  4.8525300 10.8058672  7.8748820  5.4383693  6.6779951
#> [245,] 15.744467  0.9002567  4.2274963  1.0491103  1.3103596  0.5299318
#> [246,] 12.623149  1.1132916  1.0689828  2.3257098  1.9737437  0.6721176
#> [247,] 18.311717  2.1782659  0.7721660  1.5772403  1.9744306  1.5814112
#> [248,]  7.891069  2.6011191  0.6658007  0.6892890  3.6527245  1.4650491
#> [249,] 10.577262  1.2902152  0.9098817  1.9661307  1.0150513  1.2670211
#> [250,]  6.983317  3.0702189  3.5156871  1.0299151  0.4402529  1.0005573
#> [251,] 26.847043  0.9574092  1.2583597  0.8854832  0.6638460  0.6245115
#> [252,] 27.530204  2.3281437  1.2083304  0.6657274  0.9038459  3.2412728
#> [253,]  8.680302  1.2167412  3.2245090  3.9746106  2.1967406  1.5698681
#> [254,] 12.325762  1.5057591  2.3985135  1.5320986  1.5989463  5.8746814
#> [255,]  7.431142  3.8487919  1.2132627  1.9557103  1.6953327  1.3087082
#> [256,] 32.822026  1.2941314  1.0700714  2.3425423  1.2879222  6.1725494
#> [257,]  7.606320  2.0698435  4.2975383  1.2220018  2.1898747  0.9117449
#> [258,] 24.234033  1.4114699  1.1585779  0.4270640  1.4305706  1.2478782
#> [259,] 29.331580  1.7355371  1.5375826  0.7515187  1.0967626  1.7184589
#> [260,] 10.053029  4.3730018  4.2294527  6.6071414  3.0608702  3.2210447
#> [261,]  8.499931 11.4912831  2.0502872  1.8056526  2.1259718  1.1043757
#> [262,] 10.851902  5.8254082  2.0207699  2.0691701  1.2586559  0.8345400
#> [263,] 19.293587  2.2821577  1.5096019  2.7874076  1.4200366  2.0018038
#> [264,]  9.056283  4.9243059  0.7212451  1.6591853  1.6415201  0.9090009
#> [265,]  7.875179  3.7632388  0.9283256  1.6216672  1.0076950  5.8903289
#> [266,]  6.004867  1.5036418  0.9193255  1.0344846  0.7411894  3.4982574
#> [267,] 11.772597  2.6835294  1.9678957  1.9410260  2.2472676  1.1903182
#> [268,] 12.411254  2.3982115  2.7247262  0.5924509  3.5274896  0.8322454
#> [269,] 10.253400  2.5516030  1.9411795  1.2811025  2.8693669  2.8620428
#> [270,] 22.712024  2.4265352  1.7440864  2.5764983  0.9107234  0.9334915
#> [271,] 21.135182  2.3809307  0.6004982  0.6767858  3.1371726  2.0650270
#> [272,] 14.693395 13.8462796  4.0287769  1.0427971  0.8267063  0.6586971
#> [273,]  5.424191  3.5771239  2.9105650  1.3238626  1.5692054  0.9237990
#> [274,] 14.252208  2.2741993  1.3999661  0.5941560  2.1500662  0.4108913
#> [275,] 10.306709  6.4876238  3.7829716  2.5137480  4.6311181  0.9473802
#> [276,]  9.946645  2.3786384  3.5434367  4.3948056  0.6479888  1.0142796
#> [277,] 12.286725  1.3930892  2.5907812  1.1727239  1.2623543  2.3612167
#> [278,]  3.720747  2.8853711  1.3217524  1.0005087  1.2010181  0.6149697
#> [279,]  8.913228 21.9828487  1.7304691  0.9871103  2.1002043  1.3628218
#> [280,] 10.548186  7.2743791  1.7370733  0.9015431  2.3289145  0.6849999
#> [281,]  7.623836  3.0466692  2.5495742  2.2550550  2.5282672  1.7776721
#> [282,]  8.275864  2.6464367  1.4583294  0.9060327  2.2066952  1.3592789
#> [283,] 13.368892  3.7775101  1.2986003  0.5809267  1.0989401  0.5257420
#> [284,]  6.477014  2.7918293  0.9220450  0.8313933  1.2138314  0.9374797
#> [285,] 15.446161  7.4485029  3.1795099  1.5711346  1.8238810  0.8590157
#> [286,] 15.247879  3.6483060  3.1203649  1.2635928  1.2447073  3.8511101
#> [287,] 21.846241  1.3292101  1.4366462  2.9530674  0.5180422  1.7566368
#> [288,] 15.263625  3.1586194  0.3844471  1.7345980  4.1788231  3.0251478
#> [289,] 30.160732  2.2159173  3.9637723  8.6617321 48.0673429 18.2933174
#> [290,] 16.345482 11.3300104  3.2050731  7.8549595 10.2742497 16.0569896
#> [291,] 29.573806  2.1835586  1.5255387  1.6270706  0.9030080  6.2307615
#> [292,] 11.886073  4.1617544  1.3019903  2.2131834  1.6511418  2.5602403
#> [293,] 18.861706  6.1015902  0.8171426  1.2164276  1.1929072  2.6289660
#> [294,] 10.832342  5.0308831  1.6756774  1.3182307  1.0304017  1.0221248
#> [295,] 10.890201  3.3110316  2.2671419  0.8091926  1.6052816  1.5381761
#> [296,] 21.990054  2.7539329  1.0721874  2.4861125  2.9240065  3.4003240
#> [297,]  9.903199  5.1986570  1.5555174  1.0440645  1.7700800  2.5790989
#> [298,]  7.985402  4.7302869  2.8322971  1.7516430  0.6607856  1.0226899
#> [299,]  8.026003  3.2195524  1.9699157  0.6188450  1.0629198  0.6716458
#> [300,] 17.193879  6.8178062  2.6172105 10.3843896  3.7318951  4.3599400
#> [301,] 10.064425  2.4657656  0.7548659  3.1282650  0.8159876  1.5169603
#> [302,] 16.849357  2.1453995  1.5547705  2.5986895  1.1745652  1.6589299
#> [303,]  9.784175  3.6518511  0.7373172  2.5786007  2.8043115  0.9579906
#> [304,] 12.405713  3.9598580  1.5079755  0.7640532  3.1600401  0.9363940
#>              [,7]       [,8]       [,9]      [,10]
#>   [1,] 10.7614371  8.9939363  5.5715136 20.0030021
#>   [2,]  8.1461727  5.7375683  3.4745512  8.9856328
#>   [3,]  0.7712511  2.0777986  2.9934239  2.6283684
#>   [4,]  3.3096217  7.2910216  2.5294775  3.1023263
#>   [5,]  1.2481407  4.8925547 43.4973451  1.8513128
#>   [6,]  1.8039905  1.4043723  3.3604129  8.6674786
#>   [7,]  6.4987639  5.2055904  4.4582174  4.9343696
#>   [8,]  8.8070282 11.0432951  4.8074843  7.2239644
#>   [9,]  1.5018057  5.6896500  5.8236905  3.4028458
#>  [10,]  4.0669071  3.6529293  8.7692746 17.3740218
#>  [11,]  2.1424265  6.9642279  2.6908900  4.1533739
#>  [12,] 51.3556145 65.8209279 14.6325095 10.2552800
#>  [13,] 13.8440552  8.3411704 10.6393257 11.1180379
#>  [14,] 10.9326752  5.4645705 15.6167778 34.1490148
#>  [15,]  2.2492912  3.8556025  2.8219565 13.9760424
#>  [16,]  7.4849932  2.5201891  5.0799973  3.8739617
#>  [17,]  3.4415084  3.7560775  2.4949378  2.9031575
#>  [18,]  0.9432921  1.5448372  2.3707457  1.6574245
#>  [19,]  3.9244674  1.8669384  1.8786438  1.4338931
#>  [20,]  2.1529520  0.6358290  0.8267468  0.9681378
#>  [21,]  1.3807075  3.1080291  2.1012294  4.5791875
#>  [22,]  0.9118885  2.3221927  1.8786511  3.8554675
#>  [23,]  1.5650968  2.0528996  5.1189814  4.0239279
#>  [24,]  4.4939153 11.6443018  7.4963631  5.1077168
#>  [25,]  8.1252350  3.6596924  2.6021269  2.3133815
#>  [26,]  8.2265019  2.2996728  2.6615370  2.3177752
#>  [27,]  2.0255387  1.9561860  2.8081605  1.2259524
#>  [28,]  3.3858433  4.7953717  3.9992706  4.2636517
#>  [29,]  1.3719084  0.9729485  1.4216114  4.1785246
#>  [30,]  1.1382310  1.1824106  0.8725950  0.7143593
#>  [31,]  4.3733422  7.2721798  3.2837714  3.2082153
#>  [32,]  5.0469595  3.2506626  3.0629130  2.6806590
#>  [33,]  1.5372420  0.9255536  1.5268926  3.7970640
#>  [34,]  2.3349267  8.2132305  1.6820158  1.6728618
#>  [35,]  1.1655308  1.6136766  3.0557590  3.2924615
#>  [36,]  1.2875169  2.2103417  1.6095239  1.6298280
#>  [37,]  0.8830966  3.4745614  6.0317888  3.6057346
#>  [38,]  1.9580197  1.2919013  2.8655280  1.8119830
#>  [39,]  2.0481649  2.0916437  3.9147381  2.0665991
#>  [40,]  8.6520437  8.9842962  3.6720633  9.3233829
#>  [41,]  3.3336764  1.6155222  2.7377382  3.3036278
#>  [42,]  1.1412252  1.2294707  1.5726563  1.8993293
#>  [43,]  3.0234926  4.5519383  1.2939027  1.4481579
#>  [44,]  2.0665887  1.7131413  2.9005222  7.2893996
#>  [45,]  1.2388742  0.7207452  1.9715067  1.7975549
#>  [46,]  1.4469836  1.3788079  1.0684352  1.8457408
#>  [47,]  2.9506103  1.7503720  1.0496392  1.9496770
#>  [48,]  1.6127690 12.3335585  5.5761910  6.1558571
#>  [49,]  0.7252721  2.2450022  1.1550099  1.5983740
#>  [50,]  1.6744963  3.8880368  1.7626654  1.0394282
#>  [51,]  1.0192098  2.7472753  5.9763651 17.2207416
#>  [52,]  0.6779088  0.6134433  0.7908814  0.5763330
#>  [53,]  1.1048994  0.8633919  1.6264178  1.4597560
#>  [54,]  3.5225582  1.3099113  3.9393060  0.7654275
#>  [55,]  2.3383010  0.8645727  2.3802259  2.3174377
#>  [56,]  2.0347501  5.8794984  2.2728900  5.2371559
#>  [57,]  0.6926942  3.6665982  0.5116190  0.7490891
#>  [58,]  1.1493908  2.6125636  1.6716171  1.1031519
#>  [59,]  1.2195484  2.8631806  0.7534367  2.4088244
#>  [60,]  1.2673523  1.2355686  0.5725994  1.2281808
#>  [61,]  1.4853356  1.5611335  1.0498066  1.5532227
#>  [62,]  2.9369322  2.8044860  2.3628595  2.3882582
#>  [63,]  1.1393596  0.7135491  2.0450337  3.6904429
#>  [64,]  4.8909587  3.1917259  1.6952580  2.0131954
#>  [65,]  5.0288833  4.3466179  3.7587816  4.0091350
#>  [66,]  0.7576281  0.8722444  1.0215546  0.6138994
#>  [67,]  2.3673798  1.1059134  1.4358656  1.8545648
#>  [68,]  2.7753303  0.7582407  2.7577803  3.6866997
#>  [69,]  0.5089837  1.1444588  0.6577681  0.6109990
#>  [70,]  4.9917744  2.2726729  5.4022330  2.0188878
#>  [71,]  2.2816504  6.7296175  2.4066058  5.0696754
#>  [72,]  5.1436365  6.6381472  7.2625363  9.1513071
#>  [73,]  1.2435749  4.5846317  1.9852105  3.0215427
#>  [74,]  2.8259075  1.8065919  0.5734260  5.2530198
#>  [75,]  1.7921325  5.1956358  0.7994148  3.9142770
#>  [76,]  2.0591909  3.5675305  2.3730455  2.5863343
#>  [77,]  0.6355779  0.6113820  1.0812820  1.0548647
#>  [78,]  0.6249762  1.4633296  1.8471667  0.9218680
#>  [79,]  0.9627026  1.7397864  0.8285347  2.0498168
#>  [80,]  1.6652754  1.4690551  1.7978303  1.9467209
#>  [81,]  1.3544047  1.0737566  0.7602015  1.1328195
#>  [82,]  2.7933272  2.7872150  9.0669937  5.5410862
#>  [83,]  1.2317594  1.3775573  0.9020671  1.9537612
#>  [84,]  4.1116113  4.5818173  3.0409590  8.0576617
#>  [85,]  0.7994735  0.5012788  1.1978811  1.2953504
#>  [86,]  0.9039561  1.0351871  3.6832628  1.6452419
#>  [87,]  1.0542474  2.7599198  0.7401649  0.9609943
#>  [88,]  0.9305125  3.0866537  1.4780068  4.2238850
#>  [89,]  2.6208849  3.2112123  2.3589210  2.2354502
#>  [90,]  1.3876576  1.2367108  4.2916400 10.5136790
#>  [91,]  2.0949632  1.0499030  2.0958824  2.3333345
#>  [92,]  2.4608016  2.6234065  1.9585171  4.4484806
#>  [93,]  1.6503809  1.2255194  1.4378374  0.5565463
#>  [94,]  1.2387736  0.9671634  1.6193749  1.3269700
#>  [95,]  6.0886205  2.0277994  2.1137631  1.6513782
#>  [96,] 16.8189606  4.0454045  7.7568474  6.6561386
#>  [97,]  0.9331194  2.6294247  2.0823958  1.6592620
#>  [98,]  2.4877429  7.5180235  2.9288651  5.6396037
#>  [99,]  4.4872863  0.4777938  2.6574817  1.9862198
#> [100,]  1.8132994 13.8998595  1.5621292  2.8634129
#> [101,]  1.2494833  1.2478530  0.8885116  2.0634909
#> [102,]  5.0952040  1.8856420  2.6297639  2.2062446
#> [103,]  0.9524055  0.5284466  1.3319687  1.0634201
#> [104,]  1.7504754  6.1971347  3.8102695  2.0090698
#> [105,]  0.9839313  0.8843337  0.8927060  1.6032079
#> [106,]  2.6193896  2.3831060  1.8529000  2.3224853
#> [107,]  0.7906240  1.0000138  0.9610133  0.8257486
#> [108,]  1.0815729  7.5565337  5.8392652  2.0357638
#> [109,] 44.4719515 43.5137328 51.9302727 35.8358238
#> [110,] 14.5124516 10.4873864 14.1931533 21.2712078
#> [111,]  0.8151183  0.7802328  2.0241923  0.7191949
#> [112,]  2.1308123  2.0939490  1.7092439  1.2015209
#> [113,]  2.8082201  2.8637907  1.9044439  5.7002931
#> [114,]  2.5385115  0.5206876  0.5109085  1.9636143
#> [115,]  1.1274437  1.2263137  0.7354441  1.6078090
#> [116,]  1.6363789  1.2998933  0.7371883  1.2458622
#> [117,]  1.9222532  2.7688085  2.7238640  0.8262635
#> [118,]  2.7512091  2.0910550  0.9454208  0.6198472
#> [119,]  2.3042403  1.9808739  1.9790316  1.5429716
#> [120,]  2.2684284  1.2720897  1.6882272  1.0401091
#> [121,]  7.6706889  3.0207596 16.7720505  3.3353134
#> [122,]  0.5415469  0.9125876  1.3052450  0.6013922
#> [123,]  0.6721977  0.6864179  1.5724519  1.1774662
#> [124,]  0.7036695  1.4240288  1.8573998  1.7848314
#> [125,]  0.9287225  0.6659119  1.1751182  6.8668255
#> [126,]  0.7538788  1.4432625  0.8196084  1.5048233
#> [127,]  2.6727027  0.9456166  1.2520652  0.8932277
#> [128,]  1.3378649  1.5511576  1.2663523  7.7702757
#> [129,]  6.1915512  5.9590451  7.5498942 11.8850949
#> [130,]  2.3616876  1.9119171  1.0886658  1.4008300
#> [131,]  0.7965389  2.0635488  1.2439836  2.7662283
#> [132,]  5.1118383  1.5882198  7.5060536  1.8619354
#> [133,]  2.6756380  4.8691482  1.6502250  3.5823631
#> [134,]  1.0172435  0.7284773  8.2395477  0.6007681
#> [135,]  0.5594193  3.6669946  1.1234742  6.4716126
#> [136,]  5.3040394  3.8508755  1.2892359  2.4027596
#> [137,]  0.8845326  1.5553657  3.7296201  1.3702615
#> [138,]  2.2029316  1.8051922  1.3352195  2.6673387
#> [139,]  3.6783554  1.7131847  0.8152343  2.2759591
#> [140,]  1.4162925  1.6300727  3.2708027  0.9304335
#> [141,]  3.1564745  4.8824049  2.1245940  3.2312607
#> [142,]  6.7701857  4.6912382  4.7921042  5.3288156
#> [143,]  2.2593759  2.8641284  1.1693748  5.1900281
#> [144,]  0.9919207  3.9362736  0.7712011  3.6664537
#> [145,]  0.9772142  1.2696790  2.6781820  0.6966325
#> [146,]  0.9927855  1.1424334  2.1191215  1.3057791
#> [147,]  5.0285378  2.3066250  1.5538609  3.7474064
#> [148,]  2.6719377 11.6654951  7.4202472  3.4126392
#> [149,]  3.7583597  3.2576489  6.9269255  4.6717851
#> [150,]  6.5369557  4.4819528  1.2191254  1.8708947
#> [151,]  1.0806974  0.4270493  0.5714494  1.1918178
#> [152,]  1.7761486  1.0056209  0.9138132  0.7971305
#> [153,]  2.4586712  1.2203219  0.9535609  2.0039512
#> [154,]  1.4214592  1.4012733  1.7441511  2.0875426
#> [155,]  4.3782513  2.4994290  4.5228299  1.2249124
#> [156,]  1.2083298  0.7347067  0.9482247  0.7397719
#> [157,]  6.1839378  3.3004280  3.6796281  3.3726393
#> [158,]  2.1331788  1.4257204  2.0979231  1.4136619
#> [159,]  1.2794604  1.0452975  2.3299637  4.9720175
#> [160,]  3.4982164  1.9880322  4.8665221  3.6062083
#> [161,]  1.6681519  2.0249412  1.7101956  1.5213486
#> [162,]  1.9087439  3.1648658  4.9948313  0.4301734
#> [163,]  1.3626882  6.4016341  1.6389132  0.7942290
#> [164,]  2.0860725  2.1267830  1.6673683  2.0753349
#> [165,]  1.4020755  3.0551326  0.6224801  0.7816622
#> [166,]  0.5483812  0.8771891  0.8006467  1.5530441
#> [167,]  0.9690431  0.8684023  1.1870167  2.4196273
#> [168,]  0.6570414  2.0354656  0.4823788  3.9497351
#> [169,]  1.1646648  0.5271178  1.6791691  0.9236595
#> [170,]  1.3564455  3.0147360  1.3824722  0.9904499
#> [171,]  0.8226585  4.0828434  1.6880295  1.5361816
#> [172,]  0.8196703  1.1146375  1.5768959  2.2127317
#> [173,]  0.4233613  1.0113429  2.9534876  1.4306957
#> [174,]  1.3356867  0.4461114  0.6566575  2.2001324
#> [175,]  1.1089137  1.8297329  1.3768190  0.6269717
#> [176,]  0.7326140  0.5564395  5.3896632  1.6375327
#> [177,]  0.7029842  0.7822138  0.5433300  0.5486584
#> [178,]  0.7234277  1.5272924  1.1489973  2.6290680
#> [179,]  0.9882882  1.3146290  0.4354516  1.5929571
#> [180,]  2.2112681  4.5221338  1.3252490  1.9733402
#> [181,]  1.5604202  1.0095865  4.6057001  0.7024463
#> [182,]  0.9641705  8.6702078  0.8436562  0.7543040
#> [183,]  2.4898875  1.0065567  1.8665904  1.3106545
#> [184,]  3.4503852  1.9670997  3.0547755  3.9508786
#> [185,]  2.4672946  2.1591284  2.7192587  1.1811587
#> [186,]  1.9937568  9.2833204  1.2338676  2.4957504
#> [187,]  0.8093955  1.3232943  1.5407166  1.2452468
#> [188,]  0.3914185  0.6307864  0.9181305  0.8714468
#> [189,]  1.6866594  1.9853831  1.7009567  1.1174046
#> [190,]  1.3209390  0.9078846  1.2179704  1.1528807
#> [191,]  2.7001448  1.4661707  1.5223619  2.2028865
#> [192,]  0.6218195  1.3684604  2.7594365  0.6443708
#> [193,]  1.1639425  2.0553274  1.5836372  1.2968467
#> [194,]  0.5662157  0.6797230  3.2029114  0.5433639
#> [195,]  0.9625628  0.4898009  0.9921285  1.0447274
#> [196,]  1.9477131  2.2499200  0.9851515  3.6244319
#> [197,]  0.8618075  2.3759449  1.9054266  3.1057318
#> [198,]  0.9955051  0.8728301  1.2653556  2.0568546
#> [199,]  1.1038941  0.5486736  1.1437606  0.6299859
#> [200,]  7.8646895  1.6682091  2.0436647  2.8712966
#> [201,]  1.7240894  1.7129641  0.6317295  1.9177047
#> [202,]  1.0213874  0.4094282  1.9111443  1.8814409
#> [203,]  1.0833924  0.9093266  3.6370407  1.1237812
#> [204,]  1.2255763  1.2191476  0.3279612  0.6364686
#> [205,]  2.5235305  0.9749337  1.7014235  0.8512241
#> [206,]  1.5423427  1.7287583  0.3823560  1.5303491
#> [207,]  2.5726800  3.3707268  2.4459624  0.6146884
#> [208,]  4.6048276  3.2005713  4.9976411  9.7651499
#> [209,]  1.8803517  2.1010097  1.3265562  1.8653426
#> [210,]  2.3218829  1.8476215  0.6786922  1.3125773
#> [211,]  9.9068787  0.6264194  1.3016342  1.3559228
#> [212,]  3.8206090  1.4193249  1.1865187  1.6320784
#> [213,]  1.5733647  1.1330698  0.7261113  1.6927331
#> [214,]  7.6710201  5.6266620  9.5933421  8.8684396
#> [215,]  2.7578750  1.9481735  5.0346538  4.1385355
#> [216,] 11.7347862  7.4669642  7.1896918  8.9893665
#> [217,]  1.4973462  1.8763614  1.9676709  1.1189613
#> [218,]  0.6952972  0.6092815  2.4647004  2.1956619
#> [219,]  2.6472470  1.0064492  2.8145894  2.0853382
#> [220,]  1.7581114  1.0852186  0.9636157  2.1899301
#> [221,]  4.4955347  2.5225471  3.0112811  2.6660643
#> [222,]  6.4309663  9.0667038  3.2417556  9.8829017
#> [223,]  0.7489588  2.2368910  1.2002768  4.5964480
#> [224,]  1.2678490  1.7601222  2.4007060  0.9648959
#> [225,]  0.7208732  0.9458971  0.5035813  1.2790674
#> [226,]  1.1058459  1.5191750  0.7143829  1.1411383
#> [227,]  0.6153713  0.5149170  1.1471271  0.9360782
#> [228,]  2.9719236  1.9035711  1.2148225  3.7171461
#> [229,]  1.8307672  0.9028899  0.7904432  0.6410858
#> [230,]  1.8790781  0.6233039  0.7258483  1.6976462
#> [231,]  1.4397805  1.1816418  0.3902730  2.1879660
#> [232,]  2.5832490  1.6290372  1.7231856  1.4145253
#> [233,]  1.6838031  3.4201161  0.4501800  1.4660740
#> [234,]  1.0722238  1.0956394  0.7119144  1.9169915
#> [235,]  1.1847512  0.4675489  0.9300679  1.2266010
#> [236,]  2.0651332  6.8352834  1.2264260  0.8658375
#> [237,]  0.4392345  1.1033737  1.2316907  1.9213784
#> [238,]  1.1994743  0.9490584  0.8801157  0.4279886
#> [239,]  1.3037045  2.3089545  0.8503092  0.7855089
#> [240,]  0.8642457  1.7378206  2.6219277  3.0175832
#> [241,]  1.4500840  0.7413097  3.8651366  1.9828433
#> [242,]  2.5582222  4.1141404  0.9702550  1.0515823
#> [243,]  1.9373872  3.9111539  2.3877677  6.0161499
#> [244,]  9.2002815 10.9583277 31.0233257 13.9122646
#> [245,]  1.4873891  1.1128963  1.2074483  3.4172379
#> [246,]  0.6926324  2.1319513  1.0487673  1.6545649
#> [247,]  2.1816978  1.0927084  1.4096953  1.3557524
#> [248,]  3.7867648  1.6049100  1.0519340  1.1770351
#> [249,]  3.2261059  1.5170828  1.4247810  1.1111689
#> [250,]  0.6012602  1.8187997  1.5696916  3.0327579
#> [251,]  2.3313535  1.4521773  2.0680394  1.2777244
#> [252,]  1.6136429  1.5836155  0.3887004  1.3259504
#> [253,]  1.5575590  0.6742585  0.7546753  0.7675440
#> [254,]  1.2270373  0.9455159  1.3663192  1.8733211
#> [255,]  0.5796879  1.1062097  0.8293386  1.2578407
#> [256,]  1.9148026  0.6622532  1.0282703  0.9925181
#> [257,]  0.9816550  2.3494544  2.0336374  0.7789137
#> [258,]  0.8037492  5.4464841  0.5382334  1.7563883
#> [259,]  2.5351987  2.7752034  2.1688256  6.4397828
#> [260,]  3.7232400  2.3418231  2.4673103  7.8935936
#> [261,]  0.9467794  2.2097617  0.6897114  0.6629395
#> [262,]  0.6952175  0.8070600  1.1150198  1.1711107
#> [263,]  2.4933602  2.2766230  1.1074528  3.5047933
#> [264,]  2.1265919  1.1346899  1.8937868  1.3620431
#> [265,]  3.3086705  1.3862283  1.2171384  2.7224866
#> [266,]  2.1033369  1.3607676  1.1606884  1.1837208
#> [267,]  4.4960334  1.1794940  2.0706980  0.8150864
#> [268,]  6.6651219  3.0326533  0.9758016  0.9496630
#> [269,]  1.2001184  1.1087005  1.7523544  1.1178338
#> [270,]  2.8570649  0.4572818  1.0239030  1.2885595
#> [271,]  2.2770249  1.4073327  0.9844040  1.8607011
#> [272,]  1.1222779  1.4403437  3.9070406  1.6503427
#> [273,]  0.7280107  4.0627837  1.4991005  1.3631487
#> [274,]  1.8455701  0.7102224  1.1207290  0.6549919
#> [275,]  0.5374291  1.0702681  0.3409543  0.7952788
#> [276,]  0.4786432  2.6726589  0.9134446  0.8026083
#> [277,]  1.0925963  1.4929147  1.0056757  1.1454984
#> [278,]  2.3749924  0.8477968  0.7520357  0.8460830
#> [279,]  1.9533273  1.8891193  0.8133424  0.6902181
#> [280,]  1.1763208  1.0558617  4.9261641  0.9481777
#> [281,]  1.2934793  0.5494949  0.8020270  0.8286473
#> [282,]  1.5995579  0.6883606  0.8283860  2.4930719
#> [283,]  1.0042640  2.2612979  0.5511196  0.9009080
#> [284,]  2.3825968  2.2120819  1.1681322  3.9559934
#> [285,]  0.5604758  3.3370235  1.5760225  1.3638098
#> [286,]  0.9232039  0.6846680  1.6533076  0.7893406
#> [287,]  2.3917405  1.7481207  0.5599122  5.3884050
#> [288,]  5.5273051  3.1906303  3.6878100  3.0357011
#> [289,] 25.8932451 32.6547051 30.0828605 15.2504976
#> [290,] 34.8637186 23.7309346 17.7896075 32.3916614
#> [291,]  1.6789055  1.1413662  4.8245204  1.8557634
#> [292,]  2.6438494  3.9250813  5.7776748  7.0920691
#> [293,]  1.5404696  1.8296283  1.0143432  6.9276385
#> [294,]  0.6951845  1.2022306  0.6893954  1.8509499
#> [295,]  1.9780272  1.0485870  2.2346129  2.3641917
#> [296,]  4.8737687  9.1897774  2.6976152  5.2617398
#> [297,]  1.0362246  1.7387716  1.0926785  1.0869112
#> [298,]  0.4635223  0.7534916  1.7719465  1.3496958
#> [299,]  1.4811652  1.0512451  4.2386016  2.3577548
#> [300,]  3.6674475  3.6809222  5.8701140  7.8895360
#> [301,]  0.9847931  2.0643959  0.3183735  0.5517746
#> [302,]  1.1488204  0.8757950  3.6335961  1.6259171
#> [303,]  0.6793634  1.0922478  2.0888928  1.6752048
#> [304,]  2.5232768  2.1597820  1.9287903  1.3848473
#> 
#> $df
#>           [,1]
#>  [1,] 6.000000
#>  [2,] 5.542428
#>  [3,] 5.542428
#>  [4,] 4.232387
#>  [5,] 4.232387
#>  [6,] 4.232641
#>  [7,] 4.232641
#>  [8,] 4.232641
#>  [9,] 3.863035
#> [10,] 3.863035
#> 

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$get_last_draw`
## ------------------------------------------------

data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
set.seed(123)
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$is_normalised`
## ------------------------------------------------

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.

# estimate the model
set.seed(123)
posterior      = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# check normalisation status beforehand
posterior$is_normalised()
#> [1] TRUE

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) %*% BB         # set negative diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()
#> [1] TRUE

## ------------------------------------------------
## Method `specify_posterior_bsvar_t$set_normalised`
## ------------------------------------------------

# This is an internal function that is run while executing normalise_posterior()
# Observe its working by analysing the workflow:

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_t$new(us_fiscal_lsuw, p = 4)
#> The identification is set to the default option of lower-triangular structural matrix.
set.seed(123)

# estimate the model
posterior      = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# check normalisation status beforehand
posterior$is_normalised()
#> [1] TRUE

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) %*% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)                      # draws in posterior are normalised

# check normalisation status afterwards
posterior$is_normalised()
#> [1] TRUE