R6 Class Representing StartingValuesBSVARSV

The class StartingValuesBSVARSV presents starting values for the bsvar model with Stochastic Volatility heteroskedasticity.

Super class

bsvars::StartingValuesBSVAR -> StartingValuesBSVARSV

Public fields

A

an NxK matrix of starting values for the parameter \(A\).

B

an NxN matrix of starting values for the parameter \(B\).

hyper

a (2*N+1)x2 matrix of starting values for the shrinkage hyper-parameters of the hierarchical prior distribution.

h

an NxT matrix with the starting values of the log-volatility processes.

rho

an N-vector with values of SV autoregressive parameters.

omega

an N-vector with values of SV process conditional standard deviations.

sigma2v

an N-vector with values of SV process conditional variances.

S

an NxT integer matrix with the auxiliary mixture component indicators.

sigma2_omega

an N-vector with variances of the zero-mean normal prior for \(\omega_n\).

s_

a positive scalar with the scale of the gamma prior of the hierarchical prior for \(\sigma^2_{\omega}\).

Methods

Public methods

specify_starting_values_bsvar_sv$new()

specify_starting_values_bsvar_sv$get_starting_values()

specify_starting_values_bsvar_sv$set_starting_values()

specify_starting_values_bsvar_sv$clone()

Method new()

Create new starting values StartingValuesBSVARSV.

Usage

specify_starting_values_bsvar_sv$new(N, p, T, d = 0)

Arguments

N

a positive integer - the number of dependent variables in the model.

p

a positive integer - the autoregressive lag order of the SVAR model.

T

a positive integer - the the time series dimension of the dependent variable matrix \(Y\).

d

a positive integer - the number of exogenous variables in the model.

Returns

Starting values StartingValuesBSVARSV.

Method get_starting_values()

Returns the elements of the starting values StartingValuesBSVARSV as a list.

Usage

specify_starting_values_bsvar_sv$get_starting_values()

Examples

# starting values for a bsvar model with 1 lag for a 3-variable system
sv = specify_starting_values_bsvar_sv$new(N = 3, p = 1, T = 100)
sv$get_starting_values()   # show starting values as list

Method set_starting_values()

Returns the elements of the starting values StartingValuesBSVAR_SV as a list.

Usage

specify_starting_values_bsvar_sv$set_starting_values(last_draw)

Arguments

last_draw

a list containing the last draw of the current MCMC run.

Returns

An object of class StartingValuesBSVAR including the last draw of the current MCMC as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Examples

# starting values for a bsvar model with 1 lag for a 3-variable system
sv = specify_starting_values_bsvar_sv$new(N = 3, p = 1, T = 100)

# Modify the starting values by:
sv_list = sv$get_starting_values()   # getting them as list
sv_list$A <- matrix(rnorm(12), 3, 4) # modifying the entry
sv$set_starting_values(sv_list)      # providing to the class object

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_starting_values_bsvar_sv$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# starting values for a bsvar model for a 3-variable system
sv = specify_starting_values_bsvar_sv$new(N = 3, p = 1, T = 100)

## ------------------------------------------------
## Method `specify_starting_values_bsvar_sv$get_starting_values`
## ------------------------------------------------

# starting values for a bsvar model with 1 lag for a 3-variable system
sv = specify_starting_values_bsvar_sv$new(N = 3, p = 1, T = 100)
sv$get_starting_values()   # show starting values as list
#> $B
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> [2,]    0    1    0
#> [3,]    0    0    1
#> 
#> $A
#>         [,1]     [,2]      [,3] [,4]
#> [1,] 0.93944 0.000000 0.0000000    0
#> [2,] 0.00000 0.772823 0.0000000    0
#> [3,] 0.00000 0.000000 0.2471741    0
#> 
#> $hyper
#>      [,1] [,2]
#> [1,]   10   10
#> [2,]   10   10
#> [3,]   10   10
#> [4,]   10   10
#> [5,]   10   10
#> [6,]   10   10
#> [7,]   10   10
#> 
#> $h
#>              [,1]         [,2]         [,3]          [,4]        [,5]
#> [1,] -0.004382321  0.003104321  0.006866561 -0.0080863488  0.00845498
#> [2,] -0.014348153 -0.003901116 -0.001052677 -0.0080539085 -0.01340276
#> [3,]  0.015145357 -0.007034511  0.008557220  0.0002402033 -0.00363382
#>              [,6]          [,7]        [,8]         [,9]       [,10]
#> [1,]  0.002957100 -7.654917e-05 0.012913257 -0.009523599 0.011423154
#> [2,] -0.005594185 -7.639284e-03 0.009492643 -0.012450675 0.009616121
#> [3,] -0.001319023  2.234584e-03 0.009804601  0.004191351 0.004852678
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,]  0.012666489 -0.013398076 -0.013162488  0.005301769 -0.009323170
#> [2,]  0.005767275  0.006682450  0.008019722 -0.017808959 -0.001071962
#> [3,] -0.017832070 -0.003279624  0.018779607  0.003615080  0.014806913
#>            [,16]         [,17]         [,18]       [,19]       [,20]
#> [1,] 0.005065413 -0.0047462754 -0.0148906942 -0.01789116 0.004832875
#> [2,] 0.003034568 -0.0031777777  0.0213497403  0.00284658 0.011159474
#> [3,] 0.003797679 -0.0002858477  0.0007376726  0.01194728 0.001863944
#>             [,21]        [,22]         [,23]         [,24]        [,25]
#> [1,] -0.002245534 -0.006841452 -0.0006695077  0.0094595772  0.003464587
#> [2,]  0.018172008  0.005034092  0.0074779796  0.0006473996  0.002183629
#> [3,] -0.006871098  0.002970526  0.0056264755 -0.0174377854 -0.008900761
#>             [,26]        [,27]        [,28]       [,29]        [,30]
#> [1,] -0.004920652  0.004007803 -0.006683319 -0.00311529 0.0004487847
#> [2,]  0.009891255 -0.005918885 -0.008213835  0.00621109 0.0043546292
#> [3,] -0.017237814 -0.010280466  0.001760306 -0.01002211 0.0049672971
#>              [,31]       [,32]        [,33]        [,34]       [,35]
#> [1,]  0.0092777069 0.007405368 -0.003448997  0.002380457 0.013775828
#> [2,]  0.0008691243 0.002911925 -0.003634461 -0.010208250 0.013046012
#> [3,] -0.0072433093 0.002597670 -0.012814133 -0.001485088 0.003697177
#>              [,36]       [,37]        [,38]        [,39]       [,40]
#> [1,]  0.0100017546 -0.01122576 -0.007308356 0.0057203498 -0.01612397
#> [2,] -0.0001889773  0.01268148 -0.012161950 0.0022125582 -0.01480671
#> [3,] -0.0035368195  0.01196743 -0.012605575 0.0009295309  0.01247285
#>             [,41]       [,42]        [,43]        [,44]        [,45]
#> [1,] -0.005075658 -0.02463380 -0.021080513 0.0065251642 -0.004831030
#> [2,] -0.006691202  0.01308132 -0.001346022 0.0003750332  0.016975072
#> [3,] -0.002516170 -0.00728578 -0.005119411 0.0055516712 -0.007530216
#>              [,46]        [,47]         [,48]        [,49]        [,50]
#> [1,]  0.0009636446  0.009180714  0.0008027508  0.002600932 -0.001665706
#> [2,] -0.0031282666 -0.003309613  0.0025008911 -0.005338025 -0.014627132
#> [3,]  0.0122666849 -0.007921490 -0.0046261752 -0.006061367 -0.015187939
#>             [,51]       [,52]       [,53]        [,54]       [,55]        [,56]
#> [1,]  0.014039416 0.018163358 0.005486400  0.006447610 -0.01302362 0.0095092274
#> [2,]  0.006802832 0.007768426 0.005286509 -0.015887835 -0.01173406 0.0058698743
#> [3,] -0.009957039 0.001572557 0.005074630 -0.009738482 -0.01390215 0.0009762083
#>            [,57]         [,58]        [,59]        [,60]        [,61]
#> [1,] 0.020611611  0.0007999573 -0.006167607  0.002832582 -0.007973131
#> [2,] 0.010567444 -0.0002624856  0.003059701 -0.001883788 -0.009632329
#> [3,] 0.006346514  0.0047698736  0.007631152 -0.011301943 -0.005935892
#>            [,62]        [,63]        [,64]       [,65]         [,66]
#> [1,] 0.009278003 -0.019490731 -0.008835912 0.001255035 -0.0008763099
#> [2,] 0.004814176 -0.024066254  0.005549767 0.014565764 -0.0062084430
#> [3,] 0.012275474  0.004722107 -0.004143575 0.006645046 -0.0165615578
#>             [,67]       [,68]        [,69]        [,70]        [,71]
#> [1,]  0.005806716  0.02263684  0.009455320 -0.001142631 0.0098854912
#> [2,] -0.006323757 -0.02224525 -0.007528301  0.006727978 0.0009075175
#> [3,] -0.014446326 -0.01217662 -0.006046939  0.003885027 0.0001318380
#>            [,72]        [,73]        [,74]         [,75]         [,76]
#> [1,] 0.006309348 -0.015927340 -0.001165701 -2.318225e-02 -0.0016027526
#> [2,] 0.004988727 -0.007494646 -0.007516602  1.457966e-02  0.0029736371
#> [3,] 0.003050148  0.016321058 -0.002990816  4.965169e-05  0.0004983806
#>             [,77]         [,78]        [,79]        [,80]        [,81]
#> [1,] -0.007134890 -7.957282e-03 -0.016025031 -0.001635621  0.032177415
#> [2,] -0.007996654  4.764925e-05 -0.005292308  0.008057072 -0.003019357
#> [3,] -0.008699400  2.786890e-03  0.008494457 -0.005695601  0.020643467
#>             [,82]       [,83]         [,84]        [,85]         [,86]
#> [1,]  0.002888835 0.005505278 -0.0052531026  0.011807759 -0.0003916467
#> [2,] -0.006236668 0.015895236 -0.0047752439 -0.002741668 -0.0072308519
#> [3,]  0.004981065 0.003239365  0.0009702138  0.022313979  0.0188843014
#>             [,87]        [,88]        [,89]         [,90]        [,91]
#> [1,] -0.009290911 -0.005694537  0.002709409 -0.0080026047 -0.004680470
#> [2,]  0.013489812  0.003261412 -0.007717680 -0.0006153368  0.001219792
#> [3,]  0.008746070  0.009194541 -0.001165375  0.0044203901  0.001083984
#>              [,92]         [,93]       [,94]        [,95]        [,96]
#> [1,] -0.0123889128  0.0001799319 0.014095948 0.0012709407  0.009927484
#> [2,]  0.0071371537 -0.0022434571 0.006155422 0.0001001096  0.003870986
#> [3,] -0.0001549057  0.0171411782 0.005081584 0.0031880954 -0.007182115
#>              [,97]        [,98]        [,99]       [,100]
#> [1,] -0.0045011563 -0.010601480 -0.008247451 -0.005265305
#> [2,]  0.0006021864  0.001886641 -0.001362105 -0.002122644
#> [3,]  0.0239414696 -0.004677389 -0.018688184 -0.013334095
#> 
#> $rho
#> [1] 0.5 0.5 0.5
#> 
#> $omega
#> [1] 0.1 0.1 0.1
#> 
#> $sigma2v
#> [1] 0.01 0.01 0.01
#> 
#> $S
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    1    1    1    1    1    1    1    1    1     1     1     1     1     1
#> [2,]    1    1    1    1    1    1    1    1    1     1     1     1     1     1
#> [3,]    1    1    1    1    1    1    1    1    1     1     1     1     1     1
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>      [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62]
#> [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>      [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74]
#> [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>      [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86]
#> [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>      [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>      [,99] [,100]
#> [1,]     1      1
#> [2,]     1      1
#> [3,]     1      1
#> 
#> $sigma2_omega
#> [1] 1 1 1
#> 
#> $s_
#> [1] 0.05 0.05 0.05
#> 

## ------------------------------------------------
## Method `specify_starting_values_bsvar_sv$set_starting_values`
## ------------------------------------------------

# starting values for a bsvar model with 1 lag for a 3-variable system
sv = specify_starting_values_bsvar_sv$new(N = 3, p = 1, T = 100)

# Modify the starting values by:
sv_list = sv$get_starting_values()   # getting them as list
sv_list$A <- matrix(rnorm(12), 3, 4) # modifying the entry
sv$set_starting_values(sv_list)      # providing to the class object