Provides posterior summary of homoskedastic Structural VAR estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper parameters.

Usage

# S3 method for PosteriorBSVAR
summary(object, ...)

Arguments

object

an object of class PosteriorBSVAR obtained using the estimate() function applied to homoskedastic Bayesian Structural VAR model specification set by function specify_bsvar$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper-parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 1)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in$get_last_draw(), 20, , thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>            mean      sd 5% quantile 95% quantile
#> B[1,1] 35.07169 2.00556     32.3791     38.16358
#> 
#> $B$equation2
#>              mean      sd 5% quantile 95% quantile
#> B[2,1]  0.4746155 2.23210   -2.862022     3.225085
#> B[2,2] 38.4389879 1.74876   36.419991    41.610443
#> 
#> $B$equation3
#>               mean       sd 5% quantile 95% quantile
#> B[3,1] -14.2197285 2.540111  -18.282259   -11.118711
#> B[3,2]   0.5402698 2.597053   -3.628247     4.029777
#> B[3,3]  97.1896834 4.271249   90.539611   102.983390
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.92110150 0.02487880  0.88919625   0.95213555
#> lag1_var2 -0.00112926 0.01102134 -0.01680739   0.01500657
#> lag1_var3  0.09379220 0.03267487  0.05057149   0.13536096
#> const     -0.04543450 0.08668969 -0.15791954   0.08360013
#> 
#> $A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.01915793 0.01737171 -0.04072965   0.01049972
#> lag1_var2  0.95637289 0.01122760  0.93662872   0.97514231
#> lag1_var3  0.02835982 0.02281252 -0.00340219   0.05818256
#> const     -0.39983750 0.08432734 -0.57538349  -0.25969989
#> 
#> $A$equation3
#>                   mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.013931771 0.008005127 -0.02609203 -0.006207103
#> lag1_var2 -0.006861053 0.003790138 -0.01282486 -0.001338148
#> lag1_var3  1.015099317 0.010486399  1.00330819  1.029665621
#> const     -0.081978130 0.029806897 -0.12347171 -0.039190951
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        723.1028  622.9099   161.67043     1904.130
#> B[2,]_shrinkage        667.8606  799.9100   212.11241     1267.525
#> B[3,]_shrinkage       1563.1672 1389.7999   422.14413     3932.885
#> B[1,]_shrinkage_scale 5465.5258 5029.8175   650.42298    14524.303
#> B[2,]_shrinkage_scale 4935.2981 5597.1223   615.32648    15698.213
#> B[3,]_shrinkage_scale 7119.7817 9413.4102   963.71693    21119.530
#> B_global_scale         501.3505  526.4005    70.31452     1724.658
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.3436786 0.1827965   0.1162826    0.6366575
#> A[2,]_shrinkage       0.2779046 0.1402430   0.1178125    0.5077607
#> A[3,]_shrinkage       0.4288783 0.2379780   0.2193773    0.9066899
#> A[1,]_shrinkage_scale 4.1950990 1.6969321   1.8382150    7.2633680
#> A[2,]_shrinkage_scale 3.8230245 1.4993327   1.5297427    6.4381504
#> A[3,]_shrinkage_scale 4.8813353 1.9628925   2.7901551    7.5879831
#> A_global_scale        0.5172631 0.1152898   0.3706491    0.6863643
#> 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar$new(p = 1) |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>            mean      sd 5% quantile 95% quantile
#> B[1,1] 35.07169 2.00556     32.3791     38.16358
#> 
#> $B$equation2
#>              mean      sd 5% quantile 95% quantile
#> B[2,1]  0.4746155 2.23210   -2.862022     3.225085
#> B[2,2] 38.4389879 1.74876   36.419991    41.610443
#> 
#> $B$equation3
#>               mean       sd 5% quantile 95% quantile
#> B[3,1] -14.2197285 2.540111  -18.282259   -11.118711
#> B[3,2]   0.5402698 2.597053   -3.628247     4.029777
#> B[3,3]  97.1896834 4.271249   90.539611   102.983390
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.92110150 0.02487880  0.88919625   0.95213555
#> lag1_var2 -0.00112926 0.01102134 -0.01680739   0.01500657
#> lag1_var3  0.09379220 0.03267487  0.05057149   0.13536096
#> const     -0.04543450 0.08668969 -0.15791954   0.08360013
#> 
#> $A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.01915793 0.01737171 -0.04072965   0.01049972
#> lag1_var2  0.95637289 0.01122760  0.93662872   0.97514231
#> lag1_var3  0.02835982 0.02281252 -0.00340219   0.05818256
#> const     -0.39983750 0.08432734 -0.57538349  -0.25969989
#> 
#> $A$equation3
#>                   mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.013931771 0.008005127 -0.02609203 -0.006207103
#> lag1_var2 -0.006861053 0.003790138 -0.01282486 -0.001338148
#> lag1_var3  1.015099317 0.010486399  1.00330819  1.029665621
#> const     -0.081978130 0.029806897 -0.12347171 -0.039190951
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        723.1028  622.9099   161.67043     1904.130
#> B[2,]_shrinkage        667.8606  799.9100   212.11241     1267.525
#> B[3,]_shrinkage       1563.1672 1389.7999   422.14413     3932.885
#> B[1,]_shrinkage_scale 5465.5258 5029.8175   650.42298    14524.303
#> B[2,]_shrinkage_scale 4935.2981 5597.1223   615.32648    15698.213
#> B[3,]_shrinkage_scale 7119.7817 9413.4102   963.71693    21119.530
#> B_global_scale         501.3505  526.4005    70.31452     1724.658
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.3436786 0.1827965   0.1162826    0.6366575
#> A[2,]_shrinkage       0.2779046 0.1402430   0.1178125    0.5077607
#> A[3,]_shrinkage       0.4288783 0.2379780   0.2193773    0.9066899
#> A[1,]_shrinkage_scale 4.1950990 1.6969321   1.8382150    7.2633680
#> A[2,]_shrinkage_scale 3.8230245 1.4993327   1.5297427    6.4381504
#> A[3,]_shrinkage_scale 4.8813353 1.9628925   2.7901551    7.5879831
#> A_global_scale        0.5172631 0.1152898   0.3706491    0.6863643
#> 
#>