Provides posterior summary of homoskedastic Structural VAR estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper parameters.

Usage

# S3 method for PosteriorBSVAR
summary(object, ...)

Arguments

object

an object of class PosteriorBSVAR obtained using the estimate() function applied to homoskedastic Bayesian Structural VAR model specification set by function specify_bsvar$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper-parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 20, thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>            mean       sd 5% quantile 95% quantile
#> B[1,1] 35.34721 1.950092    33.15269     38.74541
#> 
#> $B$equation2
#>              mean       sd 5% quantile 95% quantile
#> B[2,1] -0.0599277 2.845804   -3.732495     3.694468
#> B[2,2] 38.8886179 1.670954   36.721744    41.534891
#> 
#> $B$equation3
#>               mean       sd 5% quantile 95% quantile
#> B[3,1] -14.6016855 2.273077  -17.949754   -11.765651
#> B[3,2]   0.1316957 2.361361   -3.534856     3.248074
#> B[3,3]  97.2024282 4.539362   90.784868   104.814105
#> 
#> 
#> $A
#> $A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.921502780 0.01443484  0.90076787   0.94448462
#> lag1_var2 -0.005626527 0.01195996 -0.02340468   0.01504477
#> lag1_var3  0.096522021 0.02040780  0.06289732   0.12377293
#> const     -0.065920298 0.08166447 -0.18762827   0.05250867
#> 
#> $A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.02309675 0.01858415 -0.048187854  0.008379975
#> lag1_var2  0.95214935 0.00982128  0.939433511  0.966837895
#> lag1_var3  0.03557990 0.02471657 -0.003302443  0.073106763
#> const     -0.42416014 0.07185700 -0.526388736 -0.304838242
#> 
#> $A$equation3
#>                   mean          sd 5% quantile  95% quantile
#> lag1_var1 -0.017371398 0.007761561 -0.02997809 -0.0072525647
#> lag1_var2 -0.008110085 0.005500958 -0.01799117 -0.0003848438
#> lag1_var3  1.020427452 0.009520848  1.00828835  1.0336596922
#> const     -0.086915764 0.039179501 -0.14982927 -0.0386123870
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        872.5976  588.8641    231.9587     1641.993
#> B[2,]_shrinkage        858.7865  649.9412    154.6349     1968.127
#> B[3,]_shrinkage       1917.4940  982.6032    842.3885     3221.388
#> B[1,]_shrinkage_scale 6670.6125 3410.2918   1621.4315    12485.861
#> B[2,]_shrinkage_scale 6754.1107 3858.1838   1298.1515    12086.229
#> B[3,]_shrinkage_scale 8381.9262 4707.8184   2637.1256    15618.447
#> B_global_scale         693.7659  388.7168    143.4759     1259.459
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.4784616 0.2572472   0.2040358    0.9231269
#> A[2,]_shrinkage       0.5644621 0.3321461   0.1954333    1.2071596
#> A[3,]_shrinkage       0.5290594 0.3258665   0.2505545    1.1668659
#> A[1,]_shrinkage_scale 5.9971124 2.5311093   2.7719142    9.5775780
#> A[2,]_shrinkage_scale 6.6700819 2.9828053   3.2785512   11.6590589
#> A[3,]_shrinkage_scale 6.8793052 3.3340115   3.2577238   11.0052564
#> A_global_scale        0.7801249 0.2211617   0.5554593    1.3070208
#> 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar$new() |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>            mean       sd 5% quantile 95% quantile
#> B[1,1] 35.34721 1.950092    33.15269     38.74541
#> 
#> $B$equation2
#>              mean       sd 5% quantile 95% quantile
#> B[2,1] -0.0599277 2.845804   -3.732495     3.694468
#> B[2,2] 38.8886179 1.670954   36.721744    41.534891
#> 
#> $B$equation3
#>               mean       sd 5% quantile 95% quantile
#> B[3,1] -14.6016855 2.273077  -17.949754   -11.765651
#> B[3,2]   0.1316957 2.361361   -3.534856     3.248074
#> B[3,3]  97.2024282 4.539362   90.784868   104.814105
#> 
#> 
#> $A
#> $A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.921502780 0.01443484  0.90076787   0.94448462
#> lag1_var2 -0.005626527 0.01195996 -0.02340468   0.01504477
#> lag1_var3  0.096522021 0.02040780  0.06289732   0.12377293
#> const     -0.065920298 0.08166447 -0.18762827   0.05250867
#> 
#> $A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.02309675 0.01858415 -0.048187854  0.008379975
#> lag1_var2  0.95214935 0.00982128  0.939433511  0.966837895
#> lag1_var3  0.03557990 0.02471657 -0.003302443  0.073106763
#> const     -0.42416014 0.07185700 -0.526388736 -0.304838242
#> 
#> $A$equation3
#>                   mean          sd 5% quantile  95% quantile
#> lag1_var1 -0.017371398 0.007761561 -0.02997809 -0.0072525647
#> lag1_var2 -0.008110085 0.005500958 -0.01799117 -0.0003848438
#> lag1_var3  1.020427452 0.009520848  1.00828835  1.0336596922
#> const     -0.086915764 0.039179501 -0.14982927 -0.0386123870
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        872.5976  588.8641    231.9587     1641.993
#> B[2,]_shrinkage        858.7865  649.9412    154.6349     1968.127
#> B[3,]_shrinkage       1917.4940  982.6032    842.3885     3221.388
#> B[1,]_shrinkage_scale 6670.6125 3410.2918   1621.4315    12485.861
#> B[2,]_shrinkage_scale 6754.1107 3858.1838   1298.1515    12086.229
#> B[3,]_shrinkage_scale 8381.9262 4707.8184   2637.1256    15618.447
#> B_global_scale         693.7659  388.7168    143.4759     1259.459
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.4784616 0.2572472   0.2040358    0.9231269
#> A[2,]_shrinkage       0.5644621 0.3321461   0.1954333    1.2071596
#> A[3,]_shrinkage       0.5290594 0.3258665   0.2505545    1.1668659
#> A[1,]_shrinkage_scale 5.9971124 2.5311093   2.7719142    9.5775780
#> A[2,]_shrinkage_scale 6.6700819 2.9828053   3.2785512   11.6590589
#> A[3,]_shrinkage_scale 6.8793052 3.3340115   3.2577238   11.0052564
#> A_global_scale        0.7801249 0.2211617   0.5554593    1.3070208
#> 
#>