Provides posterior summary of non-normal Structural VAR estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper parameters.

Usage

# S3 method for PosteriorBSVARMIX
summary(object, ...)

Arguments

object

an object of class PosteriorBSVARMIX obtained using the estimate() function applied to non-normal Bayesian Structural VAR model specification set by function specify_bsvar_mix$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper-parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_mix$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-finiteMIX model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 20, thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-finiteMIX model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>            mean          sd 5% quantile 95% quantile
#> B[1,1] 0.148059 0.006515433   0.1377122    0.1567447
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -8.974586 0.7853883   -10.13249    -7.691103
#> B[2,2] 29.004164 2.5408074    24.82448    32.713096
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -26.527502 2.378618  -30.620420   -23.584889
#> B[3,2]  -8.047908 1.350506   -9.885219    -5.800029
#> B[3,3]  43.810991 3.890118   38.182547    50.064276
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.17058992 0.03491433  1.13066812   1.22274260
#> lag1_var2 -0.03663762 0.01552085 -0.05814101  -0.01402295
#> lag1_var3 -1.10717351 0.04689666 -1.18244836  -1.04790154
#> const     -0.37036836 0.12893145 -0.53393365  -0.16259106
#> 
#> $A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.04567078 0.01451918  0.02838832   0.06431834
#> lag1_var2  0.95592150 0.01127490  0.94164612   0.97834181
#> lag1_var3 -0.32893812 0.02093910 -0.36053856  -0.30266894
#> const     -0.40535598 0.09080414 -0.51843571  -0.22054534
#> 
#> $A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.13684716 0.02251651  0.10786152  0.168850702
#> lag1_var2 -0.01976666 0.01037092 -0.03299991 -0.003281447
#> lag1_var3  0.23693034 0.03012152  0.19428770  0.273127762
#> const     -0.19875534 0.08532721 -0.30660046 -0.060775168
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        292.4182  214.4426    61.53563     661.6432
#> B[2,]_shrinkage        404.3663  261.4021   184.31395     768.3149
#> B[3,]_shrinkage        655.0951  373.2487   347.72224    1149.3401
#> B[1,]_shrinkage_scale 3031.5654 1829.2843   808.92849    5933.5761
#> B[2,]_shrinkage_scale 3068.9808 1680.8035  1072.71520    6071.9593
#> B[3,]_shrinkage_scale 4089.4152 2043.9241   966.37545    6977.1990
#> B_global_scale         337.2130  188.9994    92.43544     617.5004
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.6210550 0.5203199   0.2488507    0.9997862
#> A[2,]_shrinkage       0.3367933 0.2239936   0.1547297    0.9255948
#> A[3,]_shrinkage       0.3885354 0.2273155   0.1620304    0.6597644
#> A[1,]_shrinkage_scale 5.3974890 1.5314992   2.7392488    7.3065779
#> A[2,]_shrinkage_scale 4.0418964 1.2314576   2.6214907    5.5503249
#> A[3,]_shrinkage_scale 4.2382050 1.3376121   2.5333458    6.0607933
#> A_global_scale        0.5527270 0.1194241   0.3675523    0.7171201
#> 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_mix$new() |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-finiteMIX model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-finiteMIX model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>            mean          sd 5% quantile 95% quantile
#> B[1,1] 0.148059 0.006515433   0.1377122    0.1567447
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -8.974586 0.7853883   -10.13249    -7.691103
#> B[2,2] 29.004164 2.5408074    24.82448    32.713096
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -26.527502 2.378618  -30.620420   -23.584889
#> B[3,2]  -8.047908 1.350506   -9.885219    -5.800029
#> B[3,3]  43.810991 3.890118   38.182547    50.064276
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.17058992 0.03491433  1.13066812   1.22274260
#> lag1_var2 -0.03663762 0.01552085 -0.05814101  -0.01402295
#> lag1_var3 -1.10717351 0.04689666 -1.18244836  -1.04790154
#> const     -0.37036836 0.12893145 -0.53393365  -0.16259106
#> 
#> $A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.04567078 0.01451918  0.02838832   0.06431834
#> lag1_var2  0.95592150 0.01127490  0.94164612   0.97834181
#> lag1_var3 -0.32893812 0.02093910 -0.36053856  -0.30266894
#> const     -0.40535598 0.09080414 -0.51843571  -0.22054534
#> 
#> $A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.13684716 0.02251651  0.10786152  0.168850702
#> lag1_var2 -0.01976666 0.01037092 -0.03299991 -0.003281447
#> lag1_var3  0.23693034 0.03012152  0.19428770  0.273127762
#> const     -0.19875534 0.08532721 -0.30660046 -0.060775168
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        292.4182  214.4426    61.53563     661.6432
#> B[2,]_shrinkage        404.3663  261.4021   184.31395     768.3149
#> B[3,]_shrinkage        655.0951  373.2487   347.72224    1149.3401
#> B[1,]_shrinkage_scale 3031.5654 1829.2843   808.92849    5933.5761
#> B[2,]_shrinkage_scale 3068.9808 1680.8035  1072.71520    6071.9593
#> B[3,]_shrinkage_scale 4089.4152 2043.9241   966.37545    6977.1990
#> B_global_scale         337.2130  188.9994    92.43544     617.5004
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.6210550 0.5203199   0.2488507    0.9997862
#> A[2,]_shrinkage       0.3367933 0.2239936   0.1547297    0.9255948
#> A[3,]_shrinkage       0.3885354 0.2273155   0.1620304    0.6597644
#> A[1,]_shrinkage_scale 5.3974890 1.5314992   2.7392488    7.3065779
#> A[2,]_shrinkage_scale 4.0418964 1.2314576   2.6214907    5.5503249
#> A[3,]_shrinkage_scale 4.2382050 1.3376121   2.5333458    6.0607933
#> A_global_scale        0.5527270 0.1194241   0.3675523    0.7171201
#> 
#>