Provides posterior summary of non-normal Structural VAR estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper parameters.

Usage

# S3 method for PosteriorBSVARMIX
summary(object, ...)

Arguments

object

an object of class PosteriorBSVARMIX obtained using the estimate() function applied to non-normal Bayesian Structural VAR model specification set by function specify_bsvar_mix$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper-parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_mix$new(us_fiscal_lsuw, p = 1)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-finiteMIX model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in$get_last_draw(), 20, , thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-finiteMIX model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1481787 0.009814882   0.1273035    0.1611399
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -9.238632 0.6755464   -10.10740    -8.249924
#> B[2,2] 29.659042 2.1681585    26.50166    32.424030
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -25.018840 1.646357  -27.549663   -23.313544
#> B[3,2]  -7.414796 1.247150   -9.163418    -5.387297
#> B[3,3]  41.207152 2.559707   38.391448    45.055967
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.11647834 0.02446523  1.07612273   1.14998618
#> lag1_var2  0.01072407 0.02673783 -0.02329893   0.05678063
#> lag1_var3 -1.06634872 0.03170714 -1.11179241  -1.01608331
#> const     -0.12961170 0.25857943 -0.44893929   0.30502725
#> 
#> $A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.02338387 0.01621369  0.005820879   0.05222733
#> lag1_var2  0.96588379 0.01197017  0.948436817   0.98038902
#> lag1_var3 -0.30957627 0.02112417 -0.345802901  -0.28431741
#> const     -0.36525888 0.10767425 -0.541436284  -0.23671018
#> 
#> $A$equation3
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.10085595 0.01507793  0.077247235   0.12024123
#> lag1_var2  0.01403124 0.01811160 -0.007406267   0.04251045
#> lag1_var3  0.26288512 0.01962456  0.236820910   0.29309946
#> const     -0.01890137 0.17490432 -0.233002339   0.25927175
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean         sd 5% quantile 95% quantile
#> B[1,]_shrinkage        189.4881   75.80608    69.25197     300.9432
#> B[2,]_shrinkage        345.3050  148.68702   126.07438     534.0942
#> B[3,]_shrinkage        363.8659  220.99068   121.50568     618.9664
#> B[1,]_shrinkage_scale 2002.7478  929.65768   645.77261    3501.9782
#> B[2,]_shrinkage_scale 2268.3946 1174.72395   528.03324    4588.8465
#> B[3,]_shrinkage_scale 2033.0642  914.96951   758.44865    3428.1582
#> B_global_scale         203.8463   79.94743    63.08838     309.7430
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.5781518 0.2938980   0.2344258    1.0622152
#> A[2,]_shrinkage       0.2559171 0.1320011   0.1468227    0.4495747
#> A[3,]_shrinkage       0.6746118 0.4219648   0.2657338    1.3562834
#> A[1,]_shrinkage_scale 5.1483493 1.7798620   3.1636528    8.7908837
#> A[2,]_shrinkage_scale 3.3615421 1.3148939   2.0781477    5.2404178
#> A[3,]_shrinkage_scale 6.2369474 2.7099591   2.5697708   11.1402601
#> A_global_scale        0.5946045 0.1595063   0.3270677    0.8053415
#> 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_mix$new(p = 1) |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-finiteMIX model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-finiteMIX model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1481787 0.009814882   0.1273035    0.1611399
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -9.238632 0.6755464   -10.10740    -8.249924
#> B[2,2] 29.659042 2.1681585    26.50166    32.424030
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -25.018840 1.646357  -27.549663   -23.313544
#> B[3,2]  -7.414796 1.247150   -9.163418    -5.387297
#> B[3,3]  41.207152 2.559707   38.391448    45.055967
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.11647834 0.02446523  1.07612273   1.14998618
#> lag1_var2  0.01072407 0.02673783 -0.02329893   0.05678063
#> lag1_var3 -1.06634872 0.03170714 -1.11179241  -1.01608331
#> const     -0.12961170 0.25857943 -0.44893929   0.30502725
#> 
#> $A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.02338387 0.01621369  0.005820879   0.05222733
#> lag1_var2  0.96588379 0.01197017  0.948436817   0.98038902
#> lag1_var3 -0.30957627 0.02112417 -0.345802901  -0.28431741
#> const     -0.36525888 0.10767425 -0.541436284  -0.23671018
#> 
#> $A$equation3
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.10085595 0.01507793  0.077247235   0.12024123
#> lag1_var2  0.01403124 0.01811160 -0.007406267   0.04251045
#> lag1_var3  0.26288512 0.01962456  0.236820910   0.29309946
#> const     -0.01890137 0.17490432 -0.233002339   0.25927175
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean         sd 5% quantile 95% quantile
#> B[1,]_shrinkage        189.4881   75.80608    69.25197     300.9432
#> B[2,]_shrinkage        345.3050  148.68702   126.07438     534.0942
#> B[3,]_shrinkage        363.8659  220.99068   121.50568     618.9664
#> B[1,]_shrinkage_scale 2002.7478  929.65768   645.77261    3501.9782
#> B[2,]_shrinkage_scale 2268.3946 1174.72395   528.03324    4588.8465
#> B[3,]_shrinkage_scale 2033.0642  914.96951   758.44865    3428.1582
#> B_global_scale         203.8463   79.94743    63.08838     309.7430
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.5781518 0.2938980   0.2344258    1.0622152
#> A[2,]_shrinkage       0.2559171 0.1320011   0.1468227    0.4495747
#> A[3,]_shrinkage       0.6746118 0.4219648   0.2657338    1.3562834
#> A[1,]_shrinkage_scale 5.1483493 1.7798620   3.1636528    8.7908837
#> A[2,]_shrinkage_scale 3.3615421 1.3148939   2.0781477    5.2404178
#> A[3,]_shrinkage_scale 6.2369474 2.7099591   2.5697708   11.1402601
#> A_global_scale        0.5946045 0.1595063   0.3270677    0.8053415
#> 
#>