Provides posterior summary of heteroskedastic Structural VAR estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper parameters.

Usage

# S3 method for PosteriorBSVARMSH
summary(object, ...)

Arguments

object

an object of class PosteriorBSVARMSH obtained using the estimate() function applied to heteroskedastic Bayesian Structural VAR model specification set by function specify_bsvar_msh$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper-parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_msh$new(us_fiscal_lsuw, p = 1)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-stationaryMSH model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in$get_last_draw(), 20, , thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-stationaryMSH model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1495117 0.006768136   0.1405256    0.1592943
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -8.997323 0.7415495   -10.14514    -7.772178
#> B[2,2] 28.959076 2.3925140    25.01285    32.672727
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -25.901708 2.843951  -30.166573   -22.462299
#> B[3,2]  -7.351171 1.940371   -9.174592    -4.562161
#> B[3,3]  42.615494 3.999129   37.406200    49.799444
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.13708840 0.01669976   1.1154925   1.16323288
#> lag1_var2 -0.07712843 0.02619952  -0.1042296  -0.03549544
#> lag1_var3 -1.01860454 0.01786018  -1.0505957  -0.99327222
#> const     -0.43500888 0.17558337  -0.6045803  -0.13083544
#> 
#> $A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.02028028 0.01703468 -0.003603529   0.04795818
#> lag1_var2  0.94211172 0.01776461  0.924329677   0.96505319
#> lag1_var3 -0.28559701 0.02242937 -0.320229233  -0.25381030
#> const     -0.45188173 0.14265093 -0.614524846  -0.25248778
#> 
#> $A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.1106965 0.01065578  0.09531210   0.12496538
#> lag1_var2 -0.0476504 0.01820998 -0.06697974  -0.01797451
#> lag1_var3  0.3016152 0.01040283  0.28878968   0.31355813
#> const     -0.2508543 0.12239086 -0.36944701  -0.04680334
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean         sd 5% quantile 95% quantile
#> B[1,]_shrinkage        114.3015   70.51535    15.98242     211.2615
#> B[2,]_shrinkage        243.8445  210.73197    77.92626     628.0595
#> B[3,]_shrinkage        414.0962  152.06752   223.49707     666.0217
#> B[1,]_shrinkage_scale 1118.5167  643.18925   183.77152    1936.5893
#> B[2,]_shrinkage_scale 1454.6932  971.69638   366.58484    3114.6426
#> B[3,]_shrinkage_scale 2068.4430 1144.22593   569.60113    4004.4845
#> B_global_scale         147.9400   81.67465    38.52264     295.5986
#> 
#> $hyper$A
#>                            mean         sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.4513934 0.20187077  0.23892455    0.8527967
#> A[2,]_shrinkage       0.3370961 0.26994422  0.13696029    0.8888597
#> A[3,]_shrinkage       0.3538953 0.25113073  0.09505622    0.7257047
#> A[1,]_shrinkage_scale 4.4629283 1.03810588  2.94162504    6.1507351
#> A[2,]_shrinkage_scale 3.5472774 1.19913574  1.75832659    5.7954222
#> A[3,]_shrinkage_scale 3.7900870 1.72167921  1.63216934    7.0569664
#> A_global_scale        0.4971786 0.06864861  0.40588596    0.6013252
#> 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_msh$new(p = 1) |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-stationaryMSH model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-stationaryMSH model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1495117 0.006768136   0.1405256    0.1592943
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -8.997323 0.7415495   -10.14514    -7.772178
#> B[2,2] 28.959076 2.3925140    25.01285    32.672727
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -25.901708 2.843951  -30.166573   -22.462299
#> B[3,2]  -7.351171 1.940371   -9.174592    -4.562161
#> B[3,3]  42.615494 3.999129   37.406200    49.799444
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.13708840 0.01669976   1.1154925   1.16323288
#> lag1_var2 -0.07712843 0.02619952  -0.1042296  -0.03549544
#> lag1_var3 -1.01860454 0.01786018  -1.0505957  -0.99327222
#> const     -0.43500888 0.17558337  -0.6045803  -0.13083544
#> 
#> $A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.02028028 0.01703468 -0.003603529   0.04795818
#> lag1_var2  0.94211172 0.01776461  0.924329677   0.96505319
#> lag1_var3 -0.28559701 0.02242937 -0.320229233  -0.25381030
#> const     -0.45188173 0.14265093 -0.614524846  -0.25248778
#> 
#> $A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.1106965 0.01065578  0.09531210   0.12496538
#> lag1_var2 -0.0476504 0.01820998 -0.06697974  -0.01797451
#> lag1_var3  0.3016152 0.01040283  0.28878968   0.31355813
#> const     -0.2508543 0.12239086 -0.36944701  -0.04680334
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean         sd 5% quantile 95% quantile
#> B[1,]_shrinkage        114.3015   70.51535    15.98242     211.2615
#> B[2,]_shrinkage        243.8445  210.73197    77.92626     628.0595
#> B[3,]_shrinkage        414.0962  152.06752   223.49707     666.0217
#> B[1,]_shrinkage_scale 1118.5167  643.18925   183.77152    1936.5893
#> B[2,]_shrinkage_scale 1454.6932  971.69638   366.58484    3114.6426
#> B[3,]_shrinkage_scale 2068.4430 1144.22593   569.60113    4004.4845
#> B_global_scale         147.9400   81.67465    38.52264     295.5986
#> 
#> $hyper$A
#>                            mean         sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.4513934 0.20187077  0.23892455    0.8527967
#> A[2,]_shrinkage       0.3370961 0.26994422  0.13696029    0.8888597
#> A[3,]_shrinkage       0.3538953 0.25113073  0.09505622    0.7257047
#> A[1,]_shrinkage_scale 4.4629283 1.03810588  2.94162504    6.1507351
#> A[2,]_shrinkage_scale 3.5472774 1.19913574  1.75832659    5.7954222
#> A[3,]_shrinkage_scale 3.7900870 1.72167921  1.63216934    7.0569664
#> A_global_scale        0.4971786 0.06864861  0.40588596    0.6013252
#> 
#>