Provides posterior summary of heteroskedastic Structural VAR estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper parameters.

Usage

# S3 method for PosteriorBSVARMSH
summary(object, ...)

Arguments

object

an object of class PosteriorBSVARMSH obtained using the estimate() function applied to heteroskedastic Bayesian Structural VAR model specification set by function specify_bsvar_msh$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper-parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_msh$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-stationaryMSH model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 20, thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-stationaryMSH model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1465381 0.006627175   0.1376761    0.1548092
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -8.851756 0.8989556   -10.25571    -7.759311
#> B[2,2] 28.448014 2.8929037    24.93042    32.980642
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -26.203550 2.040778   -28.98457   -23.336001
#> B[3,2]  -7.724891 1.611732   -10.17135    -5.774019
#> B[3,3]  43.194049 3.354565    38.26175    47.530302
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.18225488 0.02686144  1.14888552   1.22452027
#> lag1_var2  0.04242146 0.01839318  0.02227519   0.06914864
#> lag1_var3 -1.15147329 0.03133035 -1.20006149  -1.11230918
#> const      0.19610707 0.14030268  0.05576973   0.40418763
#> 
#> $A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.05491755 0.01319579  0.03473064   0.07106499
#> lag1_var2  0.98346136 0.01399439  0.96036959   1.00588915
#> lag1_var3 -0.34995082 0.01776500 -0.37255791  -0.32172093
#> const     -0.18968627 0.12002141 -0.36467331  -0.03395072
#> 
#> $A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 0.14690373 0.01865491  0.12410041   0.17621429
#> lag1_var2 0.03431306 0.01083689  0.02104731   0.05043462
#> lag1_var3 0.20432628 0.02099601  0.16833682   0.23150814
#> const     0.19052099 0.08227734  0.09766455   0.30509802
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        145.4564 203.17265    24.89843     330.1820
#> B[2,]_shrinkage        209.2065 102.54211    90.34876     394.3303
#> B[3,]_shrinkage        413.8920 295.96453   154.27949     994.1749
#> B[1,]_shrinkage_scale 1172.2602 839.74747   411.15714    2948.6128
#> B[2,]_shrinkage_scale 1333.4500 718.11465   430.83694    2431.1375
#> B[3,]_shrinkage_scale 1356.6614 616.22478   596.50204    2604.5077
#> B_global_scale         127.8871  72.61581    38.60354     252.0162
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.6346892 0.1952808   0.3420336    0.8744203
#> A[2,]_shrinkage       0.4398107 0.2116136   0.1379731    0.7738683
#> A[3,]_shrinkage       0.4662115 0.2354953   0.1869862    0.9123358
#> A[1,]_shrinkage_scale 5.9906760 1.8310384   3.3706791    8.7678760
#> A[2,]_shrinkage_scale 5.2353141 2.3001292   2.2434635    8.4483674
#> A[3,]_shrinkage_scale 5.6023726 2.5211864   3.0462921   10.1724655
#> A_global_scale        0.6048046 0.1837860   0.3863849    0.9347859
#> 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_msh$new() |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-stationaryMSH model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-stationaryMSH model             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1465381 0.006627175   0.1376761    0.1548092
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -8.851756 0.8989556   -10.25571    -7.759311
#> B[2,2] 28.448014 2.8929037    24.93042    32.980642
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -26.203550 2.040778   -28.98457   -23.336001
#> B[3,2]  -7.724891 1.611732   -10.17135    -5.774019
#> B[3,3]  43.194049 3.354565    38.26175    47.530302
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.18225488 0.02686144  1.14888552   1.22452027
#> lag1_var2  0.04242146 0.01839318  0.02227519   0.06914864
#> lag1_var3 -1.15147329 0.03133035 -1.20006149  -1.11230918
#> const      0.19610707 0.14030268  0.05576973   0.40418763
#> 
#> $A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.05491755 0.01319579  0.03473064   0.07106499
#> lag1_var2  0.98346136 0.01399439  0.96036959   1.00588915
#> lag1_var3 -0.34995082 0.01776500 -0.37255791  -0.32172093
#> const     -0.18968627 0.12002141 -0.36467331  -0.03395072
#> 
#> $A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 0.14690373 0.01865491  0.12410041   0.17621429
#> lag1_var2 0.03431306 0.01083689  0.02104731   0.05043462
#> lag1_var3 0.20432628 0.02099601  0.16833682   0.23150814
#> const     0.19052099 0.08227734  0.09766455   0.30509802
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        145.4564 203.17265    24.89843     330.1820
#> B[2,]_shrinkage        209.2065 102.54211    90.34876     394.3303
#> B[3,]_shrinkage        413.8920 295.96453   154.27949     994.1749
#> B[1,]_shrinkage_scale 1172.2602 839.74747   411.15714    2948.6128
#> B[2,]_shrinkage_scale 1333.4500 718.11465   430.83694    2431.1375
#> B[3,]_shrinkage_scale 1356.6614 616.22478   596.50204    2604.5077
#> B_global_scale         127.8871  72.61581    38.60354     252.0162
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.6346892 0.1952808   0.3420336    0.8744203
#> A[2,]_shrinkage       0.4398107 0.2116136   0.1379731    0.7738683
#> A[3,]_shrinkage       0.4662115 0.2354953   0.1869862    0.9123358
#> A[1,]_shrinkage_scale 5.9906760 1.8310384   3.3706791    8.7678760
#> A[2,]_shrinkage_scale 5.2353141 2.3001292   2.2434635    8.4483674
#> A[3,]_shrinkage_scale 5.6023726 2.5211864   3.0462921   10.1724655
#> A_global_scale        0.6048046 0.1837860   0.3863849    0.9347859
#> 
#>