Provides posterior summary of heteroskedastic Structural VAR estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper parameters.

Usage

# S3 method for PosteriorBSVARSV
summary(object, ...)

Arguments

object

an object of class PosteriorBSVARSV obtained using the estimate() function applied to heteroskedastic Bayesian Structural VAR model specification set by function specify_bsvar_sv$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper-parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_sv$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 20, thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1455819 0.006906059   0.1366867    0.1570332
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -12.96006 0.7101617   -13.86599    -11.82598
#> B[2,2]  41.46463 2.2729604    37.86005     44.55122
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -42.605748 3.184304   -47.37540   -38.365706
#> B[3,2]  -7.163313 2.961788   -10.68584    -1.682399
#> B[3,3]  67.635649 4.540036    61.87773    74.571488
#> 
#> 
#> $A
#> $A$equation1
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  1.0736176 0.03207221   1.0314105   1.12292583
#> lag1_var2 -0.0900295 0.01990006  -0.1239641  -0.06532294
#> lag1_var3 -0.9643835 0.04543681  -1.0367134  -0.89892478
#> const     -0.7522780 0.16425303  -1.0117033  -0.52050473
#> 
#> $A$equation2
#>                  mean          sd  5% quantile 95% quantile
#> lag1_var1  0.02758689 0.015749195  0.001679211   0.05047499
#> lag1_var2  0.93383946 0.009149621  0.916340478   0.94565307
#> lag1_var3 -0.30114984 0.019797675 -0.333575754  -0.26802813
#> const     -0.56833958 0.077044342 -0.687731154  -0.43554739
#> 
#> $A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.08251693 0.02021653  0.05104172   0.10932584
#> lag1_var2 -0.05549155 0.01327854 -0.07557840  -0.03774828
#> lag1_var3  0.31817966 0.02869854  0.28373559   0.36453707
#> const     -0.45811191 0.11086830 -0.60601762  -0.30888238
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        104.3185  57.21714    29.40758     207.4399
#> B[2,]_shrinkage        451.4085 307.58977   173.12295    1099.6918
#> B[3,]_shrinkage        747.5464 270.69886   430.50200    1163.9341
#> B[1,]_shrinkage_scale 1189.9260 699.44093   323.56554    2467.0928
#> B[2,]_shrinkage_scale 1622.8752 925.21049   770.27939    3006.2229
#> B[3,]_shrinkage_scale 1927.6381 800.55487   581.68076    3192.7888
#> B_global_scale         146.7346  58.75045    50.00879     234.8103
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.6659067 0.3006225   0.3609818     1.185550
#> A[2,]_shrinkage       0.5141990 0.3191152   0.1881006     1.068408
#> A[3,]_shrinkage       0.5983275 0.3859581   0.2394958     1.171163
#> A[1,]_shrinkage_scale 6.9770521 2.0185308   4.4442454    10.505001
#> A[2,]_shrinkage_scale 6.3411529 3.2750275   2.5232409    13.501104
#> A[3,]_shrinkage_scale 6.5301699 2.5536954   3.5853892    11.667518
#> A_global_scale        0.7463883 0.2233471   0.4666108     1.155375
#> 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_sv$new() |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated             |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated             |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> Posterior summary of the parameters
#> -----------------------------------
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1455819 0.006906059   0.1366867    0.1570332
#> 
#> $B$equation2
#>             mean        sd 5% quantile 95% quantile
#> B[2,1] -12.96006 0.7101617   -13.86599    -11.82598
#> B[2,2]  41.46463 2.2729604    37.86005     44.55122
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -42.605748 3.184304   -47.37540   -38.365706
#> B[3,2]  -7.163313 2.961788   -10.68584    -1.682399
#> B[3,3]  67.635649 4.540036    61.87773    74.571488
#> 
#> 
#> $A
#> $A$equation1
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  1.0736176 0.03207221   1.0314105   1.12292583
#> lag1_var2 -0.0900295 0.01990006  -0.1239641  -0.06532294
#> lag1_var3 -0.9643835 0.04543681  -1.0367134  -0.89892478
#> const     -0.7522780 0.16425303  -1.0117033  -0.52050473
#> 
#> $A$equation2
#>                  mean          sd  5% quantile 95% quantile
#> lag1_var1  0.02758689 0.015749195  0.001679211   0.05047499
#> lag1_var2  0.93383946 0.009149621  0.916340478   0.94565307
#> lag1_var3 -0.30114984 0.019797675 -0.333575754  -0.26802813
#> const     -0.56833958 0.077044342 -0.687731154  -0.43554739
#> 
#> $A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.08251693 0.02021653  0.05104172   0.10932584
#> lag1_var2 -0.05549155 0.01327854 -0.07557840  -0.03774828
#> lag1_var3  0.31817966 0.02869854  0.28373559   0.36453707
#> const     -0.45811191 0.11086830 -0.60601762  -0.30888238
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        104.3185  57.21714    29.40758     207.4399
#> B[2,]_shrinkage        451.4085 307.58977   173.12295    1099.6918
#> B[3,]_shrinkage        747.5464 270.69886   430.50200    1163.9341
#> B[1,]_shrinkage_scale 1189.9260 699.44093   323.56554    2467.0928
#> B[2,]_shrinkage_scale 1622.8752 925.21049   770.27939    3006.2229
#> B[3,]_shrinkage_scale 1927.6381 800.55487   581.68076    3192.7888
#> B_global_scale         146.7346  58.75045    50.00879     234.8103
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.6659067 0.3006225   0.3609818     1.185550
#> A[2,]_shrinkage       0.5141990 0.3191152   0.1881006     1.068408
#> A[3,]_shrinkage       0.5983275 0.3859581   0.2394958     1.171163
#> A[1,]_shrinkage_scale 6.9770521 2.0185308   4.4442454    10.505001
#> A[2,]_shrinkage_scale 6.3411529 3.2750275   2.5232409    13.501104
#> A[3,]_shrinkage_scale 6.5301699 2.5536954   3.5853892    11.667518
#> A_global_scale        0.7463883 0.2233471   0.4666108     1.155375
#> 
#>