Provides posterior summary of heteroskedastic Structural VAR estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper parameters.

Usage

# S3 method for PosteriorBSVARSV
summary(object, ...)

Arguments

object

an object of class PosteriorBSVARSV obtained using the estimate() function applied to heteroskedastic Bayesian Structural VAR model specification set by function specify_bsvar_sv$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), and hyper-parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_sv$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated              |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 20, thin = 1)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated              |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1431795 0.007674081   0.1336533    0.1553142
#> 
#> $B$equation2
#>             mean       sd 5% quantile 95% quantile
#> B[2,1] -13.21416 0.609946   -14.20084    -12.43890
#> B[2,2]  42.69047 1.961153    40.23525     45.86416
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -41.537333 2.425838   -44.44068   -38.331401
#> B[3,2]  -8.696687 2.022140   -11.31986    -5.730334
#> B[3,3]  66.953934 3.572834    61.96817    71.637906
#> 
#> 
#> $A
#> $A$equation1
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1  0.773143830 0.02861974  0.731893056   0.81189081
#> lag1_var2 -0.001490227 0.02468449 -0.036874940   0.03367457
#> lag1_var3 -0.574675223 0.04330764 -0.625391638  -0.50513498
#> const      0.241696249 0.16154095  0.003891564   0.48313521
#> 
#> $A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.0688267 0.01778152 -0.09573319  -0.04015457
#> lag1_var2  0.9612313 0.01065573  0.94443144   0.97679158
#> lag1_var3 -0.1733763 0.02439933 -0.21516448  -0.13462896
#> const     -0.2606851 0.09104546 -0.39564580  -0.08828867
#> 
#> $A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.123325745 0.01964604 -0.15666107  -0.10024890
#> lag1_var2 -0.001398016 0.01494470 -0.01953372   0.02527741
#> lag1_var3  0.587543049 0.03099135  0.54451073   0.63126739
#> const      0.158797745 0.09500798  0.02759023   0.29229391
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean         sd 5% quantile 95% quantile
#> B[1,]_shrinkage        210.8698  121.55073    63.95338     421.2204
#> B[2,]_shrinkage        385.1534  144.85169   248.81396     647.7391
#> B[3,]_shrinkage        673.7392  230.32562   447.34967     962.7812
#> B[1,]_shrinkage_scale 1965.0038  820.77264   740.12868    3113.9756
#> B[2,]_shrinkage_scale 2061.4294  857.84459   772.31317    3418.4911
#> B[3,]_shrinkage_scale 2425.6858 1299.31262   862.04333    4946.0557
#> B_global_scale         200.5425   96.89497    73.46744     331.7725
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.5238126 0.3881165   0.1371073    1.1295956
#> A[2,]_shrinkage       0.3838431 0.1483610   0.1918165    0.5966805
#> A[3,]_shrinkage       0.4357945 0.2178203   0.1999113    0.8282578
#> A[1,]_shrinkage_scale 4.8063574 3.1185845   1.7134840   11.7211111
#> A[2,]_shrinkage_scale 4.6214546 1.8659300   1.8478155    7.0187372
#> A[3,]_shrinkage_scale 4.6903268 1.6688693   2.5296694    7.9693100
#> A_global_scale        0.5470192 0.1915078   0.3178399    0.8054769
#> 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_sv$new() |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated              |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every 10th draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated              |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>             mean          sd 5% quantile 95% quantile
#> B[1,1] 0.1431795 0.007674081   0.1336533    0.1553142
#> 
#> $B$equation2
#>             mean       sd 5% quantile 95% quantile
#> B[2,1] -13.21416 0.609946   -14.20084    -12.43890
#> B[2,2]  42.69047 1.961153    40.23525     45.86416
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -41.537333 2.425838   -44.44068   -38.331401
#> B[3,2]  -8.696687 2.022140   -11.31986    -5.730334
#> B[3,3]  66.953934 3.572834    61.96817    71.637906
#> 
#> 
#> $A
#> $A$equation1
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1  0.773143830 0.02861974  0.731893056   0.81189081
#> lag1_var2 -0.001490227 0.02468449 -0.036874940   0.03367457
#> lag1_var3 -0.574675223 0.04330764 -0.625391638  -0.50513498
#> const      0.241696249 0.16154095  0.003891564   0.48313521
#> 
#> $A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.0688267 0.01778152 -0.09573319  -0.04015457
#> lag1_var2  0.9612313 0.01065573  0.94443144   0.97679158
#> lag1_var3 -0.1733763 0.02439933 -0.21516448  -0.13462896
#> const     -0.2606851 0.09104546 -0.39564580  -0.08828867
#> 
#> $A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.123325745 0.01964604 -0.15666107  -0.10024890
#> lag1_var2 -0.001398016 0.01494470 -0.01953372   0.02527741
#> lag1_var3  0.587543049 0.03099135  0.54451073   0.63126739
#> const      0.158797745 0.09500798  0.02759023   0.29229391
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean         sd 5% quantile 95% quantile
#> B[1,]_shrinkage        210.8698  121.55073    63.95338     421.2204
#> B[2,]_shrinkage        385.1534  144.85169   248.81396     647.7391
#> B[3,]_shrinkage        673.7392  230.32562   447.34967     962.7812
#> B[1,]_shrinkage_scale 1965.0038  820.77264   740.12868    3113.9756
#> B[2,]_shrinkage_scale 2061.4294  857.84459   772.31317    3418.4911
#> B[3,]_shrinkage_scale 2425.6858 1299.31262   862.04333    4946.0557
#> B_global_scale         200.5425   96.89497    73.46744     331.7725
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.5238126 0.3881165   0.1371073    1.1295956
#> A[2,]_shrinkage       0.3838431 0.1483610   0.1918165    0.5966805
#> A[3,]_shrinkage       0.4357945 0.2178203   0.1999113    0.8282578
#> A[1,]_shrinkage_scale 4.8063574 3.1185845   1.7134840   11.7211111
#> A[2,]_shrinkage_scale 4.6214546 1.8659300   1.8478155    7.0187372
#> A[3,]_shrinkage_scale 4.6903268 1.6688693   2.5296694    7.9693100
#> A_global_scale        0.5470192 0.1915078   0.3178399    0.8054769
#> 
#>