Provides posterior summary of Structural VAR with t-distributed shocks estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), hyper-parameters, and Student-t degrees-of-freedom parameter \(\nu\).

Usage

# S3 method for class 'PosteriorBSVART'
summary(object, ...)

Arguments

object

an object of class PosteriorBSVART obtained using the estimate() function applied to homoskedastic Bayesian Structural VAR model specification set by function specify_bsvar$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), hyper-parameters, and Student-t degrees-of-freedom parameter \(\nu\).

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 20)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>            mean        sd 5% quantile 95% quantile
#> B[1,1] 3.495397 0.1512458    3.191542     3.650404
#> 
#> $B$equation2
#>            mean       sd 5% quantile 95% quantile
#> B[2,1] 30.90907 1.961008    27.83230     33.40790
#> B[2,2] 70.65000 4.488938    63.19642     77.66501
#> 
#> $B$equation3
#>             mean        sd 5% quantile 95% quantile
#> B[3,1] -87.12937  8.562466   -98.63844    -72.68972
#> B[3,2]  18.76816  2.675874    15.13831     23.29855
#> B[3,3] 121.60528 10.741138   103.17212    135.01758
#> 
#> 
#> $A
#> $A$equation1
#>                 mean          sd 5% quantile 95% quantile
#> lag1_var1  0.3320177 0.013569283   0.3094048    0.3518198
#> lag1_var2 -1.2681042 0.007502091  -1.2763503   -1.2581423
#> lag1_var3  2.1393285 0.018198137   2.1176285    2.1644979
#> const     -2.8241392 0.064846553  -2.9114333   -2.7436546
#> 
#> $A$equation2
#>                 mean          sd 5% quantile 95% quantile
#> lag1_var1  0.2538373 0.014208266   0.2374266    0.2815306
#> lag1_var2  1.5276757 0.008505344   1.5159444    1.5395729
#> lag1_var3 -0.8853320 0.020022061  -0.9243990   -0.8584071
#> const      0.9875328 0.067930666   0.8962296    1.0842548
#> 
#> $A$equation3
#>                 mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.5001605 0.012069733  -0.5174929   -0.4824274
#> lag1_var2 -0.9914909 0.008100803  -1.0033028   -0.9787891
#> lag1_var3  2.6487194 0.020551071   2.6169422    2.6741141
#> const     -2.1800358 0.050058890  -2.2486400   -2.1155189
#> 
#> 
#> $hyper
#> $hyper$B
#>                             mean         sd 5% quantile 95% quantile
#> B[1,]_shrinkage        1078.5673  1134.1851    20.94436     3382.990
#> B[2,]_shrinkage        1724.4940   912.8582   519.79705     3280.961
#> B[3,]_shrinkage        3180.0707  2101.8363  1037.26339     6972.580
#> B[1,]_shrinkage_scale  9424.4096  9823.6326   281.26399    28344.459
#> B[2,]_shrinkage_scale 10478.1711  9024.4094   339.10353    25405.583
#> B[3,]_shrinkage_scale 12368.3582 11114.9697   636.51524    33001.601
#> B_global_scale          958.6314   879.8611    38.03015     2317.524
#> 
#> $hyper$A
#>                             mean         sd 5% quantile 95% quantile
#> A[1,]_shrinkage        2.5711936  1.7576872   1.2711681     4.158282
#> A[2,]_shrinkage        0.9044899  0.4951082   0.4428915     1.603807
#> A[3,]_shrinkage        1.6313938  1.2998378   0.6571619     3.390814
#> A[1,]_shrinkage_scale 15.0461987 11.6753389   5.8174345    42.834659
#> A[2,]_shrinkage_scale 10.5723945  6.5421511   5.0801227    22.754281
#> A[3,]_shrinkage_scale 14.5892205 12.0214064   4.4504929    41.624681
#> A_global_scale         1.4180451  0.9499339   0.6657122     3.061231
#> 
#> 
#> $df
#>         mean           sd  5% quantile 95% quantile 
#>    3.3782716    0.4594437    2.6655111    3.9656234 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_t$new() |>
  estimate(S = 10) |> 
  estimate(S = 20) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>            mean        sd 5% quantile 95% quantile
#> B[1,1] 3.495397 0.1512458    3.191542     3.650404
#> 
#> $B$equation2
#>            mean       sd 5% quantile 95% quantile
#> B[2,1] 30.90907 1.961008    27.83230     33.40790
#> B[2,2] 70.65000 4.488938    63.19642     77.66501
#> 
#> $B$equation3
#>             mean        sd 5% quantile 95% quantile
#> B[3,1] -87.12937  8.562466   -98.63844    -72.68972
#> B[3,2]  18.76816  2.675874    15.13831     23.29855
#> B[3,3] 121.60528 10.741138   103.17212    135.01758
#> 
#> 
#> $A
#> $A$equation1
#>                 mean          sd 5% quantile 95% quantile
#> lag1_var1  0.3320177 0.013569283   0.3094048    0.3518198
#> lag1_var2 -1.2681042 0.007502091  -1.2763503   -1.2581423
#> lag1_var3  2.1393285 0.018198137   2.1176285    2.1644979
#> const     -2.8241392 0.064846553  -2.9114333   -2.7436546
#> 
#> $A$equation2
#>                 mean          sd 5% quantile 95% quantile
#> lag1_var1  0.2538373 0.014208266   0.2374266    0.2815306
#> lag1_var2  1.5276757 0.008505344   1.5159444    1.5395729
#> lag1_var3 -0.8853320 0.020022061  -0.9243990   -0.8584071
#> const      0.9875328 0.067930666   0.8962296    1.0842548
#> 
#> $A$equation3
#>                 mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.5001605 0.012069733  -0.5174929   -0.4824274
#> lag1_var2 -0.9914909 0.008100803  -1.0033028   -0.9787891
#> lag1_var3  2.6487194 0.020551071   2.6169422    2.6741141
#> const     -2.1800358 0.050058890  -2.2486400   -2.1155189
#> 
#> 
#> $hyper
#> $hyper$B
#>                             mean         sd 5% quantile 95% quantile
#> B[1,]_shrinkage        1078.5673  1134.1851    20.94436     3382.990
#> B[2,]_shrinkage        1724.4940   912.8582   519.79705     3280.961
#> B[3,]_shrinkage        3180.0707  2101.8363  1037.26339     6972.580
#> B[1,]_shrinkage_scale  9424.4096  9823.6326   281.26399    28344.459
#> B[2,]_shrinkage_scale 10478.1711  9024.4094   339.10353    25405.583
#> B[3,]_shrinkage_scale 12368.3582 11114.9697   636.51524    33001.601
#> B_global_scale          958.6314   879.8611    38.03015     2317.524
#> 
#> $hyper$A
#>                             mean         sd 5% quantile 95% quantile
#> A[1,]_shrinkage        2.5711936  1.7576872   1.2711681     4.158282
#> A[2,]_shrinkage        0.9044899  0.4951082   0.4428915     1.603807
#> A[3,]_shrinkage        1.6313938  1.2998378   0.6571619     3.390814
#> A[1,]_shrinkage_scale 15.0461987 11.6753389   5.8174345    42.834659
#> A[2,]_shrinkage_scale 10.5723945  6.5421511   5.0801227    22.754281
#> A[3,]_shrinkage_scale 14.5892205 12.0214064   4.4504929    41.624681
#> A_global_scale         1.4180451  0.9499339   0.6657122     3.061231
#> 
#> 
#> $df
#>         mean           sd  5% quantile 95% quantile 
#>    3.3782716    0.4594437    2.6655111    3.9656234 
#>