Provides posterior summary of Structural VAR with t-distributed shocks estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), hyper-parameters, and Student-t degrees-of-freedom parameter \(\nu\).

Usage

# S3 method for class 'PosteriorBSVART'
summary(object, ...)

Arguments

object

an object of class PosteriorBSVART obtained using the estimate() function applied to homoskedastic Bayesian Structural VAR model specification set by function specify_bsvar$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), hyper-parameters, and Student-t degrees-of-freedom parameter \(\nu\).

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 20)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>            mean       sd 5% quantile 95% quantile
#> B[1,1] 74.69871 8.433178    63.11781     87.66317
#> 
#> $B$equation2
#>              mean       sd 5% quantile 95% quantile
#> B[2,1]  0.3524171 3.853685   -5.377198     7.885433
#> B[2,2] 75.1493347 6.806240   66.684349    86.877034
#> 
#> $B$equation3
#>             mean        sd 5% quantile 95% quantile
#> B[3,1] -30.79812  4.592380   -37.22770   -24.387301
#> B[3,2] -11.87784  4.500643   -19.07397    -4.837736
#> B[3,3] 188.22241 13.898860   169.35346   210.788844
#> 
#> 
#> $A
#> $A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.95235178 0.012812373  0.93288599  0.976172320
#> lag1_var2 -0.01963324 0.009123887 -0.03558185 -0.005544448
#> lag1_var3  0.06168993 0.015807449  0.03103162  0.089011582
#> const     -0.17522313 0.081560271 -0.32673452 -0.083055699
#> 
#> $A$equation2
#>                  mean          sd  5% quantile 95% quantile
#> lag1_var1 -0.01210871 0.014443907 -0.033017522   0.00945166
#> lag1_var2  0.97375409 0.008814521  0.957403920   0.98518512
#> lag1_var3  0.01978208 0.018842660 -0.008064089   0.04648084
#> const     -0.22583651 0.079017520 -0.370383377  -0.12202626
#> 
#> $A$equation3
#>                   mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.009443179 0.006071262 -0.01956704 -0.002019030
#> lag1_var2 -0.010527804 0.003751930 -0.01762167 -0.006907209
#> lag1_var3  1.012457663 0.007951917  0.99997712  1.025463302
#> const     -0.095955725 0.027721974 -0.15500600 -0.066343611
#> 
#> 
#> $hyper
#> $hyper$B
#>                             mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        1868.8397 1350.4741    692.5561     4322.274
#> B[2,]_shrinkage        1225.2981  482.3231    371.7996     1891.985
#> B[3,]_shrinkage        4517.7552 2298.6399   2712.5316     7992.006
#> B[1,]_shrinkage_scale 10660.7101 7639.8413   3139.2441    26671.419
#> B[2,]_shrinkage_scale  9099.7348 4626.4443   2168.8420    17288.676
#> B[3,]_shrinkage_scale  9484.6772 5689.6507   1823.3661    18371.390
#> B_global_scale          896.5573  498.6766    146.7835     1766.578
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.3837151 0.2362267   0.1409643    0.9193515
#> A[2,]_shrinkage       0.3107323 0.1713383   0.1253543    0.5785431
#> A[3,]_shrinkage       0.3565428 0.1676987   0.1876333    0.5995065
#> A[1,]_shrinkage_scale 4.1145460 1.7734928   2.1080242    7.6265076
#> A[2,]_shrinkage_scale 3.9033919 1.9450506   2.0298542    7.9943773
#> A[3,]_shrinkage_scale 4.2772269 1.1129866   2.8553482    5.9678681
#> A_global_scale        0.5113035 0.1282720   0.3583145    0.7614349
#> 
#> 
#> $df
#>         mean           sd  5% quantile 95% quantile 
#>    3.9719937    0.7219846    3.1655776    5.0445172 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_t$new() |>
  estimate(S = 10) |> 
  estimate(S = 20) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>            mean       sd 5% quantile 95% quantile
#> B[1,1] 74.69871 8.433178    63.11781     87.66317
#> 
#> $B$equation2
#>              mean       sd 5% quantile 95% quantile
#> B[2,1]  0.3524171 3.853685   -5.377198     7.885433
#> B[2,2] 75.1493347 6.806240   66.684349    86.877034
#> 
#> $B$equation3
#>             mean        sd 5% quantile 95% quantile
#> B[3,1] -30.79812  4.592380   -37.22770   -24.387301
#> B[3,2] -11.87784  4.500643   -19.07397    -4.837736
#> B[3,3] 188.22241 13.898860   169.35346   210.788844
#> 
#> 
#> $A
#> $A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.95235178 0.012812373  0.93288599  0.976172320
#> lag1_var2 -0.01963324 0.009123887 -0.03558185 -0.005544448
#> lag1_var3  0.06168993 0.015807449  0.03103162  0.089011582
#> const     -0.17522313 0.081560271 -0.32673452 -0.083055699
#> 
#> $A$equation2
#>                  mean          sd  5% quantile 95% quantile
#> lag1_var1 -0.01210871 0.014443907 -0.033017522   0.00945166
#> lag1_var2  0.97375409 0.008814521  0.957403920   0.98518512
#> lag1_var3  0.01978208 0.018842660 -0.008064089   0.04648084
#> const     -0.22583651 0.079017520 -0.370383377  -0.12202626
#> 
#> $A$equation3
#>                   mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.009443179 0.006071262 -0.01956704 -0.002019030
#> lag1_var2 -0.010527804 0.003751930 -0.01762167 -0.006907209
#> lag1_var3  1.012457663 0.007951917  0.99997712  1.025463302
#> const     -0.095955725 0.027721974 -0.15500600 -0.066343611
#> 
#> 
#> $hyper
#> $hyper$B
#>                             mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage        1868.8397 1350.4741    692.5561     4322.274
#> B[2,]_shrinkage        1225.2981  482.3231    371.7996     1891.985
#> B[3,]_shrinkage        4517.7552 2298.6399   2712.5316     7992.006
#> B[1,]_shrinkage_scale 10660.7101 7639.8413   3139.2441    26671.419
#> B[2,]_shrinkage_scale  9099.7348 4626.4443   2168.8420    17288.676
#> B[3,]_shrinkage_scale  9484.6772 5689.6507   1823.3661    18371.390
#> B_global_scale          896.5573  498.6766    146.7835     1766.578
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.3837151 0.2362267   0.1409643    0.9193515
#> A[2,]_shrinkage       0.3107323 0.1713383   0.1253543    0.5785431
#> A[3,]_shrinkage       0.3565428 0.1676987   0.1876333    0.5995065
#> A[1,]_shrinkage_scale 4.1145460 1.7734928   2.1080242    7.6265076
#> A[2,]_shrinkage_scale 3.9033919 1.9450506   2.0298542    7.9943773
#> A[3,]_shrinkage_scale 4.2772269 1.1129866   2.8553482    5.9678681
#> A_global_scale        0.5113035 0.1282720   0.3583145    0.7614349
#> 
#> 
#> $df
#>         mean           sd  5% quantile 95% quantile 
#>    3.9719937    0.7219846    3.1655776    5.0445172 
#>