Provides posterior summary of Structural VAR with t-distributed shocks estimation

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), hyper-parameters, and Student-t degrees-of-freedom parameter \(\nu\).

Usage

# S3 method for class 'PosteriorBSVART'
summary(object, ...)

Arguments

object

an object of class PosteriorBSVART obtained using the estimate() function applied to homoskedastic Bayesian Structural VAR model specification set by function specify_bsvar$new() containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters: the structural matrix \(B\), autoregressive parameters \(A\), hyper-parameters, and Student-t degrees-of-freedom parameter \(\nu\).

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar_t$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.

# run the burn-in
burn_in        = estimate(specification, 10)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# estimate the model
posterior      = estimate(burn_in, 20)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>            mean       sd 5% quantile 95% quantile
#> B[1,1] 51.91362 3.758703    47.14723     57.86623
#> 
#> $B$equation2
#>              mean       sd 5% quantile 95% quantile
#> B[2,1] -0.4537775 2.390441   -3.306705     2.722118
#> B[2,2] 56.1074832 2.840258   51.870497    60.444387
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -12.297728 2.188019  -14.811455    -8.749792
#> B[3,2]  -6.649249 2.173099   -9.221663    -3.012380
#> B[3,3] 135.480163 6.263583  126.479490   141.641621
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.94422382 0.01543049  0.92214650   0.97491138
#> lag1_var2 -0.01211574 0.01812723 -0.03187688   0.01969634
#> lag1_var3  0.07053555 0.01935488  0.03535649   0.09530390
#> const     -0.11040876 0.14721797 -0.28124706   0.13972678
#> 
#> $A$equation2
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.01620647 0.012383836 -0.03476712  0.002853804
#> lag1_var2  0.96493641 0.009141184  0.95021647  0.977483399
#> lag1_var3  0.02626394 0.015816249  0.00410215  0.053080681
#> const     -0.30293175 0.080590471 -0.41581812 -0.178622509
#> 
#> $A$equation3
#>                   mean          sd 5% quantile  95% quantile
#> lag1_var1 -0.009232993 0.006990248 -0.02195842 -0.0003803604
#> lag1_var2 -0.007086324 0.003672678 -0.01124031 -0.0002753878
#> lag1_var3  1.011165987 0.008733323  1.00145878  1.0262119017
#> const     -0.069202611 0.034764772 -0.12115814 -0.0010979338
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage       1031.5209  733.9402    321.1895     2280.253
#> B[2,]_shrinkage       1187.8173 1022.3449    455.2167     2568.548
#> B[3,]_shrinkage       2433.3796 1462.9282   1345.4994     3903.015
#> B[1,]_shrinkage_scale 7961.6182 3467.9915   3012.9206    13185.697
#> B[2,]_shrinkage_scale 8934.7818 4670.0336   3242.0076    15467.551
#> B[3,]_shrinkage_scale 9488.6965 3732.5700   4437.6101    15634.181
#> B_global_scale         815.6063  309.7092    327.4834     1296.766
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.4309447 0.3887741   0.1502093    1.1838014
#> A[2,]_shrinkage       0.4593824 0.3526165   0.2544808    0.9783767
#> A[3,]_shrinkage       0.3714627 0.1562633   0.1824922    0.6382999
#> A[1,]_shrinkage_scale 5.5353921 2.7021929   2.7496572   11.4922023
#> A[2,]_shrinkage_scale 5.0237356 1.3877920   3.6363904    7.8441048
#> A[3,]_shrinkage_scale 4.7378072 1.3397061   3.1359356    7.5197216
#> A_global_scale        0.5851797 0.1256375   0.4520814    0.8108750
#> 
#> 
#> $df
#>         mean           sd  5% quantile 95% quantile 
#>    4.8615199    0.6244253    3.9291817    5.7388797 
#> 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_t$new() |>
  estimate(S = 10) |> 
  estimate(S = 20) |> 
  summary()
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 10 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR model                 |
#>     with t-distributed structural skocks          |
#> **************************************************|
#>  Progress of the MCMC simulation for 20 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#>  **************************************************|
#>  bsvars: Bayesian Structural Vector Autoregressions|
#>  **************************************************|
#>    Posterior summary of the parameters             |
#>  **************************************************|
#> $B
#> $B$equation1
#>            mean       sd 5% quantile 95% quantile
#> B[1,1] 51.91362 3.758703    47.14723     57.86623
#> 
#> $B$equation2
#>              mean       sd 5% quantile 95% quantile
#> B[2,1] -0.4537775 2.390441   -3.306705     2.722118
#> B[2,2] 56.1074832 2.840258   51.870497    60.444387
#> 
#> $B$equation3
#>              mean       sd 5% quantile 95% quantile
#> B[3,1] -12.297728 2.188019  -14.811455    -8.749792
#> B[3,2]  -6.649249 2.173099   -9.221663    -3.012380
#> B[3,3] 135.480163 6.263583  126.479490   141.641621
#> 
#> 
#> $A
#> $A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.94422382 0.01543049  0.92214650   0.97491138
#> lag1_var2 -0.01211574 0.01812723 -0.03187688   0.01969634
#> lag1_var3  0.07053555 0.01935488  0.03535649   0.09530390
#> const     -0.11040876 0.14721797 -0.28124706   0.13972678
#> 
#> $A$equation2
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.01620647 0.012383836 -0.03476712  0.002853804
#> lag1_var2  0.96493641 0.009141184  0.95021647  0.977483399
#> lag1_var3  0.02626394 0.015816249  0.00410215  0.053080681
#> const     -0.30293175 0.080590471 -0.41581812 -0.178622509
#> 
#> $A$equation3
#>                   mean          sd 5% quantile  95% quantile
#> lag1_var1 -0.009232993 0.006990248 -0.02195842 -0.0003803604
#> lag1_var2 -0.007086324 0.003672678 -0.01124031 -0.0002753878
#> lag1_var3  1.011165987 0.008733323  1.00145878  1.0262119017
#> const     -0.069202611 0.034764772 -0.12115814 -0.0010979338
#> 
#> 
#> $hyper
#> $hyper$B
#>                            mean        sd 5% quantile 95% quantile
#> B[1,]_shrinkage       1031.5209  733.9402    321.1895     2280.253
#> B[2,]_shrinkage       1187.8173 1022.3449    455.2167     2568.548
#> B[3,]_shrinkage       2433.3796 1462.9282   1345.4994     3903.015
#> B[1,]_shrinkage_scale 7961.6182 3467.9915   3012.9206    13185.697
#> B[2,]_shrinkage_scale 8934.7818 4670.0336   3242.0076    15467.551
#> B[3,]_shrinkage_scale 9488.6965 3732.5700   4437.6101    15634.181
#> B_global_scale         815.6063  309.7092    327.4834     1296.766
#> 
#> $hyper$A
#>                            mean        sd 5% quantile 95% quantile
#> A[1,]_shrinkage       0.4309447 0.3887741   0.1502093    1.1838014
#> A[2,]_shrinkage       0.4593824 0.3526165   0.2544808    0.9783767
#> A[3,]_shrinkage       0.3714627 0.1562633   0.1824922    0.6382999
#> A[1,]_shrinkage_scale 5.5353921 2.7021929   2.7496572   11.4922023
#> A[2,]_shrinkage_scale 5.0237356 1.3877920   3.6363904    7.8441048
#> A[3,]_shrinkage_scale 4.7378072 1.3397061   3.1359356    7.5197216
#> A_global_scale        0.5851797 0.1256375   0.4520814    0.8108750
#> 
#> 
#> $df
#>         mean           sd  5% quantile 95% quantile 
#>    4.8615199    0.6244253    3.9291817    5.7388797 
#>