Provides posterior summary of structural shocks' conditional standard deviations

Provides posterior summary of structural shocks' conditional standard deviations including their mean, standard deviations, as well as 5 and 95 percentiles.

Usage

# S3 method for class 'PosteriorSigma'
summary(object, ...)

Arguments

object

an object of class PosteriorSigma obtained using the compute_conditional_sd() function containing posterior draws of conditional standard deviations of structural shocks.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the structural shocks' conditional standard deviations for each of the shocks and periods.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

specification  = specify_bsvar_sv$new(us_fiscal_lsuw)
#> The identification is set to the default option of lower-triangular structural matrix.
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated              |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated              |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|

# compute structural shocks' conditional standard deviations
sigma          = compute_conditional_sd(posterior)
sigma_summary  = summary(sigma)
sigma_summary$shock1[,1] # access posterior mean of shock1
#>         1         2         3         4         5         6         7         8 
#> 1.2858636 1.4721164 1.8749484 2.0752899 1.4534846 1.4031569 1.5560218 1.4649098 
#>         9        10        11        12        13        14        15        16 
#> 1.4483785 1.4967780 1.3254420 1.2011318 1.2923687 1.0647769 1.0533771 1.0053960 
#>        17        18        19        20        21        22        23        24 
#> 1.0014379 1.0280112 1.2485945 1.0699520 1.2157862 1.1174838 1.2319058 1.2314363 
#>        25        26        27        28        29        30        31        32 
#> 1.2760461 1.1927764 1.1846372 1.3896894 1.3471548 1.3275255 1.1554765 1.0822197 
#>        33        34        35        36        37        38        39        40 
#> 1.1097451 1.3275126 1.1714078 1.1627883 1.0609024 1.0576220 1.0874272 1.4729693 
#>        41        42        43        44        45        46        47        48 
#> 1.8683855 1.6638867 1.6985215 1.5035111 1.5309895 1.5008229 1.6377543 1.5544570 
#>        49        50        51        52        53        54        55        56 
#> 1.3409437 1.2485610 1.1831611 1.3709581 1.5782038 1.6302457 1.5527580 1.5919648 
#>        57        58        59        60        61        62        63        64 
#> 1.4666082 1.5004736 1.4667139 1.5591748 1.5815305 1.3555232 1.2954047 1.3982861 
#>        65        66        67        68        69        70        71        72 
#> 1.2542188 1.4049358 1.4808153 1.1953382 1.1011670 1.1936990 1.1746190 0.9749527 
#>        73        74        75        76        77        78        79        80 
#> 1.0712843 1.2107082 1.2464861 0.9800366 1.0751662 1.0708111 1.1404629 1.0648921 
#>        81        82        83        84        85        86        87        88 
#> 1.3344271 1.3066329 1.1527633 1.0105820 1.0527967 1.1292015 0.9512488 1.1351744 
#>        89        90        91        92        93        94        95        96 
#> 1.0274720 1.0655521 0.9674628 1.0164411 1.2120510 1.1890754 1.2602107 1.3364109 
#>        97        98        99       100       101       102       103       104 
#> 1.1845516 1.1441752 1.0666973 1.0569387 0.9373361 0.9970655 1.0851096 1.3279370 
#>       105       106       107       108       109       110       111       112 
#> 1.4003030 1.3980780 1.4089621 1.3574099 1.3951713 1.3725350 1.2833543 1.7147566 
#>       113       114       115       116       117       118       119       120 
#> 1.5036778 1.5526185 1.5829983 1.3522610 1.1108681 1.1522191 1.5252063 1.4886976 
#>       121       122       123       124       125       126       127       128 
#> 1.4869673 1.4102738 1.4065707 1.5403755 1.6165784 1.4000275 1.4944204 1.5468641 
#>       129       130       131       132       133       134       135       136 
#> 1.4100414 1.3993968 1.5855966 1.6156419 1.8027014 1.8307230 1.7223280 1.6915939 
#>       137       138       139       140       141       142       143       144 
#> 1.5315683 1.2240821 1.3243541 1.4712950 1.5806042 1.5793784 1.4974610 1.4275708 
#>       145       146       147       148       149       150       151       152 
#> 1.5936508 1.2944423 1.4348997 1.5440466 1.4521280 1.4047210 1.1687340 1.1582352 
#>       153       154       155       156       157       158       159       160 
#> 1.0399504 1.1621792 1.1826900 1.1510777 1.2397504 1.2758958 1.2932948 1.0787081 
#>       161       162       163       164       165       166       167       168 
#> 0.9740709 1.1509242 1.0218033 1.0244801 1.1997437 1.3029035 1.2654809 1.2619862 
#>       169       170       171       172       173       174       175       176 
#> 1.1854100 1.2253669 1.4364663 1.6314274 1.6183663 1.5191603 1.2913640 1.1424831 
#>       177       178       179       180       181       182       183       184 
#> 1.1555867 1.1429118 1.1509543 1.0542923 1.1506400 1.1504751 1.2416466 1.4330434 
#>       185       186       187       188       189       190       191       192 
#> 1.3506527 1.6125576 1.6180051 1.6036591 1.7375193 1.6940368 1.7823734 1.4501439 
#>       193       194       195       196       197       198       199       200 
#> 1.4906160 1.2885909 1.2051153 0.9929309 0.9312806 0.9449221 1.2946224 1.5996899 
#>       201       202       203       204       205       206       207       208 
#> 1.7542154 1.6862944 1.7158426 1.5908163 1.4204982 1.6261722 1.5883229 1.3641833 
#>       209       210       211       212       213       214       215       216 
#> 1.0562634 0.8260937 0.9719752 0.8387220 1.0124027 1.0661714 1.2908364 1.3687069 
#>       217       218       219       220       221       222       223       224 
#> 1.4719412 1.7765299 2.0198360 1.6218477 1.4861136 1.3630397 1.3344718 1.4338955 
#>       225       226       227       228       229       230       231       232 
#> 1.7690842 2.0732715 1.5385664 1.2752277 1.4253305 1.3665889 1.2352825 1.0332362 
#>       233       234       235       236       237       238       239       240 
#> 1.2485256 1.2977271 1.3025561 1.2908222 1.1400729 1.2395925 1.3135772 1.1879730 
#>       241       242       243       244       245       246       247       248 
#> 1.2479818 1.4424672 1.2066252 1.0668876 1.1622634 1.4103459 1.4385760 1.2783667 
#>       249       250       251       252       253       254       255       256 
#> 1.2401520 1.1239047 1.0356965 1.0642739 1.1316393 1.1832988 1.1558394 1.1040553 
#>       257       258       259       260       261       262       263       264 
#> 1.2489479 1.2307754 1.1297814 1.1893529 1.3558372 1.3997384 1.6751761 1.2365760 
#>       265       266       267       268       269       270       271       272 
#> 1.2492738 1.4150458 1.3012698 1.4535182 1.4307822 1.3958053 1.3969153 1.6804928 
#>       273       274       275       276       277       278       279       280 
#> 1.4532737 1.3628802 1.3715672 1.4933009 1.3890303 1.2674322 1.2796168 1.1473091 
#>       281       282       283       284       285       286       287       288 
#> 1.0557035 1.2528324 1.2969293 1.4867298 1.5556635 1.8337173 1.4040351 1.3601296 
#>       289       290       291       292       293       294       295       296 
#> 1.2452177 1.1211412 0.9980942 1.1787044 1.1252522 1.0912927 1.1014558 1.3672467 
#>       297       298       299       300       301       302       303       304 
#> 1.3540241 1.3433463 1.3679006 1.2512984 1.2847371 1.3910548 1.3583153 1.5568602 
#>       305       306       307       308       309       310       311       312 
#> 1.6722733 1.7332907 1.4450658 1.0304205 1.0002816 0.9555493 0.8705383 0.9306071 

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_sv$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_conditional_sd() |>
  summary() -> sigma_summary
#> The identification is set to the default option of lower-triangular structural matrix.
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated              |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
#> **************************************************|
#> bsvars: Bayesian Structural Vector Autoregressions|
#> **************************************************|
#>  Gibbs sampler for the SVAR-SV model              |
#>    Non-centred SV model is estimated              |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
sigma_summary$shock1[,1]
#>         1         2         3         4         5         6         7         8 
#> 0.9945854 1.0080705 1.0130886 0.9935794 0.9734969 1.0094503 0.9782375 0.9986469 
#>         9        10        11        12        13        14        15        16 
#> 1.0027109 1.0119343 0.9726137 0.9996900 0.9984447 0.9926003 1.0345768 0.9742216 
#>        17        18        19        20        21        22        23        24 
#> 1.0422353 0.9819422 0.9820023 0.9899784 1.0048604 0.9867976 1.0157494 0.9957177 
#>        25        26        27        28        29        30        31        32 
#> 1.0090680 1.0054423 1.0002072 0.9990580 0.9937139 1.0059995 1.0149757 1.0029307 
#>        33        34        35        36        37        38        39        40 
#> 1.0140612 1.0039793 0.9964935 1.0042504 0.9926369 1.0170345 0.9992104 0.9903386 
#>        41        42        43        44        45        46        47        48 
#> 0.9948896 1.0041113 0.9868621 1.0242755 1.0094520 1.0048385 0.9885195 1.0079288 
#>        49        50        51        52        53        54        55        56 
#> 1.0056711 1.0014456 0.9759306 1.0448404 1.0044582 1.0052775 1.0009556 0.9988308 
#>        57        58        59        60        61        62        63        64 
#> 0.9903179 1.0100421 0.9872820 1.0002450 0.9787740 0.9922141 1.0180734 0.9901782 
#>        65        66        67        68        69        70        71        72 
#> 1.0093636 0.9956949 1.0089862 1.0015283 1.0083289 0.9903633 0.9917637 1.0285490 
#>        73        74        75        76        77        78        79        80 
#> 0.9856637 0.9940071 0.9836511 0.9910639 1.0097174 1.0071194 0.9766568 1.0084377 
#>        81        82        83        84        85        86        87        88 
#> 0.9868018 1.0050472 0.9902202 0.9979605 1.0155092 0.9897114 1.0007307 0.9862941 
#>        89        90        91        92        93        94        95        96 
#> 1.0080932 0.9527076 1.0108298 1.0042308 0.9841935 1.0250775 1.0147536 1.0104302 
#>        97        98        99       100       101       102       103       104 
#> 0.9980342 0.9940630 1.0207331 1.0117451 1.0100616 1.0453753 0.9961338 0.9828939 
#>       105       106       107       108       109       110       111       112 
#> 0.9946213 1.0188399 0.9991941 1.0066252 0.9756961 0.9832598 1.0017668 1.0035293 
#>       113       114       115       116       117       118       119       120 
#> 1.0088667 1.0054940 0.9887455 1.0214759 1.0128940 0.9911235 1.0415166 0.9572778 
#>       121       122       123       124       125       126       127       128 
#> 1.0308586 0.9951968 1.0152136 0.9862525 0.9970737 0.9776533 1.0155057 0.9961687 
#>       129       130       131       132       133       134       135       136 
#> 0.9612909 1.0227012 0.9809456 0.9791996 1.0129536 0.9987944 1.0088317 1.0194824 
#>       137       138       139       140       141       142       143       144 
#> 0.9942686 1.0169557 1.0281726 0.9858705 1.0186284 0.9720663 1.0133280 1.0002446 
#>       145       146       147       148       149       150       151       152 
#> 0.9895924 1.0190337 0.9972602 0.9981625 0.9831003 0.9933112 1.0074158 0.9901673 
#>       153       154       155       156       157       158       159       160 
#> 1.0018638 1.0141728 0.9772393 1.0194713 0.9960194 0.9925609 0.9904140 1.0251443 
#>       161       162       163       164       165       166       167       168 
#> 1.0101009 0.9980549 0.9855212 1.0410404 0.9992892 0.9936205 0.9960645 1.0001180 
#>       169       170       171       172       173       174       175       176 
#> 0.9662186 1.0257192 0.9831774 0.9945835 0.9995642 0.9882331 1.0001897 1.0067047 
#>       177       178       179       180       181       182       183       184 
#> 0.9919734 1.0019433 0.9852932 1.0138016 0.9980963 0.9758753 1.0090457 0.9892497 
#>       185       186       187       188       189       190       191       192 
#> 1.0219829 0.9925496 1.0334023 0.9945607 1.0026639 0.9944946 1.0365468 0.9654966 
#>       193       194       195       196       197       198       199       200 
#> 1.0133517 0.9961316 1.0148677 0.9881016 1.0203858 0.9945777 1.0006496 0.9826278 
#>       201       202       203       204       205       206       207       208 
#> 1.0094941 0.9904698 1.0125698 0.9795972 0.9985098 1.0029438 0.9718611 1.0200467 
#>       209       210       211       212       213       214       215       216 
#> 1.0015823 1.0005916 1.0421778 0.9783113 1.0399610 0.9802246 1.0101978 1.0247315 
#>       217       218       219       220       221       222       223       224 
#> 0.9960228 0.9991458 0.9885871 0.9844921 0.9904237 1.0091676 1.0018287 1.0097534 
#>       225       226       227       228       229       230       231       232 
#> 1.0026809 1.0005292 0.9933633 0.9894998 1.0215378 0.9823989 1.0053906 1.0399506 
#>       233       234       235       236       237       238       239       240 
#> 0.9898609 0.9815497 0.9951437 0.9992348 0.9882198 1.0331640 1.0036065 0.9951542 
#>       241       242       243       244       245       246       247       248 
#> 1.0084370 1.0000811 0.9887817 1.0337860 0.9909878 1.0162419 0.9723170 1.0309542 
#>       249       250       251       252       253       254       255       256 
#> 0.9796548 1.0039294 1.0033576 0.9762407 1.0449311 0.9830372 1.0049104 0.9917841 
#>       257       258       259       260       261       262       263       264 
#> 1.0066488 0.9980701 1.0018576 0.9988628 1.0069051 1.0007862 1.0431006 0.9732614 
#>       265       266       267       268       269       270       271       272 
#> 1.0066611 0.9976159 0.9899780 1.0526493 0.9695122 0.9857041 1.0046431 1.0007927 
#>       273       274       275       276       277       278       279       280 
#> 1.0069496 0.9957002 1.0076996 0.9953667 0.9792970 1.0337557 1.0070412 1.0118759 
#>       281       282       283       284       285       286       287       288 
#> 0.9985437 0.9946470 0.9965730 1.0283263 0.9896912 1.0179431 0.9863955 1.0220172 
#>       289       290       291       292       293       294       295       296 
#> 1.0017955 0.9740414 1.0037613 1.0107874 0.9979119 0.9972914 0.9857712 1.0013586 
#>       297       298       299       300       301       302       303       304 
#> 0.9892045 1.0140473 0.9802867 1.0013280 1.0067098 0.9827061 0.9932864 1.0288685 
#>       305       306       307       308       309       310       311       312 
#> 0.9901078 1.0318020 0.9791816 1.0361062 0.9778163 1.0045995 1.0072899 1.0061850